### 2024-06-24 #6927 @editorialbot
[PRE REVIEW]: Dolphin: A Python package for large-scale InSAR PS/DS processing
**Submitting author:** @scottstanie (<a href="http://orcid.org/0000-0002-3055-5731">Scott Staniewicz</a>)
**Repository:** https://github.com/isce-framework/dolphin
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.19.0
