### 2024-06-25 9eb81f3
Let stitching_bursts run in parallel (#325)
- Add option to run `merge_dates` with `thread_map`
- Same for `estimate_interferometric_correlations`
