### 2024-07-01 #6927 @kthyng
@scottstanie In your paper's summary or statement of need section, can you add a bit about what makes your software important for research? As in, how is it used by researchers, in what line of research, or for what purpose? Something along those lines, please.
