### 2024-07-01 #6927 @kthyng
Also @scottstanie is there a reason for your license file the way it is (select a license, both open source licenses) or is it an accident over time? If there isn't a reason, we would recommend you choose one.
