+++ src/dolphin/atmosphere/__init__.py
@@ -2,3 +2,4 @@
 
 from .ionosphere import *
 from .troposphere import *
+from ._netcdf import *
