+++ src/dolphin/atmosphere/troposphere.py
@@ -1,23 +1,24 @@
 from __future__ import annotations
 
 import datetime
-from dataclasses import dataclass
+from dataclasses import dataclass # type: ignore
 from os import fspath
 from pathlib import Path
 from typing import Mapping, Sequence
 
-import numpy as np
+import numpy as np # type: ignore
 import opera_utils as oput
-from opera_utils import get_dates
-from osgeo import gdal
-from rasterio.crs import CRS
-from rasterio.warp import transform_bounds
-from scipy.interpolate import RegularGridInterpolator
+from opera_utils import get_dates, group_by_date, DATE_FORMAT
+from osgeo import gdal # type: ignore
+from rasterio.crs import CRS # type: ignore
+from rasterio.warp import transform_bounds # type: ignore
+from scipy.interpolate import RegularGridInterpolator # type: ignore
 
 from dolphin import io
 from dolphin._log import get_log
 from dolphin._types import Bbox, Filename, TropoModel, TropoType
 from dolphin.utils import format_date_pair
+from ._netcdf import delay_from_netcdf, group_netcdf_by_date
 
 logger = get_log(__name__)
 
@@ -79,10 +80,10 @@ class DelayParams:
 def estimate_tropospheric_delay(
     ifg_file_list: Sequence[Path],
     slc_files: Mapping[tuple[datetime.datetime], Sequence[Filename]],
-    troposphere_files: Mapping[tuple[datetime.datetime], Sequence[Filename]],
+    troposphere_files: Sequence[Filename],
     geom_files: dict[str, Path],
     output_dir: Path,
-    tropo_package: str,
+    file_date_fmt: str,
     tropo_model: TropoModel,
     tropo_delay_type: TropoType,
     epsg: int,
@@ -96,15 +97,16 @@ def estimate_tropospheric_delay(
         List of interferogram files.
     slc_files : Mapping[tuple[datetime.datetime], Sequence[Filename]]
         Dictionary of SLC files indexed by date.
-    troposphere_files : Mapping[tuple[datetime.datetime], Sequence[Filename]]
-        Dictionary of troposphere files indexed by date.
+    troposphere_files : Sequence[Filename]
+        list of troposphere files.
     geom_files : dict[str, Path]
         Dictionary of geometry files with height and incidence angle, or
         los_east and los_north.
     output_dir : Path
         Output directory.
-    tropo_package : str
-        Troposphere processing package ('pyaps' or 'raider').
+    file_date_fmt : str
+        Format of the date in the filename.
+        Default is [dolphin.DEFAULT_DATETIME_FORMAT][]
     tropo_model : TropoModel
         Tropospheric model (ERA5, HRES, ...).
     tropo_delay_type : TropoType
@@ -141,15 +143,21 @@ def estimate_tropospheric_delay(
 
     wavelength = oput.get_radar_wavelength(one_of_slcs)
 
-    tropo_run = compute_pyaps if tropo_package.lower() == "pyaps" else compute_raider
-
-    # comb is short for the summation of wet and dry components
-    if (tropo_delay_type.value == "hydrostatic") and (tropo_package == "raider"):
-        delay_type = "hydro"
-    elif (tropo_delay_type.value == "hydrostatic") and (tropo_package == "pyaps"):
-        delay_type = "dry"
+    delay_type = tropo_delay_type.value
+    
+    if troposphere_files[0].suffix == '.nc':
+        tropo_files = group_netcdf_by_date(troposphere_files)
+        tropo_run = compute_tropo_delay_from_netcdf
+        if (tropo_delay_type.value == "hydrostatic"):
+            delay_type = "hydro"
     else:
-        delay_type = tropo_delay_type.value
[…]
