+++ src/dolphin/atmosphere/_netcdf.py
@@ -1,0 +1,464 @@
+from __future__ import annotations
+
+import os
+import datetime
+import numpy as np # type: ignore
+from pyproj import CRS, Transformer # type: ignore
+from typing import Union, Optional, Iterable
+import xarray # type: ignore
+from scipy.interpolate import RegularGridInterpolator as Interpolator # type: ignore
+from collections import defaultdict
+import itertools
+try:
+    import multiprocessing as mp
+except ImportError:
+    mp = None
+    
+
+from dolphin._log import get_log
+from opera_utils import reproject_bounds # type: ignore
+from dolphin._types import Filename, TropoModel
+
+logger = get_log(__name__)
+
+###########
+# Mostly inherited from RAiDER
+
+__all__ = ["delay_from_netcdf"]
+
+def delay_from_netcdf(
+        dt: datetime.datetime,
+        weather_model_file: str,
+        SNWE: tuple[float, float, float, float],
+        height_levels: Optional[list[float]],
+        out_proj: Union[int, str]=4326,
+        weather_model: TropoModel=TropoModel.ECMWF,
+    ):
+    """Calculate integrated delays on query points.
+
+    Parameters
+    ----------
+    dt: Datetime                 
+        Datetime object for determining when to calculate delays
+    weather_model_File: string  
+        Name of the NETCDF file containing a pre-processed weather model
+    aoi: AOI object             
+        AOI object
+    height_levels: list         
+        (optional) list of height levels on which to calculate delays. Only needed for cube generation.
+    out_proj: int,str           
+        (optional) EPSG code for output projection
+
+    Returns
+    -------
+        xarray Dataset *or* ndarrays:  
+            wet and hydrostatic delays at the grid nodes / query points.
+    """
+    crs = CRS(out_proj)
+    
+    # Load CRS from weather model file
+    with xarray.load_dataset(weather_model_file) as ds:
+        try:
+            wm_proj = CRS.from_wkt(ds['proj'].attrs['crs_wkt'])
+        except KeyError:
+            logger.warning("WARNING: I can't find a CRS in the weather model file, so I will assume you are using WGS84")
+            wm_proj = CRS.from_epsg(4326)
+
+        if 'hydro' in ds:
+            wmodel_file = weather_model_file
+        else:
+            wmodel_file = _convert_NetCDF(weather_model_file=weather_model_file,
+                                          lat_lon_bounds=SNWE,
+                                          weather_model_output_dir='/tmp',
+                                          weather_model=weather_model)
+    
+    # get heights
+    with xarray.load_dataset(wmodel_file) as ds:
+        wm_levels = ds.z.values
+    
+    if height_levels is None:
+        height_levels = wm_levels
+
+
+    #TODO: expose this as library function
+    ds = _get_delays_on_cube(dt, wmodel_file, wm_proj, SNWE, height_levels, crs)
+
+    
+    return ds
+
+
+def get_output_xygrid(SNWE: tuple[float, float, float, float], 
+                      cube_spacing_m: float=500, 
+                      dst_epsg:int=4326):
+    """Reads the SNWE bounding box in lat/lon and creates the output grid.
+    
+    Parameters
+    ----------
+    SNWE: tuple[float, float, float, float]
+        The bounding box in lat/lon coordinates
+    cube_spacing_m: float
+        The output grid spacing in meters
+    dest_epsg:
+        The output grid epsg code
+
+    Returns
[…]
