+++ src/dolphin/workflows/displacement.py
@@ -84,13 +84,13 @@ def run(
     else:
         grouped_amp_mean_files = defaultdict(list)
 
-    if len(cfg.correction_options.troposphere_files) > 0:
-        grouped_tropo_files = group_by_date(
-            cfg.correction_options.troposphere_files,
-            file_date_fmt=cfg.correction_options.tropo_date_fmt,
-        )
-    else:
-        grouped_tropo_files = None
+    # if len(cfg.correction_options.troposphere_files) > 0:
+    #     grouped_tropo_files = group_by_date(
+    #         cfg.correction_options.troposphere_files,
+    #         file_date_fmt=cfg.correction_options.tropo_date_fmt,
+    #     )
+    # else:
+    #     grouped_tropo_files = None
 
     grouped_iono_files: Mapping[tuple[datetime], Sequence[str | PathLike[str]]] = {}
     if len(cfg.correction_options.ionosphere_files) > 0:
@@ -298,14 +298,14 @@ def run(
                 "DEM file is not given, skip estimating tropospheric corrections..."
             )
         else:
-            if grouped_tropo_files:
+            if not cfg.correction_options.troposphere_files is None:
                 tropo_paths = estimate_tropospheric_delay(
                     ifg_file_list=ifg_filenames,
-                    troposphere_files=grouped_tropo_files,
+                    troposphere_files=cfg.correction_options.troposphere_files,
+                    file_date_fmt=cfg.correction_options.tropo_date_fmt,
                     slc_files=grouped_slc_files,
                     geom_files=frame_geometry_files,
                     output_dir=out_dir,
-                    tropo_package=cfg.correction_options.tropo_package,
                     tropo_model=cfg.correction_options.tropo_model,
                     tropo_delay_type=cfg.correction_options.tropo_delay_type,
                     epsg=epsg,
