### 2024-07-02 28e5272
Logging cleanup: only `log_runtime` during workflows
Also, coerce to list so `tqdm` can see the length
during stitching.
