### 2024-07-03 #6927 @scottstanie
> @scottstanie In your paper's summary or statement of need section, can you add a bit about what makes your software important for research? As in, how is it used by researchers, in what line of research, or for what purpose? Something along those lines, please, assuming this is true.

Thanks for the comment @kthyng, I have tried to [clarify the background and motivation in the summary and statement of need.](https://github.com/isce-framework/dolphin/commit/cfa48ee9f7d20d84adba3b127e0a16f63c06ef83)

> Also @scottstanie is there a reason for your license file the way it is (select a license, both open source licenses) or is it an accident over time? If there isn't a reason, we would recommend you choose one.

I've applied the dual Apache-2.0 and BSD-3-Clause license based on the instructions of my institution's Software Release Authority. This work is subject to an Open Source Agreement between my institution and the sponsor agency that governs which license(s) should apply.
