### 2024-07-03 4b9ed85
pass through time series paths to displacement outputs, fix `.unw.tif` naming
