### 2024-07-03 3eff1b3
Fix timeseries suffixes, log only during workflows, fix tqdm stitching bar (#328)
* Logging cleanup: only `log_runtime` during workflows

Also, coerce to list so `tqdm` can see the length
during stitching.

* pass through time series paths to displacement outputs, fix `.unw.tif` naming
