### 2024-07-03 e53ac82
remove numba gpu performance warnings from tests
