+++ tests/test_phase_link_core.py
@@ -12,11 +12,6 @@ GPU_AVAILABLE = gpu_is_available() and os.environ.get("NUMBA_DISABLE_JIT") != "1
 NUM_ACQ = 30
 simulate._seed(1234)
 
-# 'Grid size 49 will likely result in GPU under-utilization due to low occupancy.'
-pytestmark = pytest.mark.filterwarnings(
-    "ignore::numba.core.errors.NumbaPerformanceWarning"
-)
-
 
 @pytest.fixture(scope="module")
 def slc_samples(C_truth):
