### 2024-07-03 3d426e6
Set modulus of compressed SLCs to be real SLC magnitude means (#329)
* Set the amplitude of compressed SLCs to the real mean

Close #251

* remove numba gpu performance warnings from tests

* add a slc compression test with mean check
