### 2024-07-04 1d68bd1
add `band` option to `write_block` and background writers
