### 2024-07-04 b050692
compute and save amplitude dispersion from real SLCs
