### 2024-07-05 0d5ccd4
start adding truncate, use ps to calc
