### 2024-07-05 e3ce1e7
replace `get_log` calls with `logging.getLogger`
