+++ src/dolphin/interpolation.py
@@ -1,12 +1,12 @@
+import logging
+
 import numba
 import numpy as np
 from numpy.typing import ArrayLike
 
-from dolphin._log import get_log
-
 from .similarity import get_circle_idxs
 
-logger = get_log(__name__)
+logger = logging.getLogger(__name__)
 
 
 def interpolate(
