### 2024-07-05 c8277aa
redo `setup_logging`, cleanup
