### 2024-07-05 fd50f14
remove queue for now, log to a file by default
