### 2024-07-05 66d7c07
add `truncate_mantissa` option to `repack` functions
