### 2024-07-05 dccdad3
fix test, set upsampled boundary to nan in compress
