### 2024-07-05 62e1bba
Add functions in `ps` to combine amplitude dispersions
These can take older rasters, assuming the `N` was saved, and produce a new mean and dispersion
