+++ tests/test_ps.py
@@ -1,5 +1,6 @@
 import numpy as np
 import pytest
+from numpy.testing import assert_allclose
 from osgeo import gdal
 
 import dolphin.ps
@@ -105,3 +106,66 @@ def test_multilook_ps_file(tmp_path, vrt_stack):
     )
     assert io.get_raster_dtype(output_ps_file) == np.uint8
     assert io.get_raster_dtype(output_amp_disp_file) == np.float32
+
+
+def test_compute_combined_amplitude_means():
+    # Test basic functionality
+    amplitudes = np.array([[[1.0, 1.0], [1.0, 1.0]], [[6.0, 6.0], [11.0, 21.0]]])
+    N = np.array([9, 1])
+    expected = np.array([[1.5, 1.5], [2.0, 3.0]])
+    result = dolphin.ps.combine_amplitude_means(amplitudes, N)
+    assert_allclose(result, expected, rtol=1e-5)
+
+    #  Test with multiple groups
+    amplitudes = np.random.randn(10, 2, 2) ** 2
+    amp_mean_1 = np.mean(amplitudes[:5], axis=0)
+    amp_mean_2 = np.mean(amplitudes[5:9], axis=0)
+    amp_3 = amplitudes[9]
+    result = dolphin.ps.combine_amplitude_means(
+        np.stack([amp_mean_1, amp_mean_2, amp_3]), [5, 4, 1]
+    )
+    assert_allclose(result, np.mean(amplitudes, axis=0), rtol=1e-5)
+
+    # Test with all equal weights
+    expected_equal = np.mean(amplitudes, axis=0)
+    result_equal = dolphin.ps.combine_amplitude_means(
+        amplitudes, np.ones(len(amplitudes))
+    )
+    assert_allclose(result_equal, expected_equal, rtol=1e-5)
+
+
+def test_compute_combined_amplitude_dispersions():
+    # Test basic functionality
+
+    amplitudes = np.random.randn(10, 2, 2) ** 2
+
+    amp_mean, amp_disp, _ = dolphin.ps.calc_ps_block(amplitudes)
+
+    N = [5, 4, 1]
+
+    amp_mean_1, amp_disp_1, _ = dolphin.ps.calc_ps_block(amplitudes[:5])
+    amp_mean_2, amp_disp_2, _ = dolphin.ps.calc_ps_block(amplitudes[5:9])
+
+    mean_inputs = np.stack([amp_mean_1, amp_mean_2, amplitudes[9]])
+    # Note: a dispersion of N=1 isn't really defined. we dont use that
+    disp_inputs = np.stack([amp_disp_1, amp_disp_2, np.zeros_like(amplitudes[9])])
+
+    combined_disp, combined_mean = dolphin.ps.combine_amplitude_dispersions(
+        dispersions=disp_inputs, means=mean_inputs, N=N
+    )
+    assert_allclose(combined_disp, amp_disp, rtol=1e-5)
+
+
+def test_single_group():
+    """Test with a group where all N=1 (meaning we passed in just the amplitudes)."""
+    amplitudes = np.random.randn(10, 2, 2) ** 2
+    amp_mean, amp_disp, _ = dolphin.ps.calc_ps_block(amplitudes)
+    N = [1] * len(amplitudes)
+    result = dolphin.ps.combine_amplitude_means(amplitudes, N)
+    assert_allclose(result, amp_mean, rtol=1e-5)
+
+    result_disp, result_mean = dolphin.ps.combine_amplitude_dispersions(
+        np.zeros_like(amplitudes), amplitudes, N
+    )
+    assert_allclose(result_disp, amp_disp, rtol=1e-5)
+    assert_allclose(result_mean, amp_mean, rtol=1e-5)
