### 2024-07-08 4f5af43
use the `reference_idx` from each `ministack`, dont hardcode to 0
