### 2024-07-08 34b2113
remove commented `stack.py` code
