### 2024-07-08 22c0bab
pop the `file` handler if we never specify a log file
