### 2024-07-08 c388bc6
Redo logging to use `dictConfig`, always log to a `.jsonl` file (#330)
* replace `get_log` calls with `logging.getLogger`

* redo `setup_logging`, cleanup

* add json config, add to manifest

* remove queue for now, log to a file by default

* pop the `file` handler if we never specify a log file
