+++ src/dolphin/unwrap/_isce3.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
+import logging
 from os import fspath
 from pathlib import Path
 
 import numpy as np
 
 from dolphin import io
-from dolphin._log import get_log
 from dolphin._types import Filename
 from dolphin.utils import full_suffix
 from dolphin.workflows import UnwrapMethod
@@ -14,7 +14,7 @@ from dolphin.workflows import UnwrapMethod
 from ._constants import CONNCOMP_SUFFIX, DEFAULT_CCL_NODATA
 from ._utils import _redirect_unwrapping_log, _zero_from_mask
 
-logger = get_log(__name__)
+logger = logging.getLogger(__name__)
 
 __all__ = ["unwrap_isce3"]
 
