+++ src/dolphin/workflows/wrapped_phase.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import datetime
+import logging
 from pathlib import Path
 from typing import Optional, Sequence, cast
 
@@ -8,11 +9,13 @@ import numpy as np
 from opera_utils import get_dates, make_nodata_mask
 
 from dolphin import interferogram, ps, stack
-from dolphin._log import get_log, log_runtime
+from dolphin._log import log_runtime, setup_logging
 from dolphin.io import VRTStack
 
 from . import InterferogramNetwork, sequential
 from .config import DisplacementWorkflow
+
+logger = logging.getLogger(__name__)
 
 
 @log_runtime
@@ -50,9 +53,9 @@ def run(
         Path to the created SHP counts file.
 
     """
+    setup_logging(debug=debug, filename=cfg.log_file)
     if tqdm_kwargs is None:
         tqdm_kwargs = {}
-    logger = get_log(name=__name__, debug=debug)
     work_dir = cfg.work_directory
     logger.info("Running wrapped phase estimation in %s", work_dir)
 
