### 2024-07-08 ec7727f
Save each ministack's amplitude dispersion for incremental PS/DS changes (#331)
* add `band` option to `write_block` and background writers

* compute and save amplitude dispersion from real SLCs

* start adding truncate, use ps to calc

* repack the amp dispersion rasters for better compression

* add `truncate_mantissa` option to `repack` functions

* fix test, set upsampled boundary to nan in compress

* temp ignore the `repack_rasters`, change incoming
