### 2024-07-08 e4aaf8d
remove `Example` from docstring
