### 2024-07-08 dbead18
update the changelog for this release
