+++ CHANGELOG.md
@@ -1,30 +1,30 @@
 # Changelog
 
-# [v0.20.0](https://github.com/isce-framework/dolphin/compare/v0.19.0...v0.20.0)
+## [0.20.0](https://github.com/isce-framework/dolphin/compare/v0.19.0...v0.20.0)
 
-**Added**
+### Added
 - Functions in `ps` to combine amplitude dispersions from older rasters
 - `truncate_mantissa` option to `repack` functions for simple compression
 - `band` option to `write_block` and background writers
 - Option to run `merge_dates` and `estimate_interferometric_correlations` with `thread_map` for parallel processing
 - Baseline lag option for "STBAS" phase linking inversion
 
-**Changed**
+### Changed
 - Logging now uses `dictConfig` and always logs JSON to a file for the Displacement workflow
 - Set modulus of compressed SLCs to be real SLC magnitude means
 - Updated Docker requirements and specfile
 - Delete intermediate unwrapping `scratchdir`s by default
 
-**Fixed**
+### Fixed
 - `use_max_ps` would occasionally fail with certain stride/window configurations
 - Unwrapped phase files did not always contain the right geographic metadata
 - Filenames in the `timeseries/` folder were wrong
 - Set upsampled boundary to `nan` in `compress`
 - Unwrapped file output path
 
-# [v0.19.0](https://github.com/isce-framework/dolphin/compare/v0.18.0...v0.19.0) - 2024-06-21
+## [0.19.0](https://github.com/isce-framework/dolphin/compare/v0.18.0...v0.19.0) - 2024-06-21
 
-**Added**
+### Added
 - `filtering` module for filtering out long wavelength signals from the unwrapped phase
 - `baseline` module for computing the perpendicular baseline. Initial version has logic for OPERA CSLCs, uses `isce3`
 - Interface only for 3D unwrapping
@@ -33,38 +33,38 @@
 - Eigenvalue solver speedups of 3-9x
 - Initial version of 3D unwrapping using `spurt`
 
-**Removed**
+### Removed
 - the KL-divergence SHP estimator has been removed. GLRT is recommended instead.
 
-**Fixed**
+### Fixed
 - `reproject_bounds` uses the `rasterio` version, which densifies points along edges for a more accurate bounding box
 - The output SHP rasters now output 0 if there was no valid input data
 - Logic for filling PS pixels, with and without setting the amplitudes to be the original SLC amplitudes
 - `ReferencePointError` during the displacement workflow now will fall back look only at the `condition_file` (i.e. choose the point with highest temporal coherence by default)
 
-**Changed**
+### Changed
 - The configuration options for unwrapping have been refactored. Options unique to each unwrapper are grouped into subclasses.
   - Note that older `dolphin_config.yaml` files will error after this refactor.
 - Unweighted time series inversion will make one batch call, providing a large speedup over the `vmap` version for weighted least squares
 
-# [v0.18.0](https://github.com/isce-framework/dolphin/compare/v0.17.0...v0.18.0) - 2024-05-07
+## [0.18.0](https://github.com/isce-framework/dolphin/compare/v0.17.0...v0.18.0) - 2024-05-07
 
-**Added**
+### Added
 - `dolphin timeseries` command line tool for inverting unwrapped interferogram network and estimating velocity
 
-**Fixed**
+### Fixed
 - Parse the file names correctly to find compressed SLCs and read dates based on production file naming convention
 
 
[…]
