### 2024-07-08 9d6a676
remove commented `N` shape check
