### 2024-07-08 c8a5297
add latex documentation
