### 2024-07-08 d89f13e
remove helper function for var combination
