### 2024-07-08 4168608
fix amp mean function name change
