### 2024-07-08 8bba665
`ps`: functions for combining dispersions/means (#332)
* add `band` option to `write_block` and background writers

* compute and save amplitude dispersion from real SLCs

* start adding truncate, use ps to calc

* repack the amp dispersion rasters for better compression

* add `truncate_mantissa` option to `repack` functions

* fix test, set upsampled boundary to nan in compress

* Add functions in `ps` to combine amplitude dispersions

These can take older rasters, assuming the `N` was saved, and produce a new mean and dispersion

* add make the truncate into `round`

* remove `Example` from
[…]
