+++ src/dolphin/ps.py
@@ -6,7 +6,7 @@ import logging
 import shutil
 import warnings
 from pathlib import Path
-from typing import Optional
+from typing import Optional, Sequence
 
 import numpy as np
 from numpy.typing import ArrayLike
@@ -24,8 +24,7 @@ NODATA_VALUES = {"ps": 255, "amp_dispersion": 0.0, "amp_mean": 0.0}
 
 FILE_DTYPES = {"ps": np.uint8, "amp_dispersion": np.float32, "amp_mean": np.float32}
 _EXTRA_COMPRESSION = {
-    "max_error": 0.005,
-    "compression_type": "lerc_deflate",
+    "significant_bits": 10,
     "predictor": 3,
 }
 REPACK_OPTIONS = {
@@ -170,7 +169,7 @@ def create_ps(
     logger.info("Repacking PS rasters for better compression")
     for fn, opt in zip(file_list, REPACK_OPTIONS.values()):
         # Repack to a temp, then overwrite
-        repack_raster(Path(fn), output_dir=None, **opt)  # type: ignore[arg-type]
+        repack_raster(Path(fn), output_dir=None, **opt)
     logger.info("Finished writing out PS files")
 
 
@@ -358,3 +357,133 @@ def multilook_ps_files(
             nodata=NODATA_VALUES["amp_dispersion"],
         )
     return ps_out_path, amp_disp_out_path
+
+
+def combine_means(means: ArrayLike, N: ArrayLike) -> np.ndarray:
+    r"""Compute the combined mean from multiple `mu_i` values.
+
+    This function calculates the weighted average of amplitudes based on the
+    number of original data points (N) that went into each mean.
+
+    Parameters
+    ----------
+    means : ArrayLike
+        A 3D array of mean values.
+        Shape: (n_images, rows, cols)
+    N : np.ndarray
+        A list/array of weights indicating the number of original images.
+        Shape: (depth,)
+
+    Returns
+    -------
+    np.ndarray
+        The combined mean.
+        Shape: (height, width)
+
+    Notes
+    -----
+    Both input arrays are expected to have the same shape.
+    The operation is performed along axis=0.
+
+    The combined mean is calculated as
+
+    \begin{equation}
+        E[X] = \frac{\sum_i N_i\mu_i}{\sum_i N_i}
+    \end{equation}
+
+    """
+    N = np.asarray(N)
+    if N.shape[0] != means.shape[0]:
+        raise ValueError("Size of N must match the number of images in means.")
+    if N.ndim == 1:
+        N = N[:, None, None]
+
+    weighted_sum = np.sum(means * N, axis=0)
+    total_N = np.sum(N, axis=0)
+
+    return weighted_sum / total_N
+
+
+def combine_amplitude_dispersions(
+    dispersions: np.ndarray, means: np.ndarray, N: ArrayLike | Sequence
+) -> np.ndarray:
+    r"""Compute the combined amplitude dispersion from multiple groups.
+
+    Given several ADs where difference numbers of images, N, went in,
+    the function computes a weighted mean/variance to calculate the combined AD.
+
+    Parameters
+    ----------
+    dispersions : np.ndarray
+        A 3D array of amplitude dispersion values for each group.
+        Shape: (depth, height, width)
+    means : np.ndarray
+        A 3D array of mean values for each group.
+        Shape: (depth, height, width)
+    N : np.ndarray
+        An array sample sizes for each group.
+        Shape: (depth, )
+
+    Returns
+    -------
+    np.ndarray
+        The combined amplitude dispersion.
+        Shape: (height, width)
+    np.ndarray
[…]
