### 2024-07-09 264747a
Add a `reference_idx` arg to `compress`
