### 2024-07-09 ac1c66c
update `MiniStackPlanner` to accept manual reference indexes
