### 2024-07-09 5217fcf
start passing through logic to account for manual reference changeovers
