+++ src/dolphin/workflows/config/_common.py
@@ -146,6 +146,15 @@ class InterferogramNetwork(BaseModel, extra="forbid"):
             " interferograms to form."
         ),
     )
+    reference_dates: Optional[list[datetime]] = Field(
+        None,
+        description=(
+            "Specify a multi-reference network: list the date/datetimes in the input"
+            " SLC stack which you would like to make single-reference interferograms for."
+            " This allows you to specify more than one network for, e.g., a large earthquake"
+            " event."
+        ),
+    )
 
     # validation
     @model_validator(mode="after")
