+++ src/dolphin/workflows/single.py
@@ -189,8 +189,6 @@ def run_wrapped_phase_single(
             amp_stack=amp_stack,
             method=shp_method,
         )
-        # Run the phase linking process on the current ministack
-        reference_idx = max(0, first_real_slc_idx - 1)
         try:
             pl_output = run_phase_linking(
                 cur_data,
@@ -198,7 +196,9 @@ def run_wrapped_phase_single(
                 strides=strides_tup,
                 use_evd=use_evd,
                 beta=beta,
-                reference_idx=reference_idx,
+                # TODO: Do we really need to pass in `reference_idx` to here?
+                # We can (and do) re-reference, if needed, from the result
+                # reference_idx=reference_idx,
                 nodata_mask=nodata_mask[in_rows, in_cols],
                 ps_mask=ps_mask[in_rows, in_cols],
                 neighbor_arrays=neighbor_arrays,
@@ -243,6 +243,7 @@ def run_wrapped_phase_single(
             cur_data[first_real_slc_idx:, in_trim_rows, in_trim_cols],
             pl_output.cpx_phase[first_real_slc_idx:, out_trim_rows, out_trim_cols],
             slc_mean=cur_data_mean,
+            reference_idx=reference_idx,
         )
         # TODO: truncate
 
