### 2024-07-10 cb886c6
get the linked phase part working with manual references
