+++ src/dolphin/workflows/single.py
@@ -43,6 +43,7 @@ def run_wrapped_phase_single(
     half_window: dict,
     strides: Optional[dict] = None,
     reference_idx: int = 0,
+    compressed_reference_idx: int = 0,
     beta: float = 0.00,
     use_evd: bool = False,
     mask_file: Optional[Filename] = None,
@@ -60,9 +61,14 @@ def run_wrapped_phase_single(
 
     Output files will all be placed in the provided `output_folder`.
     """
-    # TODO: extract common stuff between here and sequential
+    # Unclear now if we will ever need to pass through the `reference_idx`
+    # For now, make the user explicitly use "compressed_reference_idx"
+    if reference_idx:
+        logger.warning(f"reference_idx set to {reference_idx} will not be used.")
+
     if strides is None:
         strides = {"x": 1, "y": 1}
+    # TODO: extract common stuff between here and sequential
     strides_tup = Strides(y=strides["y"], x=strides["x"])
     half_window_tup = HalfWindow(y=half_window["y"], x=half_window["x"])
     output_folder = Path(output_folder)
@@ -243,7 +249,7 @@ def run_wrapped_phase_single(
             cur_data[first_real_slc_idx:, in_trim_rows, in_trim_cols],
             pl_output.cpx_phase[first_real_slc_idx:, out_trim_rows, out_trim_cols],
             slc_mean=cur_data_mean,
-            reference_idx=reference_idx,
+            reference_idx=compressed_reference_idx,
         )
         # TODO: truncate
 
