### 2024-07-10 a64add2
Edit inline citation formatting
