### 2024-07-10 #6927 @scottstanie
> Sent out the first round of reviewer invitations today.
>
> Quick formatting nit on the paper -- I think you want to remove the `[]` around the Mirzaee 2023 reference on Line 25, to drop the `()` in the PDF. I'll do a more thorough copyediting scan after the review process, but wanted to note this one in case I didn't see it again later 😄

👍  I edited the inline citations that I saw to remove the parens
