### 2024-07-10 6dcf3da
Use `numcodecs` implementation of rounding, export `round_mantissa`
