### 2024-07-10 87febfa
rename `significant_bits` to `keep_bits` for clarity
