+++ src/dolphin/ps.py
@@ -24,7 +24,7 @@ NODATA_VALUES = {"ps": 255, "amp_dispersion": 0.0, "amp_mean": 0.0}
 
 FILE_DTYPES = {"ps": np.uint8, "amp_dispersion": np.float32, "amp_mean": np.float32}
 _EXTRA_COMPRESSION = {
-    "significant_bits": 10,
+    "keep_bits": 10,
     "predictor": 3,
 }
 REPACK_OPTIONS = {
