### 2024-07-10 19a8f2b
Use `numcodecs` implementation of rounding, export `round_mantissa` (#335)
* Use `numcodecs` implementation of rounding, export `round_mantissa`

* rename `significant_bits` to `keep_bits` for clarity
