### 2024-07-10 c854a83
`mkdir` for log directory before configuring
