### 2024-07-13 7dc2980
do scratchdir in init
