+++ src/dolphin/unwrap/_unwrap_3d.py
@@ -43,11 +43,11 @@ def unwrap_spurt(
         _mask = io.load_gdal(mask_filename)
 
     gen_settings = GeneralSettings(
-        output_folder=output_path, **options.general_settings.model_dump(by_alias=True)
+        output_folder=output_path,
+        intermediate_folder=scratchdir,
+        **options.general_settings.model_dump(by_alias=True),
     )
 
-    if scratchdir is not None:
-        gen_settings.intermediate_folder = scratchdir
     tile_settings = TilerSettings(**options.tiler_settings.model_dump(by_alias=True))
     slv_settings = SolverSettings(**options.solver_settings.model_dump(by_alias=True))
     mrg_settings = MergerSettings(**options.merger_settings.model_dump(by_alias=True))
