### 2024-07-15 1b90a01
fix `is` for pre-commit
