### 2024-07-15 cfddece
Revert "fix `is` for pre-commit"
This reverts commit 1b90a01f58d15296a0eedfede99c2a8c1a3e2d19.
