+++ src/dolphin/interpolation.py
@@ -2,9 +2,9 @@ import logging
 
 import numba
 import numpy as np
+import pandas as pd
 from numpy.typing import ArrayLike
 from scipy.interpolate import interp1d
-import pandas as pd
 
 from .similarity import get_circle_idxs
 
@@ -139,56 +139,73 @@ def _interp_loop(
             interpolated_ifg[r0, c0] = np.abs(ifg[r0, c0]) * np.exp(1j * np.angle(csum))
 
 
+def interpolate_along_axis(oldCoord, newCoord, data, axis=2):
+    """Interpolate an array of 3-D data along one axis. This function.
 
-def interpolate_along_axis(oldCoord, newCoord, data, axis=2, pad=False):
-    '''
-    Interpolate an array of 3-D data along one axis. This function
     assumes that the x-coordinate increases monotonically.
-    '''
+    """
     if oldCoord.ndim > 1:
         stackedData = np.concatenate([oldCoord, data, newCoord], axis=axis)
-        out = np.apply_along_axis(interpVector, axis=axis, arr=stackedData, Nx=oldCoord.shape[axis])
+        out = np.apply_along_axis(
+            interp_vector, axis=axis, arr=stackedData, Nx=oldCoord.shape[axis]
+        )
     else:
-        out = np.apply_along_axis(interpV, axis=axis, arr=data, old_x=oldCoord, new_x=newCoord,
-                                  left=np.nan, right=np.nan)
+        out = np.apply_along_axis(
+            interp_v,
+            axis=axis,
+            arr=data,
+            old_x=oldCoord,
+            new_x=newCoord,
+            left=np.nan,
+            right=np.nan,
+        )
 
     return out
 
-def interpVector(vec, Nx):
-    '''
-    Interpolate data from a single vector containing the original
+
+def interp_vector(vec, Nx):
+    """Interpolate data from a single vector containing the original.
+
     x, the original y, and the new x, in that order. Nx tells the
     number of original x-points.
-    '''
+    """
     x = vec[:Nx]
-    y = vec[Nx:2 * Nx]
-    xnew = vec[2 * Nx:]
+    y = vec[Nx : 2 * Nx]
+    xnew = vec[2 * Nx :]
     f = interp1d(x, y, bounds_error=False, copy=False, assume_sorted=True)
     return f(xnew)
 
-def interpV(y, old_x, new_x, left=None, right=None, period=None):
-    '''
-    Rearrange np.interp's arguments
-    '''
+
+def interp_v(y, old_x, new_x, left=None, right=None, period=None):
+    """Rearrange np.interp's arguments."""
     return np.interp(new_x, old_x, y, left=left, right=right, period=period)
 
 
-def fillna3D(array, axis=-1, fill_value=0.):
-    '''
-    This function fills in NaNs in 3D arrays, specifically using the nearest non-nan value
-    for "low" NaNs and 0s for "high" NaNs. 
+def fillna3d(array: np.ndarray, axis: int = -1, fill_value: float = 0.0) -> np.ndarray:
+    """Fill in NaNs in 3D arrays using the nearest non-NaN value for "low" NaNs.
 
-    Arguments: 
-        array   - 3D array, where the last axis is the "z" dimension
-    
-    Returns: 
-        3D array with low NaNs filled as nearest neighbors and high NaNs filled as 0s
-    '''
+    and a specified fill value (default 0.0) for "high" NaNs.
 
+    Parameters
+    ----------
+    array : np.ndarray
+        3D array, where the last axis is the "z" dimension.
+    axis : int, optional
+        The axis along which to fill values. Default is -1 (the last axis).
+    fill_value : float, optional
[…]
