+++ src/dolphin/atmosphere/_netcdf.py
@@ -1,177 +1,222 @@
 from __future__ import annotations
 
-import os
 import datetime
-import numpy as np # type: ignore
-from pyproj import CRS, Transformer # type: ignore
-from typing import Union, Optional, Iterable
-import xarray # type: ignore
-from scipy.interpolate import RegularGridInterpolator as Interpolator # type: ignore
-from collections import defaultdict
 import itertools
-try:
-    import multiprocessing as mp
-except ImportError:
-    mp = None
-    
+import logging
+import multiprocessing as mp
+from collections import defaultdict
+from pathlib import Path
+from typing import Iterable, Optional, Union
 
-from dolphin._log import get_log
-from opera_utils import reproject_bounds # type: ignore
+import numpy as np
+import xarray
+from pyproj import CRS, Transformer
+from scipy.interpolate import RegularGridInterpolator as Interpolator
+
 from dolphin._types import Filename, TropoModel
 
-logger = get_log(__name__)
+logger = logging.getLogger(__name__)
 
 ###########
 # Mostly inherited from RAiDER
 
 __all__ = ["delay_from_netcdf"]
 
+
 def delay_from_netcdf(
-        dt: datetime.datetime,
-        weather_model_file: str,
-        SNWE: tuple[float, float, float, float],
-        height_levels: Optional[list[float]],
-        out_proj: Union[int, str]=4326,
-        weather_model: TropoModel=TropoModel.ECMWF,
-    ):
+    dt: datetime.datetime,
+    weather_model_file: Filename,
+    SNWE: tuple[float, float, float, float],
+    height_levels: Optional[list[float]],
+    out_proj: Union[int, str] = 4326,
+    weather_model: TropoModel = TropoModel.ECMWF,
+):
     """Calculate integrated delays on query points.
 
     Parameters
     ----------
-    dt: Datetime                 
+    dt: Datetime
         Datetime object for determining when to calculate delays
-    weather_model_File: string  
+    weather_model_file: string
         Name of the NETCDF file containing a pre-processed weather model
-    aoi: AOI object             
-        AOI object
-    height_levels: list         
-        (optional) list of height levels on which to calculate delays. Only needed for cube generation.
-    out_proj: int,str           
+    SNWE: tuple[float]
+        Bounding box in lat/lon
+    height_levels: list
+        (optional) list of height levels on which to calculate delays.
+        Only needed for cube generation.
+    out_proj: int,str
         (optional) EPSG code for output projection
+    weather_model: TropoModel
+        weather model  (ECMWF or HRES)
 
     Returns
     -------
-        xarray Dataset *or* ndarrays:  
-            wet and hydrostatic delays at the grid nodes / query points.
+    xarray Dataset *or* ndarrays:
+        wet and hydrostatic delays at the grid nodes / query points.
+
     """
-    crs = CRS(out_proj)
-    
     # Load CRS from weather model file
     with xarray.load_dataset(weather_model_file) as ds:
-        try:
-            wm_proj = CRS.from_wkt(ds['proj'].attrs['crs_wkt'])
-        except KeyError:
-            logger.warning("WARNING: I can't find a CRS in the weather model file, so I will assume you are using WGS84")
-            wm_proj = CRS.from_epsg(4326)
 
-        if 'hydro' in ds:
+        if "hydro" in ds:
             wmodel_file = weather_model_file
         else:
[…]
