+++ src/dolphin/atmosphere/__init__.py
@@ -1,5 +1,5 @@
 """Atmospheric corrections."""
 
+from ._netcdf import *
 from .ionosphere import *
 from .troposphere import *
-from ._netcdf import *
