### 2024-07-15 1bc12d6
remove unnecessary lines
