+++ src/dolphin/atmosphere/_utils.py
@@ -42,79 +42,6 @@ def find_svp(t):
     return svp.astype(np.float32)
 
 
-# def get_mapping(proj):
-#     '''Get CF-complient projection information from a proj'''
-#     # In case of WGS-84 lat/lon, keep it simple
-#     if proj.to_epsg()==4326:
-#         return 'WGS84'
-#     else:
-#         return proj.to_wkt()
-
-
-# def checkContainment_raw(path_wm_raw,
-#                         ll_bounds,
-#                         buffer_deg: float = 1e-5) -> bool:
-#     """"
-#     Checks if existing raw weather model contains
-#     requested ll_bounds
-
-#     Args:
-#     ----------
-#     path_wm_raw : path to downloaded, uncropped weather model file
-#     ll_bounds: an array of floats (SNWE) demarcating bbox of targets
-#     buffer_deg : float
-#         For x-translates for extents that lie outside of world bounding box,
-#         this ensures that translates have some overlap. The default is 1e-5
-#         or ~11.1 meters.
-
-#     Returns:
-#     -------
-#     bool
-#         True if weather model contains bounding box of OutLats and outLons
-#         and False otherwise.
-#     """
-#     import xarray as xr
-#     ymin_input, ymax_input, xmin_input, xmax_input = ll_bounds
-#     input_box   = box(xmin_input, ymin_input, xmax_input, ymax_input)
-
-#     with xr.open_dataset(path_wm_raw) as ds:
-#         try:
-#             ymin, ymax = ds.latitude.min(), ds.latitude.max()
-#             xmin, xmax = ds.longitude.min(), ds.longitude.max()
-#         except:
-#             ymin, ymax = ds.y.min(), ds.y.max()
-#             xmin, xmax = ds.x.min(), ds.x.max()
-
-#         xmin, xmax = np.mod(np.array([xmin, xmax])+180, 360) - 180
-#         weather_model_box = box(xmin, ymin, xmax, ymax)
-
-#     world_box  = box(-180, -90, 180, 90)
-
-#     # Logger
-#     input_box_str = [f'{x:1.2f}' for x in [xmin_input, ymin_input,
-#                                             xmax_input, ymax_input]]
-#     weath_box_str = [f'{x:1.2f}' for x in [xmin, ymin, xmax, ymax]]
-
-#     weath_box_str = ', '.join(weath_box_str)
-#     input_box_str = ', '.join(input_box_str)
-
-
-#     # If the bounding box goes beyond the normal world extents
-#     # Look at two x-translates, buffer them, and take their union.
-#     if not world_box.contains(weather_model_box):
-#         logger.info('Considering x-translates of weather model +/-360 '
-#                     'as bounding box outside of -180, -90, 180, 90')
-#         translates = [weather_model_box.buffer(buffer_deg),
-#                         translate(weather_model_box,
-#                                 xoff=360).buffer(buffer_deg),
-#                         translate(weather_model_box,
-#                                 xoff=-360).buffer(buffer_deg)
-#                         ]
-#         weather_model_box = unary_union(translates)
-
-#     return weather_model_box.contains(input_box)
-
-
 def calcgeoh(
     lnsp: np.ndarray,
     t: np.ndarray,
