### 2024-07-15 2b5adfc
[pre-commit.ci] pre-commit autoupdate (#336)
* [pre-commit.ci] pre-commit autoupdate

updates:
- [github.com/astral-sh/ruff-pre-commit: v0.5.1 → v0.5.2](https://github.com/astral-sh/ruff-pre-commit/compare/v0.5.1...v0.5.2)

* fix `is` for pre-commit

* Revert "fix `is` for pre-commit"

This reverts commit 1b90a01f58d15296a0eedfede99c2a8c1a3e2d19.

* test only fix

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Scott Staniewicz <scott.j.staniewicz@jpl.nasa.gov>
