### 2024-07-16 #6997 @editorialbot
[REVIEW]: Dolphin: A Python package for large-scale InSAR PS/DS processing
**Submitting author:** @scottstanie (<a href="http://orcid.org/0000-0002-3055-5731">Scott Staniewicz</a>)
**Repository:** https://github.com/isce-framework/dolphin
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.32.0
