### 2024-07-16 609290d
avoid internal `opera_utils` function
