### 2024-07-16 1e40c7c
review suggestions for optional dependencies
