### 2024-07-16 8dd0149
Netcdf support (#337)
* add netcdf support for troposphere

* reformat for pre-commit

* remove unnecessary lines

* avoid internal `opera_utils` function

* add xarray for atmosphere test reqs

* review suggestions for optional dependencies

* remove comment

---------

Co-authored-by: mirzaees <smirzaee@aurora.jpl.nasa.gov>
Co-authored-by: Scott Staniewicz <scott.stanie@gmail.com>
Co-authored-by: Scott Staniewicz <scott.j.staniewicz@jpl.nasa.gov>
