+++ src/dolphin/atmosphere/troposphere.py
@@ -9,7 +9,7 @@ from typing import Mapping, Sequence
 
 import numpy as np
 import opera_utils as oput
-from opera_utils import get_dates
+from opera_utils import get_dates, group_by_date
 from osgeo import gdal
 from rasterio.crs import CRS
 from rasterio.warp import transform_bounds
@@ -18,6 +18,8 @@ from scipy.interpolate import RegularGridInterpolator
 from dolphin import io
 from dolphin._types import Bbox, Filename, TropoModel, TropoType
 from dolphin.utils import format_date_pair
+
+from ._netcdf import delay_from_netcdf, group_netcdf_by_date
 
 logger = logging.getLogger(__name__)
 
@@ -54,17 +56,17 @@ class DelayParams:
     wavelength: float
     """Radar wavelength."""
 
-    tropo_model: str
+    tropo_model: TropoModel
     """Model used for tropospheric correction."""
 
     delay_type: str
     """Type of tropospheric delay."""
 
-    reference_file: Sequence[Filename]
-    """Sequence of filenames used as reference files."""
+    reference_file: Filename
+    """Filename used as reference files."""
 
-    secondary_file: Sequence[Filename]
-    """Sequence of filenames used as secondary files."""
+    secondary_file: Filename
+    """Filename used as secondary files."""
 
     interferogram: str
     """Identifier for the interferogram."""
@@ -79,10 +81,10 @@ class DelayParams:
 def estimate_tropospheric_delay(
     ifg_file_list: Sequence[Path],
     slc_files: Mapping[tuple[datetime.datetime], Sequence[Filename]],
-    troposphere_files: Mapping[tuple[datetime.datetime], Sequence[Filename]],
+    troposphere_files: Sequence[Filename],
     geom_files: dict[str, Path],
     output_dir: Path,
-    tropo_package: str,
+    file_date_fmt: str,
     tropo_model: TropoModel,
     tropo_delay_type: TropoType,
     epsg: int,
@@ -96,15 +98,16 @@ def estimate_tropospheric_delay(
         List of interferogram files.
     slc_files : Mapping[tuple[datetime.datetime], Sequence[Filename]]
         Dictionary of SLC files indexed by date.
-    troposphere_files : Mapping[tuple[datetime.datetime], Sequence[Filename]]
-        Dictionary of troposphere files indexed by date.
+    troposphere_files : Sequence[Filename]
+        list of troposphere files.
     geom_files : dict[str, Path]
         Dictionary of geometry files with height and incidence angle, or
         los_east and los_north.
     output_dir : Path
         Output directory.
-    tropo_package : str
-        Troposphere processing package ('pyaps' or 'raider').
+    file_date_fmt : str
+        Format of the date in the filename.
+        Default is [dolphin.DEFAULT_DATETIME_FORMAT][]
     tropo_model : TropoModel
         Tropospheric model (ERA5, HRES, ...).
     tropo_delay_type : TropoType
@@ -141,15 +144,20 @@ def estimate_tropospheric_delay(
 
     wavelength = oput.get_radar_wavelength(one_of_slcs)
 
-    tropo_run = compute_pyaps if tropo_package.lower() == "pyaps" else compute_raider
+    delay_type = tropo_delay_type.value
 
-    # comb is short for the summation of wet and dry components
-    if (tropo_delay_type.value == "hydrostatic") and (tropo_package == "raider"):
-        delay_type = "hydro"
-    elif (tropo_delay_type.value == "hydrostatic") and (tropo_package == "pyaps"):
-        delay_type = "dry"
+    if str(troposphere_files[0]).endswith(".nc"):
+        tropo_files = group_netcdf_by_date(troposphere_files)
[…]
