+++ pyproject.toml
@@ -96,7 +96,7 @@ ignore = [
 exclude = ["scripts"]
 
 [tool.ruff.lint.per-file-ignores]
-"**/__init__.py" = ["F403"]
+"**/__init__.py" = ["F401", "F403"]
 "tests/**" = ["D", "N", "PTH"]
 "benchmarks/**" = ["D", "N", "PTH", "RUF", "ARG002"]
 

[…]
