### 2024-07-16 533a364
chore: vendor mdx_bib to avoid direct url in `docs/requirements.txt`
