+++ mdx_bib.py
@@ -1,0 +1,265 @@
+import re
+import string
+from collections import Counter, OrderedDict
+from xml.etree import ElementTree as etree
+
+from markdown.extensions import Extension
+from markdown.inlinepatterns import Pattern
+from markdown.preprocessors import Preprocessor
+from markdown.treeprocessors import Treeprocessor
+from pybtex.database.input import bibtex
+from pybtex.exceptions import PybtexError
+
+BRACKET_RE = re.compile(r"\[([^\[]+)\]")
+CITE_RE = re.compile(r"@(\w+)")
+DEF_RE = re.compile(r"\A {0,3}\[@(\w+)\]:\s*(.*)")
+INDENT_RE = re.compile(r"\A\t| {4}(.*)")
+
+CITATION_RE = r"@(\w+)"
+
+
+class Bibliography(object):
+    """Keep track of document references and citations for exporting."""
+
+    def __init__(self, extension, bibtex_file, order):
+        self.extension = extension
+        self.order = order
+
+        self.citations = OrderedDict()
+        self.references = {}
+
+        if bibtex_file:
+            try:
+                parser = bibtex.Parser()
+                self.bibsource = parser.parse_file(bibtex_file).entries
+                self.labels = {
+                    id: self.formatCitation(self.bibsource[id]) for id in self.bibsource
+                }
+                for value, occurrences in Counter(self.labels.values()).items():
+                    if occurrences > 1:
+                        for xkey, xvalue in self.labels.items():
+                            i = 0
+                            if xvalue == value:
+                                self.labels[xkey] = (
+                                    f"{xvalue}{string.ascii_lowercase[i]}"
+                                )
+                                i += 1
+
+            except PybtexError:
+                print("Error loading bibtex file")
+                self.bibsource = {}
+                self.labels = {}
+        else:
+            self.bibsource = {}
+
+    def addCitation(self, citekey):
+        self.citations[citekey] = self.citations.get(citekey, 0) + 1
+
+    def setReference(self, citekey, reference):
+        self.references[citekey] = reference
+
+    def citationID(self, citekey):
+        return "cite-" + citekey
+
+    def referenceID(self, citekey):
+        return "ref-" + citekey
+
+    def formatAuthor(self, author):
+        out = f"{author.last_names[0]} {author.first_names[0][0]}."
+        if author.middle_names:
+            out += f"{author.middle_names[0][0]}."
+        return out.replace("{", "").replace("}", "")
+
+    def formatAuthorSurname(self, author):
+        out = author.last_names[0]
+        return out.replace("{", "").replace("}", "")
+
+    def formatReference(self, ref):
+        author_list = list(map(self.formatAuthor, ref.persons["author"]))
+
+        if len(author_list) == 1:
+            authors = author_list[0]
+        else:
+            authors = ", ".join(author_list[:-1])
+            authors += f" and {author_list[-1]}"
+
+        # Harvard style
+        # Surname, Initial, ... and Last_Surname,
+        # Initial, Year. Title. Journal, Volume(Issue), pages. doi.
+
+        title = ref.fields["title"].replace("{", "").replace("}", "")
+        journal = ref.fields.get("journal", "")
+        volume = ref.fields.get("volume", "")
+        issue = ref.fields.get("issue", "")
+        year = ref.fields.get("year")
+        pages = ref.fields.get("pages")
+        doi = ref.fields.get("doi")
+
+        reference = f"<p>{authors}, {year}. {title}."
+        if journal:
+            reference += f" <i>{journal}</i>."
+            if volume:
+                reference += f" <i>{volume}</i>"
+            if issue:
+                reference += f"({issue})"
+            if pages:
+                reference += f", pp.{pages}"
[…]
