### 2024-07-16 ce65da4
rename spatial/temporal to match `spurt` for less confusion
