### 2024-07-16 c6ad708
add mdx_bib reqs
