### 2024-07-16 9e6effa
Minor changes in `spurt` config to match its naming onventions (#342)
* do scratchdir in init

* rename spatial/temporal to match `spurt` for less confusion
