### 2024-07-16 3275c37
chore: vendor `mdx_bib` to avoid direct url in `docs/requirements.txt` (#340)
* chore: vendor mdx_bib to avoid direct url in `docs/requirements.txt`

* move into docs folder

* add git source

* add mdx_bib reqs

* ignore mdx code
