+++ pyproject.toml
@@ -99,6 +99,7 @@ exclude = ["scripts"]
 "**/__init__.py" = ["F401", "F403"]
 "tests/**" = ["D", "N", "PTH"]
 "benchmarks/**" = ["D", "N", "PTH", "RUF", "ARG002"]
+"docs/mdx_bib.py" = ["ALL"] # ignore the vendored code
 
 [tool.ruff.lint.flake8-unused-arguments]
 # Whether to allow unused variadic arguments, like *args and **kwargs.
