### 2024-07-16 5f139b3
[skip ci] Add note on per-year scaling
