### 2024-07-16 07c289e
`timeseries`: output velocities in units per year (#341)
* `timeseries`: output velocities in units per year

* [skip ci] Add note on per-year scaling
