### 2024-07-17 fa01875
show_versions: remove hardcoding, add separate unwrapper info
