### 2024-07-17 d919f02
use `round_mantissa` instead of `lerc` compression
not as standard a gdal option as I thought
