### 2024-07-17 98951a1
use `re.split` instead
