### 2024-07-17 fb0abcf
Let `show_versions` read `importlib.metadata`, use `keep_bits`, not `lerc` (#344)
* show_versions: remove hardcoding, add separate unwrapper info

* use `round_mantissa` instead of `lerc` compression

not as standard a gdal option as I thought

* add spaces for reqs?

* use `re.split` instead

* remove bad assert
