### 2024-07-17 7b64e2e
fix spurt unit tests, dont pass `None` to scratchdir
