### 2024-07-17 7a500bd
Add whirlwind unwrapper (#345)
* Add whirlwind unwrapper

* Add whirlwind to CI environment
