### 2024-07-17 e456247
install spurt for github
