+++ .github/workflows/test-build-push.yml
@@ -43,12 +43,13 @@ jobs:
               - conda-forge
       - name: Install
         run: |
-          pip install --no-deps "snaphu>=0.4.0" "opera-utils>=0.4.1" git+https://github.com/isce-framework/tophu@main
+          pip install  "snaphu>=0.4.0" "opera-utils>=0.4.1" \
+            git+https://github.com/isce-framework/tophu@main \
+            git+https://github.com/isce-framework/spurt@main
           pip install --no-deps .
       - name: Install test dependencies
         run: |
           micromamba install -f tests/requirements.txt -c conda-forge
-          pip install --no-deps git+https://github.com/isce-framework/tophu@main
       - name: Enable numba boundscheck for better error catching
         run: |
           echo "NUMBA_BOUNDSCHECK=1" >> $GITHUB_ENV
