### 2024-07-17 43b42c5
Remove whirlwind from the CI environment
The whirlwind repo is still private within isce-framework and therefore
requires a token to access via GitHub Actions. According to
https://github.com/orgs/community/discussions/46566, the one-time-use
`GITHUB_TOKEN` that's supplied to each GHA job doesn't have permissions
to access private repos within the same organization. Instead, we would
need to setup a PAT or some other less-secure solution.

With this in mind, I think my preferred approach is to disable whirlwind
test(s) in CI until the repo becomes publicly accessible.
