### 2024-07-18 0de4bb7
Add whirlwind to Docker image and CI environments (#349)
* Add whirlwind to Docker image

Updates the Dockerfile to install whirlwind.

In order to build whirlwind from source, we need a C++20 compiler, so I
updated the specfile using (from the repo's top-level directory)

```
$ ./docker/create-lockfile --file conda-env.yml --pkgs cxx-compiler > docker/specfile.txt
```

* Add whirlwind to CI environment
