### 2024-07-18 b45d95d
specify pytest module in warning
