### 2024-07-18 1b5c42c
Fix spurt tests (#346)
* fix spurt unit tests, dont pass `None` to scratchdir

* install spurt for github

* add missing `run | ` line

* no deps for `tophu`

* ignore fork warning

* specify pytest module in warning
