+++ .github/workflows/test-build-push.yml
@@ -52,7 +52,6 @@ jobs:
       - name: Install test dependencies
         run: |
           micromamba install -f tests/requirements.txt -c conda-forge
-          pip install --no-deps git+https://github.com/isce-framework/tophu@main
       - name: Enable numba boundscheck for better error catching
         run: |
           echo "NUMBA_BOUNDSCHECK=1" >> $GITHUB_ENV
