### 2024-07-19 a45a450
Rename `create_temporal_average` to just `create_average`. Use in `sequential` instead of `gdal_calc`
