### 2024-07-19 1842cf1
Allow `create_average` to work with masked arrays using a `mask_average_func`
