+++ src/dolphin/timeseries.py
@@ -593,6 +593,7 @@ def create_average(
     block_shape: tuple[int, int] = (512, 512),
     num_threads: int = 5,
     average_func: Callable[[ArrayLike, int], np.ndarray] = np.nanmean,
+    mask_average_func: Callable[[ArrayLike, int], np.ndarray] = np.any,
     read_masked: bool = False,
 ) -> None:
     """Average all images in `file_list` to create a 2D image in `output_file`.
@@ -612,6 +613,11 @@ def create_average(
     average_func : Callable[[ArrayLike, int], np.ndarray], optional
         The function to use to average the images.
         Default is `np.nanmean`, which calls `np.nanmean(arr, axis=0)` on each block.
+    mask_average_func : Callable[[ArrayLike, int], np.ndarray], optional
+        If `read_masked` is true, the function to use to average the masks.
+        Default is `np.any`, which calls `np.any(masks, axis=0)` on each block
+        and masks *any* pixel that is masked in one of the images.
+        Use `np.all` to have more valid pixels (a smaller masked region).
     read_masked : bool, optional
         If True, reads the data as a masked array based on the rasters' nodata values.
         Default is False.
@@ -623,6 +629,9 @@ def create_average(
     ) -> tuple[slice, slice, np.ndarray]:
         chunk = readers[0][:, rows, cols]
         out_chunk = average_func(chunk, 0), rows, cols
+        if isinstance(chunk, np.ma.MaskedArray):
+            mask = mask_average_func(chunk.mask, 0)
+            out_chunk = np.ma.MaskedArray(data=out_chunk, mask=mask)
         return out_chunk
 
     writer = io.BackgroundRasterWriter(output_file, like_filename=file_list[0])
