### 2024-07-19 e96760b
add weighted average
