+++ src/dolphin/unwrap/_unwrap_3d.py
@@ -30,12 +30,12 @@ def unwrap_spurt(
         MergerSettings,
         SolverSettings,
         TilerSettings,
+        compute_phasediff_deciles,
+        get_bulk_offsets,
+        get_tiles,
+        merge_tiles,
+        unwrap_tiles,
     )
-    from spurt.workflows.emcf._bulk_offset import get_bulk_offsets
-    from spurt.workflows.emcf._merge import merge_tiles
-    from spurt.workflows.emcf._overlap import compute_phasediff_deciles
-    from spurt.workflows.emcf._tiling import get_tiles
-    from spurt.workflows.emcf._unwrap import unwrap_tiles
 
     if cor_filenames is not None:
         assert len(ifg_filenames) == len(cor_filenames)
