### 2024-07-19 5c550c6
add test for `create_average`
