### 2024-07-19 e281990
add weighted tests
