### 2024-07-19 857430b
bring back protocol for weighted version
