+++ src/dolphin/workflows/sequential.py
@@ -125,11 +125,9 @@ def run_wrapped_phase_sequential(
     # Average the temporal coherence files in each ministack
     full_span = ministack_planner.real_slc_date_range_str
     output_temp_coh_file = output_folder / f"temporal_coherence_average_{full_span}.tif"
-    output_shp_count_file = output_folder / f"shp_counts_average_{full_span}.tif"
-
-    # we can pass the list of files to gdal_calc, which interprets it
-    # as a multi-band file
     create_average(temp_coh_files, output_file=output_temp_coh_file)
+
+    output_shp_count_file = output_folder / f"shp_counts_average_{full_span}.tif"
     create_average(shp_count_files, output_file=output_shp_count_file)
 
     # Combine the separate SLC output lists into a single list
