### 2024-07-21 d6a0635
first version for local cookoff
