+++ scripts/unwrap_cookoff.py
@@ -1,0 +1,236 @@
+import argparse
+import logging
+import time
+from pathlib import Path
+
+from pydantic import BaseModel, Field
+
+from dolphin._log import setup_logging
+from dolphin.utils import get_max_memory_usage
+from dolphin.workflows import (
+    SnaphuOptions,
+    SpurtOptions,
+    UnwrapMethod,
+    UnwrapOptions,
+    unwrapping,
+)
+
+logger = logging.getLogger("dolphin")
+
+
+NO_TILES = (1, 1)
+
+
+# @dataclass(config=ConfigDict(arbitrary_types_allowed=True))
+class CookoffRun(BaseModel):
+    """Configuration for one unwrapping test run."""
+
+    path: Path
+    unwrapper: UnwrapMethod
+    ntiles: tuple[int, int] = NO_TILES
+    run_interpolation: bool = False
+    run_goldstein: bool = False
+    n_parallel_jobs: int = 1
+    spurt_cost: SpurtOptions = Field(default_factory=SpurtOptions)
+    spurt_coh_threshold: float = 0.6
+    tile_overlap: tuple[int, int] = (400, 400)
+
+    @property
+    def name(self) -> str:
+        """Name of the unwrapping configuration."""
+        n = str(self.unwrapper.value)
+        if self.ntiles != NO_TILES:
+            n += f"_tiling_{self.ntiles[0]}-{self.ntiles[1]}"
+        if self.run_interpolation:
+            n += "_interpolated"
+        if self.run_goldstein:
+            n += "_goldstein"
+        if self.unwrapper == UnwrapMethod.SPURT:
+            n += f"_threshold_{self.spurt_coh_threshold}"
+            cost = self.spurt_cost.solver_settings.t_cost_type
+            if cost != "constant":
+                n += f"_cost_{cost}"
+
+        return n
+
+    @property
+    def work_dir(self) -> Path:
+        """Working directory to store results."""
+        return self.path / self.name
+
+    def get_options(self) -> UnwrapOptions:
+        """Convert the specified options to an `UnwrapOptions` object."""
+        common_options = {
+            "run_unwrap": True,
+            "run_interpolation": self.run_interpolation,
+            "run_goldstein": self.run_goldstein,
+            "unwrap_method": self.unwrapper,
+            "n_parallel_jobs": self.n_parallel_jobs,
+        }
+
+        if self.unwrapper == UnwrapMethod.SNAPHU:
+            snaphu_options = SnaphuOptions(
+                ntiles=self.ntiles,
+                tile_overlap=self.tile_overlap,
+            )
+            opts = UnwrapOptions(**common_options, snaphu_options=snaphu_options)
+        elif self.unwrapper == UnwrapMethod.SPURT:
+            spurt_options = SpurtOptions(
+                temporal_coherence_threshold=self.spurt_coh_threshold,
+                solver_settings={
+                    "solver_settings": self.spurt_cost,
+                },
+            )
+            opts = UnwrapOptions(**common_options, spurt_options=spurt_options)
+        else:
+            opts = UnwrapOptions(**common_options)
+        opts._directory = self.work_dir
+        return opts
+
+
+def create_unwrap_options(
+    base_dir: Path = Path(),
+    n_parallel_jobs: int = 1,
+    ntiles: tuple[int, int] = (2, 2),
+    tile_overlap: tuple[int, int] = (300, 300),
+) -> list[CookoffRun]:
+    """Generate all unwrapping configurations to test."""
+    base_dir.mkdir(exist_ok=True)
+
+    common = {
+        "path": base_dir,
+        "n_parallel_jobs": n_parallel_jobs,
+        "tile_overlap": tile_overlap,
+    }
[…]
