### 2024-07-22 eaf83e3
add `parse_log_files` function to assemble results
