### 2024-07-22 7655f9c
Make network formation `staticmethod`s public (#355)
They can be generally useful for computing indexes without forming actual VRT files
