### 2024-07-22 38e5d38
add `--run` for subselecting cookoff runs
