### 2024-07-22 0db1793
Set the `final_arr` nodata to be same as original ifg mask
closes #353
