### 2024-07-22 6211c17
add more tests for combinations of goldstein/interp
