+++ src/dolphin/unwrap/_unwrap.py
@@ -412,7 +412,7 @@ def unwrap(
             "Transferring ambiguity numbers from filtered/interpolated"
             f" ifg {unwrapper_unw_filename}"
         )
-        unw_arr = io.load_gdal(unwrapper_unw_filename)
+        unw_arr = io.load_gdal(unwrapper_unw_filename, masked=True).filled(unw_nodata)
 
         final_arr = np.angle(ifg) + (unw_arr - np.angle(modified_ifg))
         final_arr[ifg.mask] = unw_nodata
