### 2024-07-22 a6c9ce1
`_log`: add option to redirect other loggers to same json file
