### 2024-07-22 eedb605
add `extra` options for logging to json
