### 2024-07-22 015bf84
Set the `final_arr` nodata to be same as original ifg mask (#356)
* Set the `final_arr` nodata to be same as original ifg mask

closes #353

* move unwrapped, not ifg

* add more tests for combinations of goldstein/interp
