### 2024-07-22 98471f6
Fix filtering ambiguity transfer for interp/goldstein
