### 2024-07-23 86e1bf1
Add shrinkage to `nlooks` for whirlwind only
