### 2024-07-23 ea8d958
prepend `scratch-` to scratch dir to avoid confusing filename
