### 2024-07-23 816bae1
Interleave GTiffs by band (#363)
