### 2024-07-24 21b1074
Update `links_per_batch` to match cli of spurt
