### 2024-07-29 ee8724b
pull mask/numpy conversion to `masking.py`, remove middle pixel check
