### 2024-07-29 a862d52
add tests for masking
