### 2024-07-29 50e1e2c
Move mask loading function into `masking.py` (#365)
* pull mask/numpy conversion to `masking.py`, remove middle pixel check

* fix GPU readme link

* dont use private `opera-utils` import

* add tests for masking
