+++ README.md
@@ -1,16 +1,18 @@
 # dolphin
+
 [![Pytest and build docker image](https://github.com/isce-framework/dolphin/actions/workflows/test-build-push.yml/badge.svg?branch=main)](https://github.com/isce-framework/dolphin/actions/workflows/test-build-push.yml)
 [![pre-commit.ci status](https://results.pre-commit.ci/badge/github/isce-framework/dolphin/main.svg)](https://results.pre-commit.ci/latest/github/isce-framework/dolphin/main)
 [![Conda-Forge][conda-badge]][conda-link]
-<!-- [![PyPI version][pypi-version]][pypi-link] -->
-<!-- [![PyPI platforms][pypi-platforms]][pypi-link] -->
+[![PyPI platforms][pypi-platforms]][pypi-link]
+[![GitHub Discussion][github-discussions-badge]][github-discussions-link]
 
 <!-- prettier-ignore-start -->
 [conda-badge]:              https://img.shields.io/conda/vn/conda-forge/dolphin
 [conda-link]:               https://github.com/conda-forge/dolphin-feedstock
 [pypi-link]:                https://pypi.org/project/dolphin/
 [pypi-platforms]:           https://img.shields.io/pypi/pyversions/dolphin
-[pypi-version]:             https://img.shields.io/pypi/v/dolphin
+[github-discussions-badge]: https://img.shields.io/static/v1?label=Discussions&message=Ask&color=blue&logo=github
+[github-discussions-link]:  https://github.com/isce-framework/dolphin/discussions
 <!-- prettier-ignore-end -->
 
 High resolution wrapped phase estimation for InSAR using combined PS/DS processing.
