### 2024-07-30 7dfac12
add `input_options.wavelength`, auto-recognize OPERA CSLC-S1
