### 2024-07-30 bd5c581
add `wavelength` as arguments in `timeseries` functions
