### 2024-07-30 0bab1bc
add `units` and `description` options to writing rasters
