### 2024-07-30 4b8c0c7
Skip correlation estimate for files that exist  (#368)
* Skip correlation estimate for files that exist

* export `setup_logging`
