### 2024-07-30 340684c
Clip `nlooks` to a max of 20 for `whirlwind` (#369)
