### 2024-07-31 3b40a93
simplify `filter_long_wavelength` to use `np.pad(..., mode="reflect")`
