+++ src/dolphin/filtering.py
@@ -1,9 +1,9 @@
 import numpy as np
 from numpy.typing import ArrayLike
-from scipy import ndimage
+from scipy import fft, ndimage
 
 
-def filter_long_wavelength(
+def filter_long_wavelength_old(
     unwrapped_phase: ArrayLike,
     correlation: ArrayLike,
     mask_cutoff: float = 0.5,
@@ -38,120 +38,40 @@ def filter_long_wavelength(
 
     """
     nrow, ncol = correlation.shape
-
-    mask = (correlation > mask_cutoff).astype("bool")
-
-    # Create Boolean mask for Zero-filled boundary area to be False
-    # and the rest to be True
-    mask_boundary = ~(correlation == 0).astype("bool")
+    mask = correlation > mask_cutoff
+    mask_boundary = ~(correlation == 0)
 
     plane = fit_ramp_plane(unwrapped_phase, mask)
 
-    # Replace masked out pixels with the ramp plane
-    unw_ifg_interp = np.copy(unwrapped_phase)
-    unw_ifg_interp[~mask * mask_boundary] = plane[~mask * mask_boundary]
+    unw_ifg_interp = np.where(mask & mask_boundary, unwrapped_phase, plane)
 
-    # Copy the edge pixels for the boundary area before filling them by reflection
-    EV_fill = np.copy(unw_ifg_interp)
+    # Pad the array with edge values
+    pad_width = ((nrow // 2, nrow // 2), (ncol // 2, ncol // 2))
+    # See here for illustration of `mode="reflect"`
+    # https://scikit-image.org/docs/stable/auto_examples/transform/plot_edge_modes.html#interpolation-edge-modes
+    padded = np.pad(unw_ifg_interp, pad_width, mode="reflect")
 
-    Xdata = np.argwhere(mask)  # Get indices of non-NaN & masked pixels
-    NW = Xdata[np.argmin(Xdata[:, 0])]  # Get indices of upper left corner pixel
-    SE = Xdata[np.argmax(Xdata[:, 0])]  # Get indices of lower right corner pixel
-    SW = Xdata[np.argmin(Xdata[:, 1])]  # Get indices of lower left corner pixel
-    NE = Xdata[np.argmax(Xdata[:, 1])]  # Get indices of upper left corner pixel
+    sigma = _compute_filter_sigma(wavelength_cutoff, pixel_spacing, cutoff_value=0.5)
+    # Apply Gaussian filter
+    input_ = fft.fft2(padded, workers=6)
+    result = ndimage.fourier_gaussian(input_, sigma=sigma)
+    result = np.fft.ifft2(result).real.astype(unwrapped_phase.dtype)
 
-    for k in range(NW[1], NE[1] + 1):
-        n_zeros = np.count_nonzero(
-            unw_ifg_interp[0 : NE[0] + 1, k] == 0
-        )  # count zeros in North direction
-        if n_zeros == 0:
-            continue
-        EV_fill[0:n_zeros, k] = EV_fill[n_zeros, k]
-    for k in range(SW[1], SE[1] + 1):
-        n_zeros = np.count_nonzero(
-            unw_ifg_interp[SW[0] + 1 :, k] == 0
-        )  # count zeros in South direction
-        if n_zeros == 0:
-            continue
-        EV_fill[-n_zeros:, k] = EV_fill[-n_zeros - 1, k]
-    for k in range(NW[0], SW[0] + 1):
-        n_zeros = np.count_nonzero(
[…]
