+++ src/dolphin/filtering.py
@@ -6,7 +6,8 @@ from scipy import fft, ndimage
 def filter_long_wavelength(
     unwrapped_phase: ArrayLike,
     correlation: ArrayLike,
-    mask_cutoff: float = 0.5,
+    correlation_cutoff: float = 0.5,
+    connected_component_labels: ArrayLike | None = None,
     wavelength_cutoff: float = 50 * 1e3,
     pixel_spacing: float = 30,
     workers: int = 1,
@@ -19,11 +20,14 @@ def filter_long_wavelength(
         Unwrapped interferogram phase to filter.
     correlation : Arraylike, 2D
         Array of interferometric correlation from 0 to 1.
-    mask_cutoff: float
+    correlation_cutoff : float
         Threshold to use on `correlation` so that pixels where
-        `correlation[i, j] > mask_cutoff` are used and the rest are ignored.
+        `correlation[i, j] > correlation_cutoff` are used and the rest are ignored.
         The default is 0.5.
-    wavelength_cutoff: float
+    connected_component_labels : ArrayLike, optional
+        Integer labels of connected components from the unwrapped interferogram.
+        If provided, ignores pixels with label 0.
+    wavelength_cutoff : float
         Spatial wavelength threshold to filter the unwrapped phase.
         Signals with wavelength longer than 'wavelength_cutoff' are filtered out.
         The default is 50*1e3 (m).
@@ -42,29 +46,45 @@ def filter_long_wavelength(
 
     """
     nrow, ncol = correlation.shape
-    mask = correlation > mask_cutoff
-    mask_boundary = ~(correlation == 0)
+    good_pixel_mask = correlation > correlation_cutoff
+    non_boundary_mask = ~(correlation == 0)
 
-    plane = fit_ramp_plane(unwrapped_phase, mask)
+    if connected_component_labels is not None:
+        good_pixel_mask = good_pixel_mask & (connected_component_labels != 0)
 
-    unw_ifg_interp = np.where(mask & mask_boundary, unwrapped_phase, plane)
+    unw0 = np.nan_to_num(unwrapped_phase)
+    unw_valid_mask = unw0 != 0
+    total_valid_mask = unw_valid_mask & good_pixel_mask
+
+    # Shift to be zero mean, then reset the borders to 0
+    offset = np.mean(unw0[total_valid_mask])
+    unw0 -= offset
+    unw0[~total_valid_mask] = 0
+
+    plane = fit_ramp_plane(unw0, total_valid_mask)
+
+    unw_ifg_interp = np.where(total_valid_mask, unw0, plane)
 
     # Pad the array with edge values
-    pad_width = ((nrow // 2, nrow // 2), (ncol // 2, ncol // 2))
+    pad_rows = nrow // 4
+    pad_cols = ncol // 4
     # See here for illustration of `mode="reflect"`
     # https://scikit-image.org/docs/stable/auto_examples/transform/plot_edge_modes.html#interpolation-edge-modes
-    padded = np.pad(unw_ifg_interp, pad_width, mode="reflect")
+    padded = np.pad(
+        unw_ifg_interp, ((pad_rows, pad_rows), (pad_cols, pad_cols)), mode="reflect"
+    )
 
     sigma = _compute_filter_sigma(wavelength_cutoff, pixel_spacing, cutoff_value=0.5)
     # Apply Gaussian filter
     input_ = fft.fft2(padded, workers=workers)
     result = ndimage.fourier_gaussian(input_, sigma=sigma)
-    result = np.fft.ifft2(result).real.astype(unwrapped_phase.dtype)
+    # Make sure to only take the real part (ifft returns complex)
[…]
