### 2024-07-31 #6997 @scottstanie
Thanks for the quick response @margauxmouchene ! Sorry for the slow response, but I should be able to get to the rest of the comments in a few days.

For the first couple points,

> can you please clarify the paper's authors contributions to the code:

I wanted to check-  did you want a clarification added to the paper? or just here for the review? The short answer is that Heresh Fattahi made the original `fringe` and has been the project lead for dolphin, and Talib Oliver-Cabrera made/tested [the first version of the `interpolation`](https://github.com/taliboliver/ph_interp) module, but somehow isn't the commit author on github (might have been me who moved it due to odd logistics). The others hadn't been on the algorithm team making `dolphin` code when I wrote the draft in Jan/Feb., but I can revisit the author list with Heresh.

> the tests require snaphu-py, isce3 and tophu to run but they are not listed in tests\requirements.txt

That's a good idea to add `snaphu-py` to the test requirements, but `tophu`/`isce3` are unforunately difficult to install outside of conda, and `tophu` can't be `pip installed` without already having `isce3` in your environment. It's why the [github action is set up with an extra spec](https://github.com/isce-framework/dolphin/blob/340684cdf50d4bf25b1a3051477d13e0cdca867d/.github/workflows/test-build-push.yml#L20-L23) at the moment.

> but the [link](https://dolphin-insar.readthedocs.io/en/latest/getting-started/dolphin-insar.readthedocs.io/gpu-setup.md) is broken

Thank you! That's now fixed [on the live docs](https://dolphin-insar.readthedocs.io/en/latest/getting-started/)

I'll try to respond to the other couple comments ASAP, and a newer test dataset will be added to the walkthrough that should be easily downloadable/runnable on Colab
