### 2024-07-31 e1e532b
replace correlation/conncomp with `bad_pixel_mask`. Deramp result.
