### 2024-07-31 6b16677
Simplify `filter_long_wavelength` to use `np.pad(..., mode="reflect")` (#372)
* simplify `filter_long_wavelength` to use `np.pad(..., mode="reflect")`

* fix naming, add `workers`

* padd less, use scipy `ifft2`, add `connected_component_label` option

* replace correlation/conncomp with `bad_pixel_mask`. Deramp result.

* fix test
