+++ src/dolphin/filtering.py
@@ -1,34 +1,34 @@
 import numpy as np
 from numpy.typing import ArrayLike
-from scipy import ndimage
+from scipy import fft, ndimage
 
 
 def filter_long_wavelength(
     unwrapped_phase: ArrayLike,
-    correlation: ArrayLike,
-    mask_cutoff: float = 0.5,
+    bad_pixel_mask: ArrayLike,
     wavelength_cutoff: float = 50 * 1e3,
     pixel_spacing: float = 30,
+    workers: int = 1,
 ) -> np.ndarray:
     """Filter out signals with spatial wavelength longer than a threshold.
 
     Parameters
     ----------
-    unwrapped_phase : np.ndarray, 2D complex array
+    unwrapped_phase : ArrayLike
         Unwrapped interferogram phase to filter.
-    correlation : Arraylike, 2D
-        Array of interferometric correlation from 0 to 1.
-    mask_cutoff: float
-        Threshold to use on `correlation` so that pixels where
-        `correlation[i, j] > mask_cutoff` are used and the rest are ignored.
-        The default is 0.5.
-    wavelength_cutoff: float
+    bad_pixel_mask : ArrayLike
+        Boolean array with same shape as `unwrapped_phase` where `True` indicates a
+        pixel to ignore during ramp fitting
+    wavelength_cutoff : float
         Spatial wavelength threshold to filter the unwrapped phase.
         Signals with wavelength longer than 'wavelength_cutoff' are filtered out.
         The default is 50*1e3 (m).
     pixel_spacing : float
         Pixel spatial spacing. Assume same spacing for x, y axes.
         The default is 30 (m).
+    workers : int
+        Number of `fft` workers to use for `scipy.fft.fft2`.
+        Default is 1.
 
     Returns
     -------
@@ -37,121 +37,53 @@ def filter_long_wavelength(
         longer than a threshold.
 
     """
-    nrow, ncol = correlation.shape
+    good_pixel_mask = ~bad_pixel_mask
 
-    mask = (correlation > mask_cutoff).astype("bool")
+    unw0 = np.nan_to_num(unwrapped_phase)
+    # Take either nan or 0 pixels in `unwrapped_phase` to be nodata
+    nodata_mask = unw0 == 0
+    in_bounds_pixels = ~nodata_mask
 
-    # Create Boolean mask for Zero-filled boundary area to be False
-    # and the rest to be True
-    mask_boundary = ~(correlation == 0).astype("bool")
+    total_valid_mask = in_bounds_pixels & good_pixel_mask
 
-    plane = fit_ramp_plane(unwrapped_phase, mask)
+    plane = fit_ramp_plane(unw0, total_valid_mask)
+    # Remove the plane, setting to 0 where we had no data for the plane fit:
+    unw_ifg_interp = np.where((~nodata_mask & good_pixel_mask), unw0, plane)
 
-    # Replace masked out pixels with the ramp plane
-    unw_ifg_interp = np.copy(unwrapped_phase)
-    unw_ifg_interp[~mask * mask_boundary] = plane[~mask * mask_boundary]
+    # Find the filter `sigma` which gives the correct cutoff in meters
+    sigma = _compute_filter_sigma(wavelength_cutoff, pixel_spacing, cutoff_value=0.5)
 
-    # Copy the edge pixels for the boundary area before filling them by reflection
[…]
