### 2024-07-31 6ffc1ab
Remove the all-zero reference date raster in `timeseries`
