### 2024-07-31 e579860
dont pass N dates, let `create_velocity` form `N-1` dates
