### 2024-07-31 c6766f6
Skip zero raster output by timeseries inversions (#373)
* Remove the all-zero reference date raster in `timeseries`

* fix tests for new shape

* dont pass N dates, let `create_velocity` form `N-1` dates

* wait on `process_blocks` futures and raise their exceptions

they were getting silently ignored during tests before
