### 2024-07-31 b062af8
Save `ReferencePoint` into file in `timeseries/` folder
