### 2024-07-31 41851e0
only compute the reference if needed
