### 2024-07-31 decc8fa
better logging of reference point
