### 2024-07-31 c7c4a80
Save `ReferencePoint` into file in `timeseries/` folder (#374)
* Save `ReferencePoint` into file in `timeseries/` folder

* only compute the reference if needed

* fix cor vrt name typo

* better logging of reference point
