### 2024-07-31 8ff92eb
Use `label` to count labels (#375)
