### 2024-08-01 8ff9963
move reference point selection back to always run if not specified (#376)
