+++ src/dolphin/timeseries.py
@@ -83,18 +83,15 @@ def run(
 
     """
     condition_func = argmax_index if condition == CallFunc.MAX else argmin_index
-
-    def _get_reference():
-        # First we find the reference point for the unwrapped interferograms
-        if reference_point == (-1, -1):
-            return select_reference_point(
-                condition_file=condition_file,
-                output_dir=Path(output_dir),
-                condition_func=condition_func,
-                ccl_file_list=conncomp_paths,
-            )
-        else:
-            return ReferencePoint(row=reference_point[0], col=reference_point[1])
+    if reference_point == (-1, -1):
+        ref = select_reference_point(
+            condition_file=condition_file,
+            output_dir=Path(output_dir),
+            condition_func=condition_func,
+            ccl_file_list=conncomp_paths,
+        )
+    else:
+        ref = ReferencePoint(row=reference_point[0], col=reference_point[1])
 
     Path(output_dir).mkdir(exist_ok=True, parents=True)
 
@@ -111,7 +108,7 @@ def run(
         logger.info("Inverting network of %s unwrapped ifgs", len(unwrapped_paths))
         inverted_phase_paths = invert_unw_network(
             unw_file_list=unwrapped_paths,
-            reference=_get_reference(),
+            reference=ref,
             output_dir=output_dir,
             num_threads=num_threads,
         )
@@ -145,7 +142,7 @@ def run(
         create_velocity(
             unw_file_list=inverted_phase_paths,
             output_file=velocity_file,
-            reference=_get_reference(),
+            reference=ref,
             cor_file_list=cor_file_list,
             cor_threshold=correlation_threshold,
             num_threads=num_threads,
