### 2024-08-01 d6ff5d7
Fix `timeseries` to `mkdir` right away
