### 2024-08-01 ad17ca8
pass through the `ReferencePoint` found from timeseries
