### 2024-08-01 22ec97e
convert ionosphere to meters
