### 2024-08-01 721839f
add subtraction of reference_point to `ionosphere` and `troposphere`
