+++ src/dolphin/atmosphere/ionosphere.py
@@ -16,6 +16,7 @@ from scipy import interpolate
 
 from dolphin import io
 from dolphin._types import Bbox, Filename
+from dolphin.timeseries import ReferencePoint
 from dolphin.utils import format_date_pair
 
 logger = logging.getLogger(__name__)
@@ -35,6 +36,7 @@ def estimate_ionospheric_delay(
     slc_files: Mapping[tuple[datetime.datetime], Sequence[Filename]],
     tec_files: Mapping[tuple[datetime.datetime], Sequence[Filename]],
     geom_files: dict[str, Path],
+    reference_point: ReferencePoint | None,
     output_dir: Path,
     epsg: int,
     bounds: Bbox,
@@ -52,6 +54,9 @@ def estimate_ionospheric_delay(
     geom_files : list[Path]
         Dictionary of geometry files with height and incidence angle, or
         los_east and los_north.
+    reference_point : tuple[int, int], optional
+        Reference point (row, col) used during time series inversion.
+        If not provided, no spatial referencing is performed.
     output_dir : Path
         Output directory.
     epsg : int
@@ -169,6 +174,10 @@ def estimate_ionospheric_delay(
         ifg_iono_range_delay_radians = range_delay_reference - range_delay_secondary
         # Convert to meters, where positive corresponds to motion toward the satellite
         ifg_iono_range_delay = -wavelength / (4 * np.pi) * ifg_iono_range_delay_radians
+
+        if reference_point is not None:
+            ref_row, ref_col = reference_point
+            ifg_iono_range_delay -= ifg_iono_range_delay[ref_row, ref_col]
 
         # Write 2D tropospheric correction layer to disc
         io.write_arr(
