+++ src/dolphin/workflows/displacement.py
@@ -308,6 +308,7 @@ def run(
                     file_date_fmt=cfg.correction_options.tropo_date_fmt,
                     slc_files=grouped_slc_files,
                     geom_files=frame_geometry_files,
+                    reference_point=reference_point,
                     output_dir=out_dir,
                     tropo_model=cfg.correction_options.tropo_model,
                     tropo_delay_type=cfg.correction_options.tropo_delay_type,
@@ -324,6 +325,7 @@ def run(
                 slc_files=grouped_slc_files,
                 tec_files=grouped_iono_files,
                 geom_files=frame_geometry_files,
+                reference_point=reference_point,
                 output_dir=out_dir,
                 epsg=epsg,
                 bounds=out_bounds,
