### 2024-08-01 cafba88
add functions for units and description
