+++ src/dolphin/io/_core.py
@@ -406,6 +406,48 @@ def set_raster_metadata(
     ds = None
 
 
+def set_raster_description(filename: Filename, description: str, band: int = 1):
+    """Set description on a raster band.
+
+    Parameters
+    ----------
+    filename : Filename
+        Path to the file to load.
+    description : str
+        Description to set.
+    band : int, optional
+        Band to set description for. Default is 1.
+
+    """
+    ds = gdal.Open(fspath(filename), gdal.GA_Update)
+    bnd = ds.GetRasterBand(band)
+    bnd.SetDescription(description)
+    bnd.FlushCache()
+    ds = None
+
+
+def set_raster_units(filename: Filename, units: str, band: int | None = None) -> None:
+    """Set units on a raster band.
+
+    Parameters
+    ----------
+    filename : Filename
+        Path to the file to load.
+    units : str
+        Units to set.
+    band : int, optional
+        Band to set units for. Default is None, which sets for all bands.
+
+    """
+    ds = gdal.Open(fspath(filename), gdal.GA_Update)
+    if band is None:
+        bands = range(1, ds.RasterCount + 1)
+    for i in bands:
+        bnd = ds.GetRasterBand(i)
+        bnd.SetUnitType(units)
+        bnd.FlushCache()
+
+
 def rowcol_to_xy(
     row: int,
     col: int,
@@ -467,6 +509,8 @@ def write_arr(
     strides: Optional[dict[str, int]] = None,
     projection: Optional[Any] = None,
     nodata: Optional[float] = None,
+    units: Optional[str] = None,
+    description: Optional[str] = None,
 ):
     """Save an array to `output_name`.
 
@@ -507,6 +551,11 @@ def write_arr(
     nodata : float, optional
         Nodata value to save.
         Default is the nodata of band 1 of `like_filename` (if provided), or None.
+    units : str, optional
+        Units of the data. Default is None.
+        Value is stored in the metadata as "units".
+    description : str, optional
+        Description of the raster bands stored in the metadata.
 
     """
     fi = FileInfo.from_user_inputs(
@@ -552,12 +601,17 @@ def write_arr(
             bnd = ds_out.GetRasterBand(i + 1)
             bnd.WriteArray(arr[i])
 
-    # Set the nodata value for each band
-    if fi.nodata is not None:
-        for i in range(fi.nbands):
-            logger.debug(f"Setting nodata for band {i+1}/{fi.nbands}")
-            bnd = ds_out.GetRasterBand(i + 1)
+    # Set the nodata/units/description for each band
+    for i in range(fi.nbands):
+        logger.debug(f"Setting nodata for band {i+1}/{fi.nbands}")
+        bnd = ds_out.GetRasterBand(i + 1)
+        # Note: right now we're assuming the nodata/units/description
+        if fi.nodata is not None:
             bnd.SetNoDataValue(fi.nodata)
+        if units is not None:
+            bnd.SetUnitType(units)
+        if description is not None:
+            bnd.SetDescription(description)
 
     ds_out.FlushCache()
     ds_out = None
