### 2024-08-01 bcee7b3
add units to `write_arr` for iono/tropo
