### 2024-08-01 26003bf
remove chatting wgs84 warning, add tropo logging
