### 2024-08-01 f8d2aae
use `timeseries_paths` for corrections, not ifg paths
