### 2024-08-01 507b05e
Reference corrections to point, convert to meters (#377)
* pass through the `ReferencePoint` found from timeseries

* convert `troposphere` to output delay in meters

* convert ionosphere to meters

* convert troposphere to meters

* add subtraction of reference_point to `ionosphere` and `troposphere`

* add functions for units and description

* add `get_` functions, export to `__all__`

* add units to `write_arr` for iono/tropo

* remove chatting wgs84 warning, add tropo logging

* use `timeseries_paths` for corrections, not ifg paths

* fix new test
