+++ src/dolphin/workflows/displacement.py
@@ -18,6 +18,7 @@ from tqdm.auto import tqdm
 from dolphin import __version__, io, timeseries, utils
 from dolphin._log import log_runtime, setup_logging
 from dolphin.atmosphere import estimate_ionospheric_delay, estimate_tropospheric_delay
+from dolphin.timeseries import ReferencePoint
 from dolphin.workflows import CallFunc
 
 from . import stitching_bursts, unwrapping, wrapped_phase
@@ -42,6 +43,7 @@ class OutputPaths(NamedTuple):
     timeseries_paths: list[Path] | None
     tropospheric_corrections: list[Path] | None
     ionospheric_corrections: list[Path] | None
+    reference_point: ReferencePoint | None
 
 
 @log_runtime
@@ -224,6 +226,7 @@ def run(
             timeseries_paths=None,
             tropospheric_corrections=None,
             ionospheric_corrections=None,
+            reference_point=None,
         )
 
     row_looks, col_looks = cfg.phase_linking.half_window.to_looks()
@@ -246,7 +249,7 @@ def run(
     if len(unwrapped_paths) > 1 and (ts_opts.run_inversion or ts_opts.run_velocity):
         # the output of run_timeseries is not currently used so pre-commit removes it
         # let's add back if we need it
-        timeseries_paths = timeseries.run(
+        timeseries_paths, reference_point = timeseries.run(
             unwrapped_paths=unwrapped_paths,
             conncomp_paths=conncomp_paths,
             corr_paths=stitched_cor_paths,
@@ -262,6 +265,7 @@ def run(
 
     else:
         timeseries_paths = None
+        reference_point = None
 
     # ##############################################
     # 5. Estimate corrections for each interferogram
@@ -269,8 +273,6 @@ def run(
     tropo_paths: list[Path] | None = None
     iono_paths: list[Path] | None = None
     if len(cfg.correction_options.geometry_files) > 0:
-        stitched_ifg_dir = cfg.interferogram_network._directory
-        ifg_filenames = sorted(Path(stitched_ifg_dir).glob("*.int.tif"))
         out_dir = cfg.work_directory / cfg.correction_options._atm_directory
         out_dir.mkdir(exist_ok=True)
         grouped_slc_files = group_by_date(cfg.cslc_file_list)
@@ -278,13 +280,14 @@ def run(
         # Prepare frame geometry files
         geometry_dir = out_dir / "geometry"
         geometry_dir.mkdir(exist_ok=True)
-        crs = io.get_raster_crs(ifg_filenames[0])
+        assert timeseries_paths is not None
+        crs = io.get_raster_crs(timeseries_paths[0])
         epsg = crs.to_epsg()
-        out_bounds = io.get_raster_bounds(ifg_filenames[0])
+        out_bounds = io.get_raster_bounds(timeseries_paths[0])
         frame_geometry_files = utils.prepare_geometry(
             geometry_dir=geometry_dir,
             geo_files=cfg.correction_options.geometry_files,
-            matching_file=ifg_filenames[0],
+            matching_file=timeseries_paths[0],
             dem_file=cfg.correction_options.dem_file,
             epsg=epsg,
             out_bounds=out_bounds,
@@ -294,16 +297,22 @@ def run(
         # Troposphere
         if "height" not in frame_geometry_files:
             logger.warning(
-                "DEM file is not given, skip estimating tropospheric corrections..."
+                "DEM file is not given, skip estimating tropospheric corrections."
             )
         else:
             if cfg.correction_options.troposphere_files is not None:
+                assert timeseries_paths is not None
+                logger.info(
+                    "Calculating tropospheric corrections for %s files.",
+                    len(timeseries_paths),
+                )
[…]

[…]
