### 2024-08-01 038bcc5
change datetime format (#378)
Co-authored-by: mirzaees <smirzaee@aurora.jpl.nasa.gov>
