### 2024-08-01 5114c74
Still copy over unwrapped to `timeseries/` even with 1 file
