### 2024-08-01 4cbd5b4
Copy over unwrapped to `timeseries/` even with 1 file (#379)
* Still copy over unwrapped to `timeseries/` even with 1 file

* fix comment
