### 2024-08-02 e71ca8e
Format `timestamp` first in logs
