### 2024-08-02 d042b8d
Add `map` option for `DummyProcessPoolExecutor`
