### 2024-08-02 624373c
add a `dolphin filter` CLI for long wavelegnth filtering
