+++ src/dolphin/_cli_filter.py
@@ -1,0 +1,173 @@
+import argparse
+from pathlib import Path
+from typing import TYPE_CHECKING, Any
+
+from dolphin.filtering import filter_long_wavelength
+
+if TYPE_CHECKING:
+    _SubparserType = argparse._SubParsersAction[argparse.ArgumentParser]
+else:
+    _SubparserType = Any
+
+
+def get_parser(subparser=None, subcommand_name="unwrap") -> argparse.ArgumentParser:
+    """Set up the command line interface."""
+    metadata = {
+        "description": "Create a configuration file for a displacement workflow.",
+        "formatter_class": argparse.ArgumentDefaultsHelpFormatter,
+        # https://docs.python.org/3/library/argparse.html#fromfile-prefix-chars
+        "fromfile_prefix_chars": "@",
+    }
+    if subparser:
+        # Used by the subparser to make a nested command line interface
+        parser = subparser.add_parser(subcommand_name, **metadata)
+    else:
+        parser = argparse.ArgumentParser(**metadata)  # type: ignore[arg-type]
+
+    # parser._action_groups.pop()
+    parser.add_argument(
+        "-o",
+        "--output-dir",
+        type=Path,
+        help="Path to output directory to store results. None stores in same location as inputs",
+    )
+    # Get Inputs from the command line
+    inputs = parser.add_argument_group("Input options")
+    inputs.add_argument(
+        "--unw-filenames",
+        nargs=argparse.ONE_OR_MORE,
+        type=Path,
+        help=(
+            "List the paths of unwrapped files to filter. Can pass a newline delimited"
+            " file with @ifg_filelist.txt"
+        ),
+    )
+    inputs.add_argument(
+        "--temporal-coherence-filename",
+        type=Path,
+        help="Optionally, list the path of the temporal coherence to mask.",
+    )
+    inputs.add_argument(
+        "--cor-filenames",
+        nargs=argparse.ZERO_OR_MORE,
+        help="Optionally, list the paths of the correlation files to use for masking",
+    )
+    inputs.add_argument(
+        "--conncomp-filenames",
+        nargs=argparse.ZERO_OR_MORE,
+        help="Optionally, list the paths of the connected component labels for masking",
+    )
+    parser.add_argument(
+        "--wavelength-cutoff",
+        type=float,
+        default=50_000,
+        help="Spatial wavelength_cutoff (in meters) of filter to use.",
+    )
+
+    parser.add_argument(
+        "--max-jobs",
+        type=int,
+        default=1,
+        help="Number of parallel files to filter.",
+    )
+
+    parser.set_defaults(run_func=_run_filter)
+
+    return parser
+
+
+def filter_rasters(
+    unw_filenames: list[Path],
+    temporal_coherence_filename: Path | None,
+    wavelength_cutoff: float = 50_000,
+    output_dir: Path | None = None,
+    num_threads: int = 4,
+    **kwargs,
+):
+    """Recreate and compress a list of raster files.
+
+    Useful for rasters which were created in block and lost
+    the full effect of compression.
+
+    Parameters
+    ----------
+    raster_files : List[Path]
+        List of paths to the input raster files.
+    output_dir : Path, optional
+        Directory to save the processed rasters or None for in-place processing.
+    num_threads : int, optional
+        Number of threads to use (default is 4).
+    keep_bits : int, optional
+        Number of bits to preserve in mantissa. Defaults to None.
+        Lower numbers will truncate the mantissa more and enable more compression.
+
+    Returns
+    -------
+    output_dir : Path
[…]
