+++ src/dolphin/_cli_filter.py
@@ -1,7 +1,15 @@
 import argparse
+import multiprocessing as mp
+from itertools import repeat
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
+import numpy as np
+from numpy.typing import NDArray
+from tqdm.contrib.concurrent import process_map
+
+from dolphin import io
+from dolphin._overviews import Resampling, create_image_overviews
 from dolphin.filtering import filter_long_wavelength
 
 if TYPE_CHECKING:
@@ -13,7 +21,9 @@ else:
 def get_parser(subparser=None, subcommand_name="unwrap") -> argparse.ArgumentParser:
     """Set up the command line interface."""
     metadata = {
-        "description": "Create a configuration file for a displacement workflow.",
+        "description": (
+            "Filter unwrapped interferograms using a long-wavelength filter."
+        ),
         "formatter_class": argparse.ArgumentDefaultsHelpFormatter,
         # https://docs.python.org/3/library/argparse.html#fromfile-prefix-chars
         "fromfile_prefix_chars": "@",
@@ -29,7 +39,10 @@ def get_parser(subparser=None, subcommand_name="unwrap") -> argparse.ArgumentPar
         "-o",
         "--output-dir",
         type=Path,
-        help="Path to output directory to store results. None stores in same location as inputs",
+        help=(
+            "Path to output directory to store results. None stores in same location as"
+            " inputs"
+        ),
     )
     # Get Inputs from the command line
     inputs = parser.add_argument_group("Input options")
@@ -78,82 +91,114 @@ def get_parser(subparser=None, subcommand_name="unwrap") -> argparse.ArgumentPar
 
 def filter_rasters(
     unw_filenames: list[Path],
+    cor_filenames: list[Path] | None,
+    conncomp_filenames: list[Path] | None,
     temporal_coherence_filename: Path | None,
     wavelength_cutoff: float = 50_000,
+    correlation_cutoff: float = 0.5,
     output_dir: Path | None = None,
     num_threads: int = 4,
-    **kwargs,
 ):
-    """Recreate and compress a list of raster files.
+    """Filter a list of unwrapped interferogram files using a long-wavelength filter.
 
-    Useful for rasters which were created in block and lost
-    the full effect of compression.
+    This function applies a spatial filter to remove long-wavelength components from
+    unwrapped interferograms. It can optionall use temporal coherence, correlation, and
+    connected component information for masking.
 
     Parameters
     ----------
-    raster_files : List[Path]
-        List of paths to the input raster files.
-    output_dir : Path, optional
-        Directory to save the processed rasters or None for in-place processing.
+    unw_filenames : list[Path]
+        List of paths to unwrapped interferogram files to be filtered.
+    cor_filenames : list[Path] | None
+        List of paths to correlation files
+        Passing None skips filtering on correlation.
+    conncomp_filenames : list[Path] | None
+        List of paths to connected component files, filters any 0 labelled pixels.
+        Passing None skips filtering on connected component labels.
+    temporal_coherence_filename : Path | None
+        Path to the temporal coherence file for masking.
+        Passing None skips filtering on temporal coherence.
+    wavelength_cutoff : float, optional
+        Spatial wavelength cutoff (in meters) for the filter. Default is 50,000 meters.
+    correlation_cutoff : float, optional
+        Threshold of correlation (if passing `cor_filenames`) to use to ignore pixels
+        during filtering.
[…]
