### 2024-08-02 ea8a3f6
move filtering with io functions into `filtering` away from cli
