### 2024-08-02 3b76779
add check for too large cutoff
