### 2024-08-02 5cc6f4a
Add `dolphin filter` cli to run a long wavelgnth filter on a set of rasters (#382)
* add a `dolphin filter` CLI for long wavelegnth filtering

* optionally pass through cor/conncomp, add overviews

* fix cli test

* move filtering with io functions into `filtering` away from cli

* add test

* avoid set ctx

* add check for too large cutoff
