+++ src/dolphin/filtering.py
@@ -1,5 +1,10 @@
+import multiprocessing as mp
+from concurrent.futures import ProcessPoolExecutor
+from itertools import repeat
+from pathlib import Path
+
 import numpy as np
-from numpy.typing import ArrayLike
+from numpy.typing import ArrayLike, NDArray
 from scipy import fft, ndimage
 
 
@@ -36,9 +41,15 @@ def filter_long_wavelength(
         filtered interferogram that does not contain signals with spatial wavelength
         longer than a threshold.
 
+    Raises
+    ------
+    ValueError
+        If wavelength_cutoff too large for image size/pixel spacing.
+
     """
     good_pixel_mask = ~bad_pixel_mask
 
+    rows, cols = unwrapped_phase.shape
     unw0 = np.nan_to_num(unwrapped_phase)
     # Take either nan or 0 pixels in `unwrapped_phase` to be nodata
     nodata_mask = unw0 == 0
@@ -53,6 +64,10 @@ def filter_long_wavelength(
     # Find the filter `sigma` which gives the correct cutoff in meters
     sigma = _compute_filter_sigma(wavelength_cutoff, pixel_spacing, cutoff_value=0.5)
 
+    if sigma > unw0.shape[0] or sigma > unw0.shape[0]:
+        msg = f"{wavelength_cutoff = } too large for image."
+        msg += f"Shape = {(rows, cols)}, and {pixel_spacing = }"
+        raise ValueError(msg)
     # Pad the array with edge values
     # The padding extends further than the default "radius = 2*sigma + 1",
     # which given specified in `gaussian_filter`
@@ -124,3 +139,123 @@ def fit_ramp_plane(unw_ifg: ArrayLike, mask: ArrayLike) -> np.ndarray:
     plane = np.reshape(X_ @ theta, (nrow, ncol))
 
     return plane
+
+
+def filter_rasters(
+    unw_filenames: list[Path],
+    cor_filenames: list[Path] | None = None,
+    conncomp_filenames: list[Path] | None = None,
+    temporal_coherence_filename: Path | None = None,
+    wavelength_cutoff: float = 50_000,
+    correlation_cutoff: float = 0.5,
+    output_dir: Path | None = None,
+    max_workers: int = 4,
+) -> list[Path]:
+    """Filter a list of unwrapped interferogram files using a long-wavelength filter.
+
+    Remove long-wavelength components from each unwrapped interferogram.
+    It can optionally use temporal coherence, correlation, and connected component
+    information for masking.
+
+    Parameters
+    ----------
+    unw_filenames : list[Path]
+        List of paths to unwrapped interferogram files to be filtered.
+    cor_filenames : list[Path] | None
+        List of paths to correlation files
+        Passing None skips filtering on correlation.
+    conncomp_filenames : list[Path] | None
+        List of paths to connected component files, filters any 0 labelled pixels.
+        Passing None skips filtering on connected component labels.
+    temporal_coherence_filename : Path | None
+        Path to the temporal coherence file for masking.
+        Passing None skips filtering on temporal coherence.
+    wavelength_cutoff : float, optional
+        Spatial wavelength cutoff (in meters) for the filter. Default is 50,000 meters.
+    correlation_cutoff : float, optional
+        Threshold of correlation (if passing `cor_filenames`) to use to ignore pixels
+        during filtering.
+    output_dir : Path | None, optional
+        Directory to save the filtered results.
+        If None, saves in the same location as inputs with .filt.tif extension.
+    max_workers : int, optional
+        Number of parallel images to process. Default is 4.
[…]
