### 2024-08-06 0eabea9
fix setuptools
