+++ .pre-commit-config.yaml
@@ -35,7 +35,7 @@ repos:
     hooks:
       - id: mypy
         additional_dependencies:
-        - types-pkg_resources
+        - types-setuptools
         - types-requests
         - "pydantic>=2.4"
         - types-markdown
