+++ .pre-commit-config.yaml
@@ -20,22 +20,22 @@ repos:
       - id: trailing-whitespace
 
   - repo: https://github.com/psf/black
-    rev: "24.4.2"
+    rev: "24.8.0"
     hooks:
       - id: black
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.5.5
+    rev: v0.5.6
     hooks:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix, --show-fixes]
 
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: "v1.11.0"
+    rev: "v1.11.1"
     hooks:
       - id: mypy
         additional_dependencies:
-        - types-pkg_resources
+        - types-setuptools
         - types-requests
         - "pydantic>=2.4"
         - types-markdown
