### 2024-08-07 7316b6c
Temporarily remove resampling during stitching final crop (#385)
* Temporarily remove resampling during stitching final crop

For now https://github.com/OSGeo/gdal/issues/10536 is causing issues for this step.
We may not ever need to resample here, but we should look at the differences.

* need bilinear to get right sizes with `-tap`
