### 2024-08-07 8cbd34e
add more timeseries logging
