### 2024-08-07 b9f4807
Add ability to mask a subset of input SLCs (#386)
* Add function to create mask from bounding box

* skip nodata in PS iteration

* add bounds to cli

* add more timeseries logging

* refactor `wrapped_phase` to make combined mask

* add new test data
