+++ src/dolphin/masking.py
@@ -1,15 +1,20 @@
 from __future__ import annotations
 
 import logging
+import tempfile
 from enum import IntEnum
+from os import fspath
 from pathlib import Path
 from typing import Optional, Sequence
 
 import numpy as np
 from osgeo import gdal
+from pyproj import CRS, Transformer
+from shapely import to_geojson
+from shapely.geometry import box
 
 from dolphin import io
-from dolphin._types import PathOrStr
+from dolphin._types import Bbox, PathOrStr
 
 gdal.UseExceptions()
 
@@ -158,3 +163,72 @@ def load_mask_as_numpy(mask_file: PathOrStr) -> np.ndarray:
     # invert the mask so Trues are the missing data pixels
     nodata_mask = ~nodata_mask
     return nodata_mask
+
+
+def create_bounds_mask(
+    bounds: Bbox | tuple[float, float, float, float],
+    output_filename: PathOrStr,
+    like_filename: PathOrStr,
+    bounds_epsg: int = 4326,
+    overwrite: bool = False,
+) -> None:
+    """Create a boolean raster mask where 1 is inside the given bounds and 0 is outside.
+
+    Parameters
+    ----------
+    bounds : tuple
+        (min x, min y, max x, max y) of the area of interest
+    like_filename : Filename
+        Reference file to copy the shape, extent, and projection.
+    output_filename : Filename
+        Output filename for the mask
+    bounds_epsg : int, optional
+        EPSG code of the coordinate system of the bounds.
+        Default is 4326 (lat/lon coordinates for the bounds).
+    overwrite : bool, optional
+        Overwrite the output file if it already exists, by default False
+
+    """
+    if Path(output_filename).exists():
+        if not overwrite:
+            logger.info(f"Skipping {output_filename} since it already exists.")
+            return
+        else:
+            logger.info(f"Overwriting {output_filename} since overwrite=True.")
+            Path(output_filename).unlink()
+
+    # Transform bounds if necessary
+    # Geojson default is 4326, and GDAL handles the conversion to, e.g., UTM
+    if bounds_epsg != 4326:
+        transformer = Transformer.from_crs(
+            CRS.from_epsg(bounds_epsg), 4326, always_xy=True
+        )
+        bounds = transformer.transform_bounds(*bounds)
+
+    logger.info(f"Creating mask for bounds {bounds}")
+
+    # Create a polygon from the bounds
+    bounds_poly = box(*bounds)
+
+    # Create the output raster
+    io.write_arr(
+        arr=None,
+        output_name=output_filename,
+        dtype=bool,
+        nbands=1,
+        like_filename=like_filename,
+    )
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        temp_vector_file = Path(tmpdir) / "temp.geojson"
+        with open(temp_vector_file, "w") as f:
+            f.write(to_geojson(bounds_poly))
+
+        # Open the input vector file
+        src_ds = gdal.OpenEx(fspath(temp_vector_file), gdal.OF_VECTOR)
+        dst_ds = gdal.Open(fspath(output_filename), gdal.GA_Update)
+
+        # Now burn in the union of all polygons
+        gdal.Rasterize(dst_ds, src_ds, burnValues=[1])
+
+    logger.info(f"Created {output_filename}")
