+++ src/dolphin/workflows/wrapped_phase.py
@@ -9,7 +9,7 @@ from typing import Optional, Sequence, cast
 import numpy as np
 from opera_utils import get_dates, make_nodata_mask
 
-from dolphin import interferogram, ps, stack
+from dolphin import Bbox, Filename, interferogram, masking, ps, stack
 from dolphin._log import log_runtime, setup_logging
 from dolphin.io import VRTStack
 
@@ -73,6 +73,28 @@ def run(
         outfile=cfg.work_directory / "slc_stack.vrt",
     )
 
+    # Mark any files beginning with "compressed" as compressed
+    is_compressed = ["compressed" in str(f).lower() for f in input_file_list]
+    input_dates = _get_input_dates(
+        input_file_list, is_compressed, cfg.input_options.cslc_date_fmt
+    )
+    reference_date, reference_idx = _get_reference_date_idx(
+        input_file_list, is_compressed, input_dates
+    )
+
+    non_compressed_slcs = [
+        f for f, is_comp in zip(input_file_list, is_compressed) if not is_comp
+    ]
+
+    # Create a mask file from input bounding polygons and/or specified output bounds
+    mask_filename = _get_mask(
+        output_dir=cfg.work_directory,
+        output_bounds=cfg.output_options.bounds,
+        like_filename=vrt_stack.outfile,
+        cslc_file_list=non_compressed_slcs,
+    )
+
+    nodata_mask = masking.load_mask_as_numpy(mask_filename) if mask_filename else None
     # ###############
     # PS selection
     # ###############
@@ -97,6 +119,7 @@ def run(
             output_amp_dispersion_file=cfg.ps_options._amp_dispersion_file,
             amp_dispersion_threshold=cfg.ps_options.amp_dispersion_threshold,
             existing_amp_dispersion_file=existing_disp,
+            nodata_mask=nodata_mask,
             existing_amp_mean_file=existing_amp,
             block_shape=cfg.worker_settings.block_shape,
             **kwargs,
@@ -116,15 +139,6 @@ def run(
     pl_path = cfg.phase_linking._directory
     pl_path.mkdir(parents=True, exist_ok=True)
 
-    # Mark any files beginning with "compressed" as compressed
-    is_compressed = ["compressed" in str(f).lower() for f in input_file_list]
-    input_dates = _get_input_dates(
-        input_file_list, is_compressed, cfg.input_options.cslc_date_fmt
-    )
-    reference_date, reference_idx = _get_reference_date_idx(
-        input_file_list, is_compressed, input_dates
-    )
-
     ministack_planner = stack.MiniStackPlanner(
         file_list=input_file_list,
         dates=input_dates,
@@ -134,19 +148,6 @@ def run(
         reference_date=reference_date,
         reference_idx=reference_idx,
     )
-
-    # Make the nodata mask from the polygons, if we're using OPERA CSLCs
-    non_compressed_slcs = [
-        f for f, is_comp in zip(input_file_list, is_compressed) if not is_comp
-    ]
-    try:
-        nodata_mask_file = cfg.work_directory / "nodata_mask.tif"
-        make_nodata_mask(
-            non_compressed_slcs, out_file=nodata_mask_file, buffer_pixels=200
-        )
-    except Exception as e:
-        logger.warning(f"Could not make nodata mask: {e}")
-        nodata_mask_file = None
 
     phase_linked_slcs = sorted(pl_path.glob("2*.tif"))
     if len(phase_linked_slcs) > 0:
@@ -171,7 +172,7 @@ def run(
                 strides=strides,
                 use_evd=cfg.phase_linking.use_evd,
                 beta=cfg.phase_linking.beta,
-                mask_file=nodata_mask_file,
+                mask_file=mask_filename,
                 ps_mask_file=ps_output,
[…]
