### 2024-08-12 352f556
Clarify "heavy computational demand" review note
