### 2024-08-12 151cbc4
set the default `input_conventions` to be same as `output_conventions`
