### 2024-08-12 #6997 @scottstanie
Thanks for the patience @margauxmouchene !

>dolphin was developed to meet the need for a tool capable of handling the heavy computational demands of large-scale InSAR processing: how does dolphin address this need? what technology/characteristics of this code make it computationally performant for large datasets (such as OPERA's products)?

I've added [some clarifying points to the Overview](https://github.com/isce-framework/dolphin/commit/352f5561e2c33729a21feab118d85dd961d38d87) to try and answer this:

> what is `dolphin`'s place within the `isce-framework` project? how does it relate to `isce2`?

I think it may be a little out of scope to fully explain isce2/3? But here's a rundown, and you can let me know if you think it's worth adding:

`isce2` and `isce3` are broad, framework-like libraries which contain low-level components for building InSAR processing applications. For instance, both libraries have C++ (or Fortran for `isce2`) modules to focus raw data into SAR images, or signal processing algorithms to resample one image onto another for co-registration. `isce3` is full re-write of `isce2` that has been funded by [NISAR](nisar.jpl.nasa.gov/) algorithm development, while `isce2` is now run by the InSAR community.

Dolphin starts with a stack of focused, co-registered SAR images; for example, you can use the outputs of the ["Coregistered SLC stack workflow" of `isce2`](https://github.com/isce-framework/isce2/blob/b231d6e80a9f9d40a0ea95a54f00f57b9a11c058/contrib/stack/topsStack/README.md#42-example-workflow-coregistered-stack-of-slc-with-modified-parameters) and run `dolphin` on these images. However, since there are now Sentinel-1 geocoded SLC images (which are already co-registered) available on ASF, you can get a displacement output in a fraction of the time by downloading these and running `dolphin` (see the walkthrough notebook).

Both libraries have the pieces to run a multi-temporal analysis, but the workflows which put them together for `isce2` are in the user-contributed ["contrib"](https://github.com/isce-framework/isce2/tree/main/contrib/stack) folder. The highest level that isce2 is set up to process is interferogram formation/unwrapping; most users then pass the results onto to [`MintPy`](https://github.com/insarlab/MintPy) to obtain one displacement time series. The time series algorithms that Dolphin runs are newer and more advanced than what exists in isce2/3, though it also has the pieces to do simple multi-looked interferogram formation like `isce2`.

> can you please clarify the paper's authors contributions to the code:

A bit more on the author contributions:

The others who are on the committer list but not here as authors were either added much later to the Algorithm Team or are on separate teams building future validation for `dolphin`.

Also, after a little more internal discussion, [Emre Havazli](https://github.com/ehavazli) was added to the author list as one of the main testers/debuggers of `dolphin` for many case studies over the past year. It was an oversight to not have him on in the first place. He is not one of the committers (similar to Talib), but has been very helpful reporting all errors to me and finding improvements for me to implement.

> is the dataset used to produce Figure 1 in the paper available? otherwise, is there another sample dataset available for testing?

I have updated the walkthrough notebook [in this PR](https://github.com/isce-framework/dolphin/pull/390); it should even be [testable on Colab now](https://colab.research.google.com/github/isce-framework/dolphin/blob/main/docs/notebooks/walkthrough-basic.ipynb)

------------

No problem at all on the later start @McWhity , I was slow to answer @margauxmouchene 's points anyway. Sorry to push changes right after you started to address her points, but it should now be simpler to test the basic walkthrough.
