### 2024-08-12 b91bbb8
Fix plotting/imports for walkthrough (#393)
