### 2024-08-13 #6997 @margauxmouchene
> I've added [some clarifying points to the Overview](https://github.com/isce-framework/dolphin/commit/352f5561e2c33729a21feab118d85dd961d38d87) to try and answer this

Thank you for the clarification. I think this part is clear enough now.

> I think it may be a little out of scope to fully explain isce2/3?

Yes, exactly, that was more for my own understanding and your explanation was super clear. Thank you!

> A bit more on the author contributions

Looks good to me!

>I have updated the walkthrough notebook

I am testing it right now, this is looking good and very user-friendly!

Some errors I ran into:
In the section *Input dataset*
- Code block where ASF username and password are asked: password is assigned to the variable name `password` but called in the following code block as `asf_password`, causing an error to be raised.
- Code block where data is downloaded:

```
with open(url_file, "w") as f:
    f.write("\n".join(best_option.inputs))
```

raises `AttributeError: 'BurstSubsetOption' object has no attribute 'inputs'`

I could not find the urls in `best_option` so I searched the same polygon on the ASF website and put relevant urls in the `urls.txt` file to be able to download the dataset and run the rest of the code without too much change.

I have to stop here and I will be away from my computer for the rest of the week but I'll resume testing/reviewing early next week.

Have a good week!
