### 2024-08-13 #6997 @scottstanie
> I am testing it right now, this is looking good and very user-friendly!
>
> Some errors I ran into: In the section _Input dataset_

Shoot I'm sorry I couldn't get these changes in- I saw that [password typo](https://github.com/isce-framework/dolphin/commit/e6c5ecb3d8ab05eb65dc00e92310722c70015902) after I pushed up the first version, and I made a release of [the `opera-utils` library that does that data search](https://github.com/opera-adt/opera-utils/releases/tag/v0.12.0) yesterday. Maybe I should have just hard-coded the URLs for this demo.

> Also, on the paper:
>
>     1. Line 21: period missing at the end of the sentence
>
>     2. Fattahi et al. ref: missing the mention "AGU Fall Meeting Abstracts", maybe? (suggested citation [here](https://ui.adsabs.harvard.edu/abs/2019AGUFM.G11B0514F/exportcitation))
>
>     3. Rosen et al. ref: missing "IGARSS 2018 - 2018 IEEE International Geoscience and Remote Sensing Symposium", maybe?

I'll double check these citations and make the fixes.

> Have a good week!

you as well!
