### 2024-08-14 #6997 @McWhity
Hi @scottstanie
I run into similar problem as @margauxmouchene.

Here are some first remarks/questions
1. @scottstanie please let me know as soon as you fixed the AttributeError for downloading the test dataset
> raises AttributeError: 'BurstSubsetOption' object has no attribute 'inputs'
2. I assume that the branch "main" is the one that we should test. Am I correct?
3. Regarding the [readme](https://github.com/isce-framework/dolphin/blob/main/README.md)
- Please add the full name for InSAR and PS/DS as especially PS/DS are not known outside of the community (similar to how it is mentioned in the paper manuscript)
4. Regarding the [license](https://github.com/isce-framework/dolphin/blob/main/LICENSE)
- For clarification I would recommend to indicate that the first section is the BSC-3-Clause. You could add the name before “Copyright (c) 2022 California …” similar to the section for the Apache License.
5. Regarding the authors contributions
    Scott J. Staniewicz lead author (511+4)
    Sara Mirzaee (11)
    Geoffrey M. Gunter (4)
    Tablib Oliver-Cabrera (not committing author)
    Heresh Fattahi (lead for dolphin project)
    Not mentioned which is ok based on the low commit count (Ryan Burns (2), Simran Sangha (4), seyeonjeon (1))
- I would suggest to include a credit of author statement in the paper manuscript. Similar to the ones, journals like elsevier (https://www.elsevier.com/researcher/author/policies-and-guidelines/credit-author-statement) or mdpi have. But of course adjusted to the special needs of software development. As I do not know the opinion/options in place by JOSS it might be good if @mikemahoney218 could comment on that to.
6. I would recommend to include the changelog in documentation
7. Regarding the checklist point "Substantial Scholarly Effect"
- If applicable please add within the paper manuscript if the software were used to process data or if results are already published in other papers. Further you could indicate/cite projects were dolphin is currently used.
8. Broken links
- https://dolphin-insar.readthedocs.io/en/latest/getting-started/ penultimate line "notebooks"
- https://dolphin-insar.readthedocs.io/en/latest/notebooks/walkthrough-basic/ under Setup the link to "Getting Started"

More comments will follow
