### 2024-08-14 0da978a
Clarify acronyms in headline of README
