### 2024-08-14 230afc9
clarify first section of license is BSD
