### 2024-08-14 9a4e356
Add link to changelog on docs
