### 2024-08-14 87ef5ed
Fix changelog link using snippets
