+++ docs/index.md
@@ -8,6 +8,6 @@
 1. [Getting started](./getting-started.md)
 1. [Tutorials](tutorials.md)
 1. [Code Reference](reference/summary.md)
-1. [Changelog](../CHANGELOG.md)
+1. [Changelog](changelog.md)
 <!-- 1. [How-To Guides](how-to-guides.md) -->
 <!-- 1. [Background theory](background-theory.md) -->
