### 2024-08-14 #6997 @scottstanie
Thank you @McWhity !

> Hi @scottstanie I run into similar problem as @margauxmouchene.
>
> Here are some first remarks/questions
>
>     1. @scottstanie please let me know as soon as you fixed the AttributeError for downloading the test dataset
>
>
> > raises AttributeError: 'BurstSubsetOption' object has no attribute 'inputs'

Sorry that it didn't work for you. Just checking, how did you run it? I just checked the ["testable on Colab now"](https://colab.research.google.com/github/isce-framework/dolphin/blob/main/docs/notebooks/walkthrough-basic.ipynb) where it runs a `pip install dolphin` `pip install opera-utils`, and I didn't get that error. the `dolphin.show_versions()` call should show [0.12.0 for opera_utils](https://pypi.org/project/opera-utils/).

Maybe the fix was poorly timed and you had already installed the software earlier?

>     2. I assume that the branch "main" is the one that we should test. Am I correct?

Yes `main` is good. `joss` has the paper, but I have been merging the changes from `main` into `joss`, so both are fine to test.

>     3. Regarding the [readme](https://github.com/isce-framework/dolphin/blob/main/README.md)
>
>
>     * Please add the full name for InSAR and PS/DS as especially PS/DS are not known outside of the community (similar to how it is mentioned in the paper manuscript)

Thanks, [added here](https://github.com/isce-framework/dolphin/commit/0da978a32b90a66b900efecdee9ef405f257b5bd)

>     4. Regarding the [license](https://github.com/isce-framework/dolphin/blob/main/LICENSE)
>
>
>     * For clarification I would recommend to indicate that the first section is the BSC-3-Clause. You could add the name before “Copyright (c) 2022 California …” similar to the section for the Apache License.

Added BSD clarification [here](https://github.com/isce-framework/dolphin/commit/230afc969dff474fafe042132e174ea2a94b7514), looks like the Apache section already starts with "Apache License"

>     5. Regarding the authors contributions
>        Scott J. Staniewicz lead author (511+4)
>        Sara Mirzaee (11)
>        Geoffrey M. Gunter (4)
>        Tablib Oliver-Cabrera (not committing author)
>        Heresh Fattahi (lead for dolphin project)
>        Not mentioned which is ok based on the low commit count (Ryan Burns (2), Simran Sangha (4), seyeonjeon (1))
>
>
>     * I would suggest to include a credit of author statement in the paper manuscript. Similar to the ones, journals like elsevier (https://www.elsevier.com/researcher/author/policies-and-guidelines/credit-author-statement) or mdpi have. But of course adjusted to the special needs of software development. As I do not know the opinion/options in place by JOSS it might be good if @mikemahoney218 could comment on that to.

I'd be interested in any JOSS suggestion about how to add that.

>     6. I would recommend to include the changelog in documentation

Added [here](https://github.com/isce-framework/dolphin/pull/395/commits/87ef5ed41e0519457fbff3a4fe9b9c3a249259aa)

>     7. Regarding the checklist point "Substantial Scholarly Effect"
>
>
>     * If applicable please add within the paper manuscript if the software were used to process data or if results are already published in other papers. Further you could indicate/cite projects were dolphin is currently used.

The main research work using Dolphin is still ongoing as part of the OPERA project. I'll double check if there are any citations that make sense to add for that which currently exist.

>     8. Broken links
>
>
>     * https://dolphin-insar.readthedocs.io/en/latest/getting-started/ penultimate line "notebooks"
>
>     * https://dolphin-insar.readthedocs.io/en/latest/notebooks/walkthrough-basic/ under Setup the link to "Getting Started"
>

Thank you! [Fixed](https://github.com/isce-framework/dolphin/pull/395/commits/52fc49c69e08572e6faeba1bbbf9e61a97b6087e)
