### 2024-08-14 4f78c48
DOCS: Add clarificaitons and suggested changes to documentation (#395)
* Clarify acronyms in headline of README

* clarify first section of license is BSD

* Add link to changelog on docs

* Fix broken tutorials and getting started links

* Fix changelog link using snippets
