### 2024-08-14 12c821b
Set `use_slc_amp` default to False (#396)
