### 2024-08-15 d44d62e
avoid recalc of redisual by passing through
