### 2024-08-15 a9122ca
add test for `invert_stack_l1` to match `invert_stack`
