### 2024-08-15 68cdf1f
Add `method` to request L1/L2 inversion in `run` and `invert_unw_network`
