### 2024-08-15 7b4e470
add method to workflow configuration
