### 2024-08-15 3b2ba0c
get basic single-reference changeover working for interferograms
