### 2024-08-16 29824e4
Fix `displacement.run` to avoid troposphere computation with empty list (#399)
as pointed out by @collinss-jpl , this used to not check for `None`. Since the default initializer is an empty list, this fails after loading the yaml and creating `[]`.
