### 2024-08-20 3720d9f
Fix `n_parallel_tiles` to be an integer, not float
