### 2024-08-20 310a1ad
Add L1-norm-based inversion to `timeseries` (#402)
* start adding `irls`

* Get IRLS loop working

* Add demo notebook about L1 inversion

* avoid recalc of redisual by passing through

* Add `irls` test

* add test for `invert_stack_l1` to match `invert_stack`

* Add `method` to request L1/L2 inversion in `run` and `invert_unw_network`

* add method to workflow configuration

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix notebook ruff errors

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.norepl
[…]
