+++ src/dolphin/timeseries.py
@@ -2,13 +2,14 @@ from __future__ import annotations
 
 import logging
 import shutil
+from enum import Enum
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 from typing import Callable, Optional, Protocol, Sequence, TypeVar
 
 import jax.numpy as jnp
 import numpy as np
-from jax import Array, jit, vmap
+from jax import Array, jit, lax, vmap
 from numpy.typing import ArrayLike
 from opera_utils import get_dates
 from scipy import ndimage
@@ -26,6 +27,13 @@ logger = logging.getLogger(__name__)
 __all__ = ["run"]
 
 
+class InversionMethod(str, Enum):
+    """Method to use for timeseries inversion."""
+
+    L1 = "L1"
+    L2 = "L2"
+
+
 class ReferencePointError(ValueError):
     pass
 
@@ -37,6 +45,7 @@ def run(
     condition_file: PathOrStr,
     condition: CallFunc,
     output_dir: PathOrStr,
+    method: InversionMethod = InversionMethod.L2,
     run_velocity: bool = False,
     velocity_file: Optional[PathOrStr] = None,
     correlation_threshold: float = 0.2,
@@ -62,6 +71,10 @@ def run(
         the options are [min, max]
     output_dir : Path
         Path to the output directory.
+    method : str, choices = "L1", "L2"
+        Inversion method to use when solving Ax = b.
+        Default is L2, which uses least squares to solve Ax = b (faster).
+        "L1" minimizes |Ax - b|_1 at each pixel.
     run_velocity : bool
         Whether to run velocity estimation on the inverted phase series
     velocity_file : Path, Optional
@@ -115,6 +128,7 @@ def run(
             reference=ref_point,
             output_dir=output_dir,
             num_threads=num_threads,
+            method=method,
         )
     else:
         logger.info(
@@ -644,6 +658,7 @@ def invert_unw_network(
     cor_threshold: float = 0.2,
     n_cor_looks: int = 1,
     ifg_date_pairs: Sequence[Sequence[DateOrDatetime]] | None = None,
+    method: InversionMethod = InversionMethod.L2,
     block_shape: tuple[int, int] = (512, 512),
     num_threads: int = 5,
     add_overviews: bool = True,
@@ -663,6 +678,10 @@ def invert_unw_network(
     ifg_date_pairs : Sequence[Sequence[DateOrDatetime]], optional
         List of date pairs to use for the inversion. If not provided, will be
         parsed from filenames in `unw_file_list`.
+    method : str, choices = "L1", "L2"
+        Inversion method to use when solving Ax = b.
+        Default is L2, which uses least squares to solve Ax = b (faster).
+        "L1" minimizes |Ax - b|_1 at each pixel.
     block_shape : tuple[int, int], optional
         The shape of the blocks to process in parallel
     cor_file_list : Sequence[PathOrStr], optional
@@ -725,7 +744,7 @@ def invert_unw_network(
     def read_and_solve(
         readers: Sequence[io.StackReader], rows: slice, cols: slice
     ) -> tuple[slice, slice, np.ndarray]:
-        if len(readers) == 2:
+        if len(readers) == 2 and method == "L2":
             unw_reader, cor_reader = readers
             stack = unw_reader[:, rows, cols]
             cor = cor_reader[:, rows, cols]
@@ -742,6 +761,10 @@ def invert_unw_network(
         # TODO: do i want to write residuals too? Do i need
         # to have multiple writers then?
[…]
