+++ tests/test_timeseries.py
@@ -112,7 +112,22 @@ class TestInvert:
         npt.assert_allclose(x, sar_phases[1:, -1, -1], atol=1e-5)
         assert np.array(residual).item() < 1e-5
 
-    def test_stack(self, data, A):
+    def test_l1_invert(self, data, A):
+        sar_dates, sar_phases, ifg_date_pairs, ifgs = data
+
+        # Get one pixel
+        b = ifgs[:, -1, -1].copy()
+        # Change one value by a lot, should ignore it
+        single_ifg_error = 100
+        b[0] += single_ifg_error
+
+        x, residuals = timeseries.irls(A, b)
+        assert x.shape[0] == len(sar_dates) - 1
+        npt.assert_allclose(x, sar_phases[1:, -1, -1], atol=1e-4)
+        npt.assert_allclose(residuals[0], single_ifg_error, atol=1e-4)
+        npt.assert_allclose(residuals[1:], 0, atol=1e-4)
+
+    def test_invert_stack(self, data, A):
         sar_dates, sar_phases, ifg_date_pairs, ifgs = data
 
         phi_stack, residuals = timeseries.invert_stack(A, ifgs)
@@ -120,6 +135,18 @@ class TestInvert:
         assert residuals.shape == sar_phases[0].shape
         npt.assert_allclose(phi_stack, sar_phases[1:], atol=1e-5)
         npt.assert_array_less(residuals, 1e-5)
+
+    def test_invert_stack_l1(self, data, A):
+        sar_dates, sar_phases, ifg_date_pairs, ifgs = data
+        ifgs = ifgs.copy()
+        single_ifg_error = 100
+        ifgs[0] += single_ifg_error
+
+        phi_stack, residuals = timeseries.invert_stack_l1(A, ifgs)
+        assert phi_stack.shape == sar_phases[1:].shape
+        assert residuals.shape == sar_phases[0].shape
+        npt.assert_allclose(phi_stack, sar_phases[1:], atol=1e-3)
+        npt.assert_allclose(residuals, single_ifg_error, atol=1e-3)
 
     def test_weighted_stack(self, data, A):
         sar_dates, sar_phases, ifg_date_pairs, ifgs = data
@@ -171,7 +198,8 @@ class TestInvert:
 
         return out
 
-    def test_invert_unw_network(self, data, unw_files, tmp_path):
+    @pytest.mark.parametrize("method", ["L1", "L2"])
+    def test_invert_unw_network(self, data, unw_files, tmp_path, method):
         output_dir = tmp_path / "output"
         output_dir.mkdir()
         ref_point = (0, 0)
@@ -179,6 +207,7 @@ class TestInvert:
             unw_file_list=unw_files,
             reference=ref_point,
             output_dir=output_dir,
+            method=method,
             # ifg_date_pairs: Sequence[Sequence[DateOrDatetime]] | None = None,
             # block_shape: tuple[int, int] = (512, 512),
             # cor_file_list: Sequence[PathOrStr] | None = None,
