### 2024-08-20 6aedb71
Skip `combine_mask_files` if the output already exists
