### 2024-08-21 isce-framework/dolphin#405 @McWhity
Issues if running docs/notebooks/walkthrough-basic.ipynb
@openjournals/joss-reviews#6997

Regarding [walkthrough-basic.ipynb](https://github.com/isce-framework/dolphin/blob/main/docs/notebooks/walkthrough-basic.ipynb)

Within section "Running the workflow" if running command `!dolphin run dolphin_config.yaml` I get following error

```
[INFO|_unwrap|L140] 2024-08-21T15:03:50+0200: 23 left to unwrap

snaphu v2.0.7
22 parameters input from file /media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/dolphin_docker/dolphin/docs/notebooks/input_slcs/work-walkthrough/unwrapped/scratch/scratch-20210606_20210712.int/snaphu.config.t09g99h6.txt (22 lines total)
Reading wrapped phase from file /media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/dolphin_docker/dolphin/docs/notebooks/input_slcs/work-walkthrough/unwrapped/scratch/scratch-20210606_20210712.int/snaphu.igram.9vp76nu1.c8
No weight file specified.  Assuming uniform weights
Reading correlation data from file /media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/dolphin_docker/dolphin/docs/notebooks/input_slcs/work-walkthrough/unwrapped/scratch/scratch-20210606_20210712.int/snaphu.corr.3t9rsnnr.f4
Calculating smooth-solution cost parameters
Initializing flows with MCF algorithm
Running nonlinear network flow optimizer
Maximum flow on network: 2
Flow increment: 1  (Total improvements: 0)
Found 1 valid set(s) of connected nodes
Growing connected component mask
Writing connected components to file /media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/dolphin_docker/dolphin/docs/notebooks/input_slcs/work-walkthrough/unwrapped/scratch/scratch-20210606_20210712.int/snaphu.conncomp.kr7b5zph.u4 as 4-byte unsigned ints
Maximum flow on network: 1
Total solution cost: 20879598
Integrating phase
Writing output to file /media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/dolphin_docker/dolphin/docs/notebooks/input_slcs/work-walkthrough/unwrapped/scratch/scratch-20210606_20210712.int/snaphu.unw.s2u_yz74.f4
Program snaphu done
Elapsed processor time:   0:00:06.42
Elapsed wall clock time:  0:00:06
[INFO|_unwrap|L480] 2024-08-21T15:03:57+0200: Setting nodata values of /media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/dolphin_docker/dolphin/docs/notebooks/input_slcs/work-walkthrough/unwrapped/20210606_20210712.unw.tif file
[INFO|_unwrap|L484] 2024-08-21T15:03:57+0200: Setting nodata values of /media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/dolphin_docker/dolphin/docs/notebooks/input_slcs/work-walkthrough/unwrapped/20210606_20210712.unw.conncomp.tif file
Traceback (most recent call last):
  File "/home/tweiss/anaconda3/envs/dolphin-env2/bin/dolphin", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/cli.py", line 32, in main
    run_func(**arg_dict)
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/workflows/_cli_run.py", line 30, in run
    displacement.run(cfg, debug=debug)
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/_log.py", line 90, in wrapper
    result = f(*args, **kwargs)
             ^^^^^^^^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/workflows/displacement.py", line 233, in run
    unwrapped_paths, conncomp_paths = unwrapping.run(
                                      ^^^^^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/_log.py", line 90, in wrapper
    result = f(*args, **kwargs)
             ^^^^^^^^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/workflows/unwrapping.py", line 82, in run
    unwrapped_paths, conncomp_paths = unwrap.run(
                                      ^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/unwrap/_unwrap.py", line 158, in run
    exc.submit(
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/utils.py", line 680, in submit
    result = fn(*args, **kwargs)
             ^^^^^^^^^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/site-packages/dolphin/unwrap/_unwrap.py", line 491, in unwrap
    shutil.rmtree(scratchdir)
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/shutil.py", line 759, in rmtree
    _rmtree_safe_fd(stack, onexc)
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/shutil.py", line 703, in _rmtree_safe_fd
    onexc(func, path, err)
  File "/home/tweiss/anaconda3/envs/dolphin-env2/lib/python3.12/shutil.py", line 662, in _rmtree_safe_fd
    os.rmdir(name, dir_fd=dirfd)
OSError: [Errno 39] Directory not empty: PosixPath('/media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/dolphin_docker/dolphin/docs/notebooks/input_slcs/work-walkthrough/unwrapped/scratch/scratch-20210606_20210712.int')
```

Somehow the directory seems to be not empty. But when I checkout the folder "scratch-20210606_20210712.int" afterwards it is empty
