### 2024-08-21 isce-framework/dolphin#406 @McWhity
Problem installing with docker
@openjournals/joss-reviews#6997

I tried to install dolphin via dockerfile [https://github.com/isce-framework/dolphin/blob/main/docker/Dockerfile](https://github.com/isce-framework/dolphin/blob/main/docker/Dockerfile)

First issue I encounter was

```
------
 > [stage-1 16/21] COPY --chown=mamba:mamba docker/specfile.txt /tmp/specfile.txt:
------
Dockerfile:64
--------------------
  62 |     # Install conda packages
  63 |     # https://github.com/mamba-org/micromamba-docker#quick-start
  64 | >>> COPY --chown=$MAMBA_USER:$MAMBA_USER docker/specfile.txt /tmp/specfile.txt
  65 |     RUN micromamba install --yes --channel conda-forge -n base -f /tmp/specfile.txt && \
  66 |         micromamba clean --all --yes
--------------------
ERROR: failed to solve: failed to compute cache key: failed to calculate checksum of ref 61b26c04-97a4-4cc6-b5ed-2b0c9ae5aa6f::z8vhf3w2wtwbn1cubgsq1rro6: "/docker/specfile.txt": not found
```
it seems to be an issue with the current directory

I change line 64 from
`COPY --chown=$MAMBA_USER:$MAMBA_USER docker/specfile.txt /tmp/specfile.txt`
to
`COPY --chown=$MAMBA_USER:$MAMBA_USER specfile.txt /tmp/specfile.txt`
which resolved the issue
