### 2024-08-21 isce-framework/dolphin#406 @McWhity
Another error that I get is
```
------
 > [stage-1 20/21] RUN python -m pip install --no-deps .:
1.114 ERROR: Directory '.' is not installable. Neither 'setup.py' nor 'pyproject.toml' found.
------
Dockerfile:81
--------------------
  79 |
  80 |     # --no-deps because they are installed with conda
  81 | >>> RUN python -m pip install --no-deps .
  82 |
  83 |     # Run the entrypoint from the /work directory
--------------------
ERROR: failed to solve: process "/usr/local/bin/_dockerfile_shell.sh python -m pip install --no-deps ." did not complete successfully: exit code: 1
```
Again it seems to be an error regarding the directory. I tried to hardcode the path but that did not resolve the error.
