### 2024-08-21 isce-framework/dolphin#405 @McWhity
[https://github.com/isce-framework/dolphin/blob/main/docs/notebooks/walkthrough-basic.ipynb](https://github.com/isce-framework/dolphin/blob/main/docs/notebooks/walkthrough-basic.ipynb)
There might be a typo in section "Bounds subsetting". Highlighted in bold below

> If you are interested in a smaller subset of the **are**, you can specify the --output-bounds to force the raster **boudaries**.
