### 2024-08-21 isce-framework/dolphin#406 @scottstanie
Hi @McWhity , How were you trying to build it? I've included a script to help with a few common options we do, though the base command is
```
docker build --network=host --tag $image_name --file docker/Dockerfile
```
Had you changed directories into the `docker` folder to build it?
