### 2024-08-21 isce-framework/dolphin#405 @scottstanie
Interesting, that error is indeed weird.. What platform (linxus/mac/windows) are you running this on? Were you trying to use docker here? (I only ask because I saw the folder with `docker` in the name)

Typos fixed here https://github.com/isce-framework/dolphin/pull/407
