### 2024-08-21 isce-framework/dolphin#405 @McWhity
I am running it on Linux. No docker usage here. At first I tried to install it via docker but as I run into some errors I switched and just created a new environment. But I used "conda" instead of "mamba" to install dolphin. Tomorrow I will try it with mamba as well.
