### 2024-08-21 4a5c275
Update OPERA acknowledgment statement
