### 2024-08-22 isce-framework/dolphin#406 @McWhity
> Hi @McWhity , How were you trying to build it? I've included the script ./docker/build-docker-image.sh to help with a few common options we do, though the base command is

Thank you for the hint. I tried the following command
`docker build docker/.`
which produces the error mentioned above. The error does not change if I am in the main directory or in the subfolder `docker.` If I use your `build-docker-image.sh` than everything works fine.

As you already prepared the relevant files for the docker creation I would recommend to add a short docker instruction guide to the Installation section.
