### 2024-08-22 isce-framework/dolphin#405 @McWhity
Right now I have to installations.
1. using your installation guide and creating a virtual environment
2. Docker container using `build-docker-image.sh` to build it

Using installation 1: I get the above mentioned error
Using installation 2: I encountered a different problem
-
```
