### 2024-08-22 isce-framework/dolphin#405 @McWhity
Right now I have to installations.

1. using your installation guide and creating a virtual environment
2. Docker container using `build-docker-image.sh` to build it

Using installation 1: I get the above mentioned error
Using installation 2: I encountered other problems

```
Cell In[4], line 25
     23 url_file = slc_dir / \"urls.txt\"
     24 with open(url_file, \"w\") as f:
---> 25     f.write(\"\
\".join(best_option.inputs))
     28 chdir(slc_dir)
     29 # Download 4in parallel

AttributeError: 'BurstSubsetOption' object has no attribute 'inputs'"

```
The problem is due to the version of package `opera-utils`. Within the docker the version 0.8.0 is installed. When I checked Installation 1 I found version 0.12.0. Thus, you might need to update the package version in [https://github.com/isce-framework/dolphin/blob/main/docker/specfile.txt](https://github.com/isce-framework/dolphin/blob/main/docker/specfile.txt)

Another problem running the notebook was that I used as as docker configuration `python3` in which `wget` is not installed. Thus I added the `wget` installation to the `Dockerfile`. You might consider adding
```
# Install wget
RUN apt-get update && apt-get install -y wget
```
to [Dockerfile](https://github.com/isce-framework/dolphin/blob/main/docker/Dockerfile)

With the mentioned adjustments I can run the notebook. However I am not able to visualize the output. There is no `timeseries` folder in the folder `work-walkthrough`. The folder structure looks like this (only 2 sub-directories are shown)
```
work-walkthrough/
├── dolphin.log
├── interferograms
│   ├── 20210606_20210618.int.cor.tif
│   ├── 20210606_20210618.int.tif
│   ├── 20210606_20210630.int.cor.tif
│   ├── 20210606_20210630.int.tif
│   ├── 20210606_20210712.int.cor.tif
│   ├── 20210606_20210712.int.tif
│   ├── 20210606_20210724.int.cor.tif
│   ├── 20210606_20210724.int.tif
│   ├── 20210606_20210805.int.cor.tif
│   ├── 20210606_20210805.int.tif
│   ├── 20210606_20210817.int.cor.tif
│   ├── 20210606_20210817.int.tif
│   ├── 20210606_20210829.int.cor.tif
│   ├── 20210606_20210829.int.tif
│   ├── 20210606_20210910.int.cor.tif
│   ├── 20210606_20210910.int.tif
│   ├── 20210606_20210922.int.cor.tif
│   ├── 20210606_20210922.int.tif
│   ├── 20210606_20211004.int.cor.tif
│   ├── 20210606_20211004.int.tif
│   ├── 20210606_20211016.int.cor.tif
│   ├── 20210606_20211016.int.tif
│   ├── 20210606_20211028.int.cor.tif
│   ├── 20210606_20211028.int.tif
│   ├── 20210606_20211109.int.cor.tif
│   ├── 20210606_20211109.int.tif
│   ├── 20210606_20211121.int.cor.tif
│   ├── 20210606_20211121.int.tif
│   ├── 20210606_20211203.int.cor.tif
│   ├── 20210606_20211203.int.tif
│   ├── 20210606_20211215.int.cor.tif
│   ├── 20210606_20211215.int.tif
│   ├── 20210606_20220219.int.cor.tif
│   ├── 20210606_20220219.int.tif
│   ├── 20210606_20220303.int.cor.tif
│   ├── 20210606_20220303.int.tif
│   ├── 20210606_20220315.int.cor.tif
│   ├── 20210606_20220315.int.tif
│   ├── 20210606_20220327.int.cor.tif
│   ├── 20210606_20220327.int.tif
│   ├── 20210606_20220408.int.cor.tif
│   ├── 20210606_20220408.int.tif
│   ├── 20210606_20220420.int.cor.tif
│   ├── 20210606_20220420.int.tif
│   ├── 20210606_20220502.int.cor.tif
│   ├── 20210606_20220502.int.tif
│   ├── 20210606_20220514.int.cor.tif
│   ├── 20210606_20220514.int.tif
│   ├── 20210606_20220526.int.cor.tif
│   ├── 20210606_20220526.int.tif
│   ├── amp_dispersion_looked.tif
│   ├── ps_mask_looked.tif
│   ├── shp_counts.tif
│   └── temporal_coherence.tif
├── t078_165573_iw2
│   ├── bounds_mask.tif
│   ├── combined_mask.tif
│   ├── interferograms
│   ├── linked_phase
│   ├── nodata_mask.tif
│   ├── PS
│   ├── slc_stack.vrt
│   ├── timeseries
│   └── unwrapped
├── t078_165574_iw2
│   ├── bounds_mask.tif
│   ├── combined_mask.tif
│   ├── interferograms
│   ├── linked_phase
│   ├── nodata_mask.tif
│   ├── PS
│   ├── slc_stack.vrt
│   ├── timeseries
│   └── unwrapped
└── unwrapped
    ├── 20210606_20210618.unw.conncomp.tif
    ├── 20210606_20210618.unw.tif
    └── scratch

```
Here is the entire folder structure
```
work-walkthrough/
├── dolphin.log
├── interferograms
│   ├── 20210606_20210618.int.cor.tif
│   ├── 20210606_20210618.int.tif
│   ├── 20210606_20210630.int.cor.tif
│   ├── 20210606_20210630.int.tif
│   ├── 20210606_20210712.int.cor.tif
│   ├── 20210606_20210712.int.tif
│   ├── 20210606_20210724.int.cor.tif
│   ├── 20210606_20210724.int.tif
│   ├── 20210606_20210805.int.cor.tif
│   ├── 20210606_20210805.int.tif
│   ├── 20210606_20210817.int.cor.tif
│   ├── 20210606_20210817.int.tif
│   ├── 20210606_20210829.int.cor.tif
│   ├── 20210606_20210829.int.tif
│   ├── 20210606_20210910.int.cor.tif
│   ├── 20210606_20210910.int.tif
│   ├── 20210606_20210922.int.cor.tif
│   ├── 20210606_20210922.int.tif
│   ├── 20210606_20211004.int.cor.tif
│   ├── 20210606_20211004.int.tif
│   ├── 20210606_20211016.int.cor.tif
│   ├── 20210606_20211016.int.tif
│   ├── 20210606_20211028.int.cor.tif
│   ├── 20210606_20211028.int.tif
│   ├── 20210606_20211109.int.cor.tif
│   ├── 20210606_20211109.int.tif
│   ├── 20210606_20211121.int.cor.tif
│   ├── 20210606_20211121.int.tif
│   ├── 20210606_20211203.int.cor.tif
│   ├── 20210606_20211203.int.tif
│   ├── 20210606_20211215.int.cor.tif
│   ├── 20210606_20211215.int.tif
│   ├── 20210606_20220219.int.cor.tif
│   ├── 20210606_20220219.int.tif
│   ├── 20210606_20220303.int.cor.tif
│   ├── 20210606_20220303.int.tif
│   ├── 20210606_20220315.int.cor.tif
│   ├── 20210606_20220315.int.tif
│   ├── 20210606_20220327.int.cor.tif
│   ├── 20210606_20220327.int.tif
│   ├── 20210606_20220408.int.cor.tif
│   ├── 20210606_20220408.int.tif
│   ├── 20210606_20220420.int.cor.tif
│   ├── 20210606_20220420.int.tif
│   ├── 20210606_20220502.int.cor.tif
│   ├── 20210606_20220502.int.tif
│   ├── 20210606_20220514.int.cor.tif
│   ├── 20210606_20220514.int.tif
│   ├── 20210606_20220526.int.cor.tif
│   ├── 20210606_20220526.int.tif
│   ├── amp_dispersion_looked.tif
│   ├── ps_mask_looked.tif
│   ├── shp_counts.tif
│   └── temporal_coherence.tif
├── t078_165573_iw2
│   ├── bounds_mask.tif
│   ├── combined_mask.tif
│   ├── interferograms
│   │   ├── 20210606_20210618.int.vrt
│   │   ├── 20210606_20210630.int.vrt
│   │   ├── 20210606_20210712.int.vrt
│   │   ├── 20210606_20210724.int.vrt
│   │   ├── 20210606_20210805.int.vrt
│   │   ├── 20210606_20210817.int.vrt
│   │   ├── 20210606_20210829.int.vrt
│   │   ├── 20210606_20210910.int.vrt
│   │   ├── 20210606_20210922.int.vrt
│   │   ├── 20210606_20211004.int.vrt
│   │   ├── 20210606_20211016.int.vrt
│   │   ├── 20210606_20211028.int.vrt
│   │   ├── 20210606_20211109.int.vrt
│   │   ├── 20210606_20211121.int.vrt
│   │   ├── 20210606_20211203.int.vrt
│   │   ├── 20210606_20211215.int.vrt
│   │   ├── 20210606_20220219.int.vrt
│   │   ├── 20210606_20220303.int.vrt
│   │   ├── 20210606_20220315.int.vrt
│   │   ├── 20210606_20220327.int.vrt
│   │   ├── 20210606_20220408.int.vrt
│   │   ├── 20210606_20220420.int.vrt
│   │   ├── 20210606_20220502.int.vrt
│   │   ├── 20210606_20220514.int.vrt
│   │   └── 20210606_20220526.int.vrt
│   ├── linked_phase
│   │   ├── 20210606_20211121
│   │   │   ├── avg_coh_20210606_20211121.tif
│   │   │   ├── eigenvalues_20210606_20211121.tif
│   │   │   ├── estimator_20210606_20211121.tif
│   │   │   ├── shp_counts_20210606_20211121.tif
│   │   │   └── temporal_coherence_20210606_20211121.tif
│   │   ├── 20210606_20211121.vrt
│   │   ├── 20210606.slc.tif
│   │   ├── 20210618.slc.tif
│   │   ├── 20210630.slc.tif
│   │   ├── 20210712.slc.tif
│   │   ├── 20210724.slc.tif
│   │   ├── 20210805.slc.tif
│   │   ├── 20210817.slc.tif
│   │   ├── 20210829.slc.tif
│   │   ├── 20210910.slc.tif
│   │   ├── 20210922.slc.tif
│   │   ├── 20211004.slc.tif
│   │   ├── 20211016.slc.tif
│   │   ├── 20211028.slc.tif
│   │   ├── 20211109.slc.tif
│   │   ├── 20211121.slc.tif
│   │   ├── 20211203_20220526
│   │   │   ├── avg_coh_20211203_20220526.tif
│   │   │   ├── eigenvalues_20211203_20220526.tif
│   │   │   ├── estimator_20211203_20220526.tif
│   │   │   ├── shp_counts_20211203_20220526.tif
│   │   │   └── temporal_coherence_20211203_20220526.tif
│   │   ├── 20211203_20220526.vrt
│   │   ├── 20211203.slc.tif
│   │   ├── 20211215.slc.tif
│   │   ├── 20220219.slc.tif
│   │   ├── 20220303.slc.tif
│   │   ├── 20220315.slc.tif
│   │   ├── 20220327.slc.tif
│   │   ├── 20220408.slc.tif
│   │   ├── 20220420.slc.tif
│   │   ├── 20220502.slc.tif
│   │   ├── 20220514.slc.tif
│   │   ├── 20220526.slc.tif
│   │   ├── compressed_20210606_20210606_20211121.tif
│   │   ├── compressed_20210606_20211203_20220526.tif
│   │   ├── shp_counts_average_20210606_20220526.tif
│   │   └── temporal_coherence_average_20210606_20220526.tif
│   ├── nodata_mask.tif
│   ├── PS
│   │   ├── amp_dispersion_looked.tif
│   │   ├── amp_dispersion.tif
│   │   ├── amp_mean.tif
│   │   ├── ps_pixels_looked.tif
│   │   └── ps_pixels.tif
│   ├── slc_stack.vrt
│   ├── timeseries
│   └── unwrapped
├── t078_165574_iw2
│   ├── bounds_mask.tif
│   ├── combined_mask.tif
│   ├── interferograms
│   │   ├── 20210606_20210618.int.vrt
│   │   ├── 20210606_20210630.int.vrt
│   │   ├── 20210606_20210712.int.vrt
│   │   ├── 20210606_20210724.int.vrt
│   │   ├── 20210606_20210805.int.vrt
│   │   ├── 20210606_20210817.int.vrt
│   │   ├── 20210606_20210829.int.vrt
│   │   ├── 20210606_20210910.int.vrt
│   │   ├── 20210606_20210922.int.vrt
│   │   ├── 20210606_20211004.int.vrt
│   │   ├── 20210606_20211016.int.vrt
│   │   ├── 20210606_20211028.int.vrt
│   │   ├── 20210606_20211109.int.vrt
│   │   ├── 20210606_20211121.int.vrt
│   │   ├── 20210606_20211203.int.vrt
│   │   ├── 20210606_20211215.int.vrt
│   │   ├── 20210606_20220219.int.vrt
│   │   ├── 20210606_20220303.int.vrt
│   │   ├── 20210606_20220315.int.vrt
│   │   ├── 20210606_20220327.int.vrt
│   │   ├── 20210606_20220408.int.vrt
│   │   ├── 20210606_20220420.int.vrt
│   │   ├── 20210606_20220502.int.vrt
│   │   ├── 20210606_20220514.int.vrt
│   │   └── 20210606_20220526.int.vrt
│   ├── linked_phase
│   │   ├── 20210606_20211121
│   │   │   ├── avg_coh_20210606_20211121.tif
│   │   │   ├── eigenvalues_20210606_20211121.tif
│   │   │   ├── estimator_20210606_20211121.tif
│   │   │   ├── shp_counts_20210606_20211121.tif
│   │   │   └── temporal_coherence_20210606_20211121.tif
│   │   ├── 20210606_20211121.vrt
│   │   ├── 20210606.slc.tif
│   │   ├── 20210618.slc.tif
│   │   ├── 20210630.slc.tif
│   │   ├── 20210712.slc.tif
│   │   ├── 20210724.slc.tif
│   │   ├── 20210805.slc.tif
│   │   ├── 20210817.slc.tif
│   │   ├── 20210829.slc.tif
│   │   ├── 20210910.slc.tif
│   │   ├── 20210922.slc.tif
│   │   ├── 20211004.slc.tif
│   │   ├── 20211016.slc.tif
│   │   ├── 20211028.slc.tif
│   │   ├── 20211109.slc.tif
│   │   ├── 20211121.slc.tif
│   │   ├── 20211203_20220526
│   │   │   ├── avg_coh_20211203_20220526.tif
│   │   │   ├── eigenvalues_20211203_20220526.tif
│   │   │   ├── estimator_20211203_20220526.tif
│   │   │   ├── shp_counts_20211203_20220526.tif
│   │   │   └── temporal_coherence_20211203_20220526.tif
│   │   ├── 20211203_20220526.vrt
│   │   ├── 20211203.slc.tif
│   │   ├── 20211215.slc.tif
│   │   ├── 20220219.slc.tif
│   │   ├── 20220303.slc.tif
│   │   ├── 20220315.slc.tif
│   │   ├── 20220327.slc.tif
│   │   ├── 20220408.slc.tif
│   │   ├── 20220420.slc.tif
│   │   ├── 20220502.slc.tif
│   │   ├── 20220514.slc.tif
│   │   ├── 20220526.slc.tif
│   │   ├── compressed_20210606_20210606_20211121.tif
│   │   ├── compressed_20210606_20211203_20220526.tif
│   │   ├── shp_counts_average_20210606_20220526.tif
│   │   └── temporal_coherence_average_20210606_20220526.tif
│   ├── nodata_mask.tif
│   ├── PS
│   │   ├── amp_dispersion_looked.tif
│   │   ├── amp_dispersion.tif
│   │   ├── amp_mean.tif
│   │   ├── ps_pixels_looked.tif
│   │   └── ps_pixels.tif
│   ├── slc_stack.vrt
│   ├── timeseries
│   └── unwrapped
└── unwrapped
    ├── 20210606_20210618.unw.conncomp.tif
    ├── 20210606_20210618.unw.tif
    └── scratch
        └── scratch-20210606_20210618.int

```
