### 2024-08-26 3a5d482
Use element-wise multiplies to reweight L1
