### 2024-08-27 df61a72
add options to pass through for block size/parallelism
