### 2024-08-27 ee5325b
Add instructions for building Docker image
closes #406
