### 2024-08-27 6d1e5a9
add `wget`, `cxx-compiler` to docker
