### 2024-08-27 isce-framework/dolphin#405 @scottstanie
Thanks for the details @McWhity ,
- I added `wget` to docker
- The `.inputs` error was from `docker/specfile.txt` being outdated. the docker build uses the explicitly saved dependencies, so I updated that in #409

The output doesn't look like it was made, since it errored during the step making `unwrapped/`.

I still don't understand what that error means, as I've never seen that and haven't run into it running it on my Mac, linux, or on Colab..  is there a chance you started running it outside of docker, then got the error during a docker run? If so, it might be a folder permissions thing.
Do you get that same error running it again?
