### 2024-08-27 9ffd00d
fix install
