### 2024-08-28 56dcfbf
DOC: Add instructions for building Docker image (#409)
* Add instructions for building Docker image

closes #406

* update specfile

* add `wget`, `cxx-compiler` to docker

* fix install
