### 2024-08-28 4c84858
update FRInGE bibtex
