+++ paper/references.bib
@@ -99,16 +99,21 @@
   keywords = {decorrelation,groundwater,InSAR deformation map,persistent scatterers}
 }
 
-@article{Fattahi2019FRInGEFullResolutionInSAR,
-  title    = {{{FRInGE}}; {{Full-Resolution InSAR}} Timeseries Using {{Generalized Eigenvectors}}},
-  author   = {Fattahi, H. and Agram, P. S. and Tymofyeyeva, E. and Bekaert, D. P.},
-  year     = {2019},
-  month    = dec,
-  volume   = {2019},
-  pages    = {G11B-0514},
-  urldate  = {2024-05-16},
-  keywords = {1209 Tectonic deformation,1211 Non-tectonic deformation,1240 Satellite geodesy: results,1241 Satellite geodesy: technical issues,GEODESY AND GRAVITY}
+@inproceedings{2019AGUFM.G11B0514F,
+  author    = {{Fattahi}, H. and {Agram}, P.S. and {Tymofyeyeva}, E. and {Bekaert}, D.P.},
+  title     = {{FRInGE; Full-Resolution InSAR timeseries using Generalized Eigenvectors}},
+  keywords  = {1209 Tectonic deformation, GEODESY AND GRAVITY, 1211 Non-tectonic deformation, GEODESY AND GRAVITY, 1240 Satellite geodesy: results, GEODESY AND GRAVITY, 1241 Satellite geodesy: technical issues, GEODESY AND GRAVITY},
+  booktitle = {AGU Fall Meeting Abstracts},
+  year      = 2019,
+  volume    = {2019},
+  month     = dec,
+  eid       = {G11B-0514},
+  pages     = {G11B-0514},
+  adsurl    = {https://ui.adsabs.harvard.edu/abs/2019AGUFM.G11B0514F},
+  adsnote   = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
 
 @article{Ferretti2011NewAlgorithmProcessing,
   title     = {A New Algorithm for Processing Interferometric Data-Stacks: {{SqueeSAR}}},
