+++ paper/references.bib
@@ -193,16 +193,17 @@
 }
 
 @inproceedings{Rosen2018InSARScientificComputing,
-  title     = {The {{InSAR Scientific Computing Environment}} 3.0: {{A Flexible Framework}} for {{NISAR Operational}} and {{User-Led Science Processing}}},
   author    = {Rosen, Paul A. and Gurrola, Eric M. and Agram, Piyush and Cohen, Joshua and Lavalle, Marco and Riel, Bryan V. and Fattahi, Heresh and Aivazis, Michael A.G. and Simons, Mark and Buckley, Sean M.},
+  booktitle = {IGARSS 2018 - 2018 IEEE International Geoscience and Remote Sensing Symposium},
+  title     = {The {{InSAR Scientific Computing Environment}} 3.0: {{A Flexible Framework}} for {{NISAR Operational}} and {{User-Led Science Processing}}},
   year      = {2018},
-  month     = jul,
-  pages     = {4897--4900},
-  publisher = {IEEE},
-  address   = {Valencia},
-  doi       = {10.1109/IGARSS.2018.8517504},
-  langid    = {english}
+  volume    = {},
+  number    = {},
+  pages     = {4897-4900},
+  keywords  = {Python;Radar imaging;Production;Graphics processing units;Orbits;NASA;InSAR processing;geodetic imaging;computational frameworks;Earth science informatics},
+  doi       = {10.1109/IGARSS.2018.8517504}
 }
+
 
 @article{Parizzi2011AdaptiveInSARStack,
   title      = {Adaptive {{InSAR Stack Multilooking Exploiting Amplitude Statistics}}: {{A Comparison Between Different Techniques}} and {{Practical Results}}},
