### 2024-08-28 #6997 @scottstanie
> Also, on the paper:
>
>     1. Line 21: period missing at the end of the sentence
>
>     2. Fattahi et al. ref: missing the mention "AGU Fall Meeting Abstracts", maybe? (suggested citation [here](https://ui.adsabs.harvard.edu/abs/2019AGUFM.G11B0514F/exportcitation))
>
>     3. Rosen et al. ref: missing "IGARSS 2018 - 2018 IEEE International Geoscience and Remote Sensing Symposium", maybe?

Sorry that these slipped, I've push up the fixes to the JOSS branch.
