### 2024-08-28 5dd88a5
Include layover shadow mask when stitching in `prepare_geometry` (#410)
This should still probably be moved elsewhere from the too-specific assumptions about OPERA data format,but that can be a separate PR
