### 2024-08-28 #6997 @margauxmouchene
Hello!

@scottstanie with the fix you made, everything runs smoothly for me (I reinstalled everything).

In the 'walkthrough-basics' notebook, in the 'Visualization the displacement' (typo maybe?) section, in the last executable code block:
- there is a missing `import numpy as np` somewhere (`np.unravel_index` is called)
- `io.load_gdal` should be `dolphin.io.load_gdal`
- `pixels` is not defined

But everything else works, it's easy to follow and it's a great starting point to learn how to use **dolphin**.

In the other notebook ('simulate-demo'):
- `samps3d = simulate.create_noisy_deformation(C, defo_stack)`
returns the error:
`AttributeError: module 'dolphin.phase_link.simulate' has no attribute 'create_noisy_deformation' `

@mikemahoney218 I am almost done with the review, I just want to test the GPU speedup, which is one of the functionality claims of the paper. I should be able to wrap this all up before the end of the week.
