### 2024-08-28 #6997 @scottstanie
Thanks @margauxmouchene !

> In the 'walkthrough-basics' notebook, in the 'Visualization the displacement' (typo maybe?) section, in the last executable code block:
>
>     * there is a missing `import numpy as np` somewhere (`np.unravel_index` is called)
>
>     * `io.load_gdal` should be `dolphin.io.load_gdal`
>
>     * `pixels` is not defined

Thanks for that, I had someone else point that out just after you started your review, [fixed in this commit](https://github.com/isce-framework/dolphin/commit/b91bbb8381caab46bd2045c7778688c8dd2e98d0)

> But everything else works, it's easy to follow and it's a great starting point to learn how to use **dolphin**.
>
> In the other notebook ('simulate-demo'):
>
>     * `samps3d = simulate.create_noisy_deformation(C, defo_stack)`
>       returns the error:
>       `AttributeError: module 'dolphin.phase_link.simulate' has no attribute 'create_noisy_deformation' `
>

Ah thanks for catching that! [I fixed my function renaming here.](https://github.com/isce-framework/dolphin/pull/412)
