### 2024-08-28 9ccb07c
Fix `simulate-demo` function usage, add link to notebook (#412)
* Fix renamed function in simulate-demo

* add link in docs to `simulate-demo`
