+++ docs/tutorials.md
@@ -1,6 +1,7 @@
 # Tutorials
 
 1. [Basic walkthrough notebook](notebooks/walkthrough-basic.ipynb)
+1. [Compare phase linking algorithms on simulated data](notebooks/simulate-demo.ipynb)
 
 <!-- This part of the project documentation will contain step-by-step instructions for learning about generating displacement using Dolphin. -->
 
