### 2024-08-29 7508e31
Use the same mask for unwrapping and conncomp regrowing
Updates the unwrapping code to make sure we pass the same mask to the
connected component regrowing step (when applicable) that we passed to
the original unwrapper. Previously in the connected component regrowing
step, we only masked out pixels with coherence equal to zero.
