### 2024-08-30 2c228d1
Add `block_shape` to timeseries parameters, use less memory for L1 inversion (#414)
* Use element-wise multiplies to reweight L1

* add options to pass through for block size/parallelism

* [skip ci] fix docstrings
