### 2024-08-30 f4e3782
Fix outdated docs and missing log statements
