### 2024-08-30 a5b5e3c
Fix outdated docs and missing log statements (#415)
