### 2024-08-30 820874d
Add `_convert_and_reference` to timeseries instead of just copying
