### 2024-08-30 831b451
fix output units for full workflow, fix single-reference check
