### 2024-09-04 isce-framework/dolphin#405 @McWhity
@scottstanie: The installation process with mamba as well as with docker works now.
However, using the workflow from the `walkthrough-basic.ipynb` the output after Section `Running the workflow` is still as mentioned above. Unfortunately, the workflow does not produce an folder `work-walkthrough/timeseries`. The folder `work-walkthough/t078_165573_iw2/timeseries` is created but there are no files stored in that folder. As you can see in the comment above the Interferograms and Linked phase files are there.
