### 2024-09-04 isce-framework/dolphin#405 @scottstanie
Hmm I'm sorry this still isn't working for you, through I'm struggling to reproduce this error.

Would you mind putting the exact steps you're running to hit this error repeatedly?  That would help me debug if it's something with the notebook, with docker, or with your setup that I've under-specified. Are you using docker, but somehow executing the notebook? Running the notebook steps manually within docker?

Something to try- stripping everything around the notebook to see if `dolphin` is failing for you. You could try (assuming that you still have the directory of HDF5 input files) just entering that directory, then running:

```bash
cd input_slcs/  # wherever the  "OPERA_L2_CSLC-S1..." files are
rm -rf work-walkthrough

dolphin config --slc-files *.h5  --subdataset "/data/VV"     --threads-per-worker 8  --strides 6 3 --n-parallel-bursts 2     --output-bounds -102.8 31.3 -102.6 31.5 --work-directory work-walkthrough
dolphin run dolphin_config.yaml
```
This runs in 4-5 mins on my macbook.

Alternatively, if you really want to use docker, pull [the github-built image](https://github.com/isce-framework/dolphin/pkgs/container/dolphin),

```bash
docker pull ghcr.io/isce-framework/dolphin:develop
docker run --user $(id -u):$(id -g) -v $PWD:/work --rm -it ghcr.io/isce-framework/dolphin:develop bash
# ...as above with config, run
```

If that still fails, then there's definitely something wrong with this setup.
