### 2024-09-04 isce-framework/dolphin#405 @McWhity
I get the same problem if I run it with docker or in a virtualenv configuration with mamba installation process. But I think I found the problem. As suggested I did run the application in the terminal and there I got an error that is unfortunately not shown if I run the command in the jupyter notebook.

```
Traceback (most recent call last):
  File "/home/tweiss/anaconda3/envs/dolphin-env/bin/dolphin", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/cli.py", line 32, in main
    run_func(**arg_dict)
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/workflows/_cli_run.py", line 30, in run
    displacement.run(cfg, debug=debug)
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/_log.py", line 90, in wrapper
    result = f(*args, **kwargs)
             ^^^^^^^^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/workflows/displacement.py", line 233, in run
    unwrapped_paths, conncomp_paths = unwrapping.run(
                                      ^^^^^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/_log.py", line 90, in wrapper
    result = f(*args, **kwargs)
             ^^^^^^^^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/workflows/unwrapping.py", line 82, in run
    unwrapped_paths, conncomp_paths = unwrap.run(
                                      ^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/unwrap/_unwrap.py", line 158, in run
    exc.submit(
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/utils.py", line 680, in submit
    result = fn(*args, **kwargs)
             ^^^^^^^^^^^^^^^^^^^
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/site-packages/dolphin/unwrap/_unwrap.py", line 491, in unwrap
    shutil.rmtree(scratchdir)
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/shutil.py", line 759, in rmtree
    _rmtree_safe_fd(stack, onexc)
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/shutil.py", line 703, in _rmtree_safe_fd
    onexc(func, path, err)
  File "/home/tweiss/anaconda3/envs/dolphin-env/lib/python3.12/shutil.py", line 662, in _rmtree_safe_fd
    os.rmdir(name, dir_fd=dirfd)
OSError: [Errno 39] Directory not empty: PosixPath('/media/tweiss/data/review_joss_dolphin/20240903_dolphin/dolphin/docs/notebooks/input_slcs/work-walkthrough~/unwrapped/scratch/scratch-20210606_20210712.int')
```
The problem is with `shutil`. By trying to delete not needed processing files ([https://github.com/isce-framework/dolphin/blob/main/src/dolphin/unwrap/_unwrap.py](https://github.com/isce-framework/dolphin/blob/main/src/dolphin/unwrap/_unwrap.py) line 493) `shutil` gives me an `OSError`. If I check the directory afterwards it is actually empty. I don't know why the Error occurs but it does.
If I remove the delete option then everything works fine. I finally get the output in the `timeseries` folder.
