### 2024-09-04 d7a9513
Create `.conncomp` files for `spurt` from temporal coherence
Closes #347
