### 2024-09-04 b7785b5
remove prints, fix other bad logging issues
