+++ src/dolphin/workflows/displacement.py
@@ -64,7 +64,8 @@ def run(
     if cfg.log_file is None:
         cfg.log_file = cfg.work_directory / "dolphin.log"
     # Set the logging level for all `dolphin.` modules
-    setup_logging(debug=debug, filename=cfg.log_file)
+    for logger_name in ["dolphin", "spurt"]:
+        setup_logging(logger_name=logger_name, debug=debug, filename=cfg.log_file)
     # TODO: need to pass the cfg filename for the logger
     logger.debug(cfg.model_dump())
 
