### 2024-09-04 isce-framework/dolphin#405 @scottstanie
huh, so that `shutil` error happens on a fresh environment after deleting the working directory, and running one unwrapping job at a time (the default)? that is so weird... I was wondering if it's some python 3.12 problem, but i don't see [any relevant changes in this version](https://docs.python.org/3/whatsnew/3.12.html#shutil)

What does your environment look like with `python -c 'import dolphin; dolphin.show_versions()'` ? here's mine after i did

```
mamba env create --file conda-env.yml
conda activate dolphin-env
pip install . snaphu
python -c 'import dolphin; dolphin.show_versions()'
```

```
dolphin version: 0.26.0.post1.dev31+g10d1019

Python deps:
        h5py: 3.11.0
         jax: 0.4.31
       numba: 0.60.0
       numpy: 2.0.2
 opera-utils: 0.12.1
    pydantic: 2.8.2
      pyproj: 3.6.1
    rasterio: 1.3.11
 ruamel.yaml: 0.18.6
       scipy: 1.14.1
threadpoolctl: 3.5.0
        tqdm: 4.66.5
  osgeo.gdal: 3.9.2

System:
      python: 3.12.5 | packaged by conda-forge | (main, Aug  8 2024, 18:36:51) [GCC 12.4.0]
  executable: /u/aurora-r0/staniewi/miniconda3/envs/dolphin-env/bin/python
     machine: Linux-3.10.0-1160.118.1.el7.x86_64-x86_64-with-glibc2.17
optional GPU info:
         jax: 0.4.31
gpu_is_available: True
```

also, could you post the `dolphin_config.yaml` you are running?
