### 2024-09-05 isce-framework/dolphin#405 @McWhity
```
dolphin version: 0.27.1.post1.dev27+geeabbf9

Python deps:
        h5py: 3.11.0
         jax: 0.4.31
       numba: 0.60.0
       numpy: 2.0.2
 opera-utils: 0.12.1
    pydantic: 2.8.2
      pyproj: 3.6.1
    rasterio: 1.3.11
 ruamel.yaml: 0.18.6
       scipy: 1.14.1
threadpoolctl: 3.5.0
        tqdm: 4.66.5
  osgeo.gdal: 3.9.2

System:
      python: 3.12.5 | packaged by conda-forge | (main, Aug  8 2024, 18:36:51) [GCC 12.4.0]
  executable: /home/tweiss/miniforge3/envs/dolphin-env/bin/python
     machine: Linux-6.8.0-40-generic-x86_64-with-glibc2.35
optional GPU info:
         jax: 0.4.31
gpu_is_available: False

```
The only difference that I see is in the dolphin version number and the used machine.
