### 2024-09-05 isce-framework/dolphin#405 @scottstanie
Ok two questions-

/media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/20240903_dolphin/dolphin/docs/notebooks/input_slcs/work-walkthrough~

This is the work directory, is there supposed to be a tilde at the end?

Also, is /media a symlink to another drive? I noticed shutil had some setting about avoid symlink attacks but I haven't tried to understand that yet
