### 2024-09-05 isce-framework/dolphin#405 @McWhity
> /media/tweiss/data/Arbeit_einordnen/review_joss_dolphin/20240903_dolphin/dolphin/docs/notebooks/input_slcs/work-walkthrough~
> This is the work directory, is there supposed to be a tilde at the end?

I don't know why there is a tilde at the end. Probably a typing error from my side. I removed the tilde and I did run the process again with the same Error as output.

> Also, is /media a symlink to another drive? I noticed shutil had some setting about avoid symlink attacks but I haven't tried to understand that yet

Ok. This seems to be the problem. `/media/tweiss/data` is not actually a symlink but a mounted internal hard drive. Nevertheless, I moved the output folder  (work-warkthrough) to my home directory and there everything works fine. It is somehow strange.
I tried a quick and dirty fix by adding `ignore_errors=True`. If I change the shutil command `shutil.rmtree(scratchdir, ignore_errors=True)`, the error is ignored. The shutil command is still removing all files within folders like `scatch-20210606_20220526.int`. But as the Error seems to occur during the removing process of the empty folder, the folder itself is not deleted.
