### 2024-09-05 isce-framework/dolphin#405 @scottstanie
Thanks for testing through to find that! A very strange bug indeed, but I don't see the hard in ignore errors for that step, so I'll add that in a PR
