### 2024-09-05 028144e
Add `ignore_errors=True` to `rmtree` for scratch removal
Closes #405
