### 2024-09-05 e84dff1
Use `label` and dilation to spread create a reasonable set of labels
