### 2024-09-05 a2dcdb8
Add note on FFT dilation, bump default to 25 pixels
