### 2024-09-06 #6997 @margauxmouchene
Hello @mikemahoney218, @scottstanie, @McWhity,
I think I am done reviewing, all my comments were addressed. I was not really able to test the GPU acceleration but this is due to config issues on my side. Hopefully, @McWhity can test this properly?
Everything else is good!
Thank you for the opportunity, this project is awesome!
