### 2024-09-06 e301f3a
Create `.conncomp` files for `spurt` from temporal coherence (#417)
* Create `.conncomp` files for `spurt` from temporal coherence

Closes #347

* remove prints, fix other bad logging issues

* Use `label` and dilation to spread create a reasonable set of labels

* Add note on FFT dilation, bump default to 25 pixels

* fix typos in comments
