### 2024-09-09 c34e914
Use Gaussian in `interferogram`
