### 2024-09-10 4e6f54f
Turn off correlation weighting for velocity by default (#423)
