### 2024-09-11 #6997 @McWhity
Hello @scottstanie, @margauxmouchene, @mikemahoney218,
I think I am done with my review. So far it looks good. However there are a view things that need to be addressed. Main issues are
- Problems with GPU configuration
- Problems running tests

I tried to group my comments based on
- Paper
- Documentation
- Testing GPU setup
- Running tests

## Paper

- You mentioned other software/libraries like Miaplpy or TomoSAR. I see the advantages of dolphin but I think it would be good to clearly state them. Thus I would like see a short discussion of advantages and disadvantages. You already state one with the demand for handling heavy computational but I think there are more (something like easy handling, compiling of different algorithm stuff like that). Just make clear why we should use dolphin.
- Regarding the former mentioned author contribution. I hope @mikemahoney218 can comment on this. Is it possible/desired from journal side to include a credit of author statement in the paper? This would somehow clarify author contribution that are not visible in the commits.

## Readme

- Please add a section about contributing. Basically the first thing many people read is the README. At a later stage some people will read the documentation and find the part how they can contribute. Thus for simplicity just put the first part of the [documentation section](https://dolphin-insar.readthedocs.io/en/latest/developer-setup/) "contributing" in the README. Please keeps the links to “issue” and “discussions” so contributors can easily start issues. Additionally, I recommend to put a link to the Contributing section of the documentation for further instructions.

## Documentation

### [Section Getting started](https://dolphin-insar.readthedocs.io/en/latest/getting-started/)

- last sentence of part “Usage”: Is the link to the documentation necessary? It just directs you to the Table of contents. I see you just use the README.md within your documentation. This is smart. In this case I think you can leave it as it is.

### [Dolphin basic walkthrough](https://dolphin-insar.readthedocs.io/en/latest/notebooks/walkthrough-basic/)
- Link to ESA’s Burst database directs you to “404 Page not found”

### [Contributing](https://dolphin-insar.readthedocs.io/en/latest/developer-setup/)
- link to mkdocstrings: page is somehow strange

### Missing section/information
- Not everyone will read the JOSS paper. I recommend to add a section/information where you clearly state the “statement of need” and “the audience” you want to address. Similar to what you have done in the JOSS Paper. I know sometimes it seems that the same information is just provided at different places. But you need to keep in mind that people will find dolphin via different channels and not everyone will read the README, Documentation and the JOSS paper. Thus the information need to be provided at different places.

- In the beginning it was hard for me to determine which algorithms are used by dolphin. I suggest to incorporate a graphical overview in the JOSS Paper as well as in the Documentation. I try to explain best I can what I mean. Basically it could be kind of a workflow diagram for the default processing chains including which parts can be easily changed.
  - What is the needed input (stacks of coregistered SLC images)
  - What are the next steps and which algorithms are used.
  - Which algorithms can be changed (e.g., SNAPHU, PHASS, ICU)
  - What is the output

What I am looking for is a graphical explanation of the capabilities that dolphin provides. Thus, users that are working in SAR but not necessarily in In-SAR can understand how dolphin can help them in archiving there desired goal.

- For users it might be good if you put a default dolphin_config.yaml file in the github repository as well as in the documentation. The explanations within the yaml file are ok but when I first read the usage part of the README “dolphin run dolphin_config.yaml” I was searching for the file to checkout the options but I couldn’t find it as it is only created when trying out the notebook.

- Some more information on how to change between SNAPHU, PHASS and ICU would be also nice. Are there different parameter that can be changed for the different algorithms? What is PHASS and ICU? In the README you link to “isce3” but that does not help me to find more information about PHASS or ICU. If there is documentation to PHASS and ICU just pinpoint directly to it.

## Testing GPU setup

### Documentation
- turn path to documentation of numba and JAX into links

### I could not get the GPU up and running (Problems)
I had to search but I found a computer with NVIDIA card with CUDA support. However, I did run in some problems.

#### Attempt 1: NVIDIA driver with cuda 11.6

`pip install --upgrade "jax[cuda11_pip]" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html`
`Warning: jax 0.4.31 does not provide the extra 'cuda11.pip'`

When running the dolphin_config.yaml it get

`Warning: An NVIDIA GPU may be present on this machine, but a CUDA-enabled jaxlib is not installed. Falling back to cpu.`

Installed
```
jax 0.4.31
jaxlib 0.4.31
```

stated dependencies within dolphin are jax>=0.4.19

#### Attempt 2+3: NVIDIA driver with cuda 12.4 / 12.6

before installing jax I get the same Warning as during with cuda 11.6
`Warning: An NVIDIA GPU may be present on this machine, but a CUDA-enabled jaxlib is not installed. Falling back to cpu`.

After installing
`
pip install --upgrade "jax[cuda12_pip]" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html`

I get the follwing Error

`E external/xla/xla/stream_executor/cuda/cuda_driver.cc:216] failed call to cuInit: INTERNAL: CUDA error: Failed call to cuInit: CUDA_ERROR_NO_DEVICE: no CUDA-capable device is detected`

By the way
`mamba install cudatoolkit=12.6`
does not work as only cudatoolkit up to 11.8 can be found. Following Error occurs
`cudotoolkit 12.6** does not exist (perhaps a typo or a missing channel).`

Following code snippit does work
`mamba install nvidia/label/cuda-12.6.0::cuda-toolkit`

However I get the Error

`E external/xla/xla/stream_executor/cuda/cuda_driver.cc:216] failed call to cuInit: INTERNAL: CUDA error: Failed call to cuInit: CUDA_ERROR_NO_DEVICE: no CUDA-capable device is detected`

```
nvcc --version provides
nvcc:: NVIDIA (R) Cuda compiler driver
Copyright (c) 2005-2024 NVIDIA Corporation
Built on Fri_Jun_14_16:36:21_PDT_2024
Cuda compilation tools, release 12.6 V12.6.20
Buzild cuda_12.6.r12.65/compiler.34431801_0
```

By the way that happens although within dolphin_config.yaml `gpu_enabled: false`
It I change it to `true` the error is not shown but the script is somehow stopping

## Running tests
I usually run the tests by using `python -m unittest discover -s tests`. But this somehow did not work for me. One problem was that I had to install further packages like pytest, boto3, moto etc. The requirements are listed in the tests folder, however as the tests are part of the software please make your that all dependencies are installed during the initial installation process.  After I installed some dependencies I did run `python -m unittest discover -s tests` again but the output was
```
Ran 0 tests in 0.000s
NO TESTS RAN
```
This is were I stopped further investigating. Thus please make your that the tests can be run properly.
Just out of curiosity: How did you run the test so far?

Final comment: This is a really good and useful python packages.
