### 2024-09-11 dc22f2d
Adjust `GDAL_MAX_DATASET_POOL_SIZE` if system file limit is low
