### 2024-09-11 be741dd
add a test for ref update workflow
