+++ tests/test_workflows_displacement.py
@@ -15,10 +15,7 @@ pytestmark = pytest.mark.filterwarnings(
 )
 
 
-def test_displacement_run_single(
-    opera_slc_files: list[Path],
-    tmpdir,
-):
+def test_displacement_run_single(opera_slc_files: list[Path], tmpdir):
     with tmpdir.as_cwd():
         cfg = config.DisplacementWorkflow(
             cslc_file_list=opera_slc_files,
@@ -183,3 +180,40 @@ def test_separate_workflow_runs(slc_file_list, tmp_path):
     assert len(ifgs3_b) == 10
     # Names should be the same as the previous run
     assert [f.name for f in ifgs3_b] == [f.name for f in ifgs3]
+
+
+def test_displacement_run_extra_reference_date(opera_slc_files: list[Path], tmpdir):
+    with tmpdir.as_cwd():
+        cfg = config.DisplacementWorkflow(
+            # start_date = 20220101
+            # shape = (4, 128, 128)
+            # First one is COMPRESSED_
+            output_options={"extra_reference_date": "2022-01-03"},
+            cslc_file_list=opera_slc_files,
+            input_options={"subdataset": "/data/VV"},
+            phase_linking={
+                "ministack_size": 4,
+            },
+        )
+        paths = displacement.run(cfg)
+
+        for slc_paths in paths.comp_slc_dict.values():
+            assert all(p.exists() for p in slc_paths)
+        assert paths.stitched_ps_file.exists()
+        assert all(p.exists() for p in paths.stitched_ifg_paths)
+        assert all(p.exists() for p in paths.stitched_cor_paths)
+        assert paths.stitched_temp_coh_file.exists()
+        assert paths.stitched_ps_file.exists()
+        assert paths.stitched_amp_dispersion_file.exists()
+        assert paths.unwrapped_paths is not None
+        assert paths.conncomp_paths is not None
+        assert paths.timeseries_paths is not None
+        assert all(p.exists() for p in paths.conncomp_paths)
+        assert all(p.exists() for p in paths.unwrapped_paths)
+        assert all(full_suffix(p) == ".unw.tif" for p in paths.unwrapped_paths)
+        assert all(p.exists() for p in paths.conncomp_paths)
+        assert all(p.exists() for p in paths.timeseries_paths)
+        assert all(full_suffix(p) == ".tif" for p in paths.timeseries_paths)
+
+        # check the network size
+        assert len(paths.unwrapped_paths) > len(paths.timeseries_paths)
