### 2024-09-11 f7112fd
Setup test to check names of output ifgs
