### 2024-09-11 #6997 @margauxmouchene
> What I am looking for is a graphical explanation of the capabilities that dolphin provides

+1, great idea!

Also, I had similar issues installing cuda and ended up installing cuda-toolkit 12.2 (tried different options), then the script runs without failing but never finishes.
