### 2024-09-11 #6997 @scottstanie
Thank you for all the comments! i'll try to get more detailed responses to more of it shortly, but about your last point on the tests and unit test setup-

https://dolphin-insar.readthedocs.io/en/latest/developer-setup/

I've linked the [`CONTRIBUTING`](https://github.com/isce-framework/dolphin/blob/main/CONTRIBUTING.md) info here:
- we use [`pytest`](https://docs.pytest.org/en/stable/) for the test setup
- [the extra installs required](https://dolphin-insar.readthedocs.io/en/latest/developer-setup/#running-tests) for just the unit tests can be done with `python -m pip install -e ".[test]"`, which uses the ["Optional extra requirements" feature of the pyproject.toml setup](https://github.com/isce-framework/dolphin/blob/main/pyproject.toml#L36-L40), so then you can just be in the root of the repo and run `pytest`

I think the fact that this was unclear to you is a point in favor of what you said- "the info is scattered between JOSS, README, and the docs now". I'll make another pass and try to make sure that the development setup is clear in both docs and the readme.

As for the GPU setup... :\ I'm currently annoyed at how much of a moving target it feels like setting up any GPU libraries are. The first attempt I had on our linux server, setting up the conda environment Just Worked. But then the next time, I also had to reinstall JAX via pip to make sure it picked the right one for the CUDA libraries we had.

I'll take another look at the errors you posted and see if I can think of a clear way to augment the JAX/Numba GPU setup instructions.
