### 2024-09-12 3e9be6d
Add an `output_reference_idx` to `PhaseLinkingOptions`
This accounts for the fact that we may pass in several compressed SLCs for a
single run, and we wish to make the phase linking outputs referenced to
something besides the first SLC.

For example, if we had
   1_1_3
   1_4_6
   7_7_9
   10
   11
   12

And we wanted to output interferograms relative to `7`, we need to
specify that to the phase linking functions; otherwise, it will
output things relative to day 1. Since we dont write out rasters
corresponding to compressed SLC inputs, we'd have nothing from
day 7 to re-refrence later.

Note 
[…]
