+++ .pre-commit-config.yaml
@@ -23,6 +23,7 @@ repos:
     rev: "24.8.0"
     hooks:
       - id: black
+        args: [--preview]
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.6.3
