### 2024-09-13 013f526
Make 2 ministack idxs- one for `compressed_reference`, one for `output_reference`
