+++ src/dolphin/stack.py
@@ -49,10 +49,6 @@ class BaseStack(BaseModel):
         Path(),
         description="Folder/location where ministack will write outputs to.",
     )
-    reference_idx: int = Field(
-        0,
-        description="Index of the SLC to use as reference during phase linking",
-    )
 
     model_config = {
         # For the `Filename, so it can handle the `GeneralPath` protocol`
@@ -83,26 +79,6 @@ class BaseStack(BaseModel):
             msg = f"dates and file_list must be the same length. Got {lengths}"
             raise ValueError(msg)
         return self
-
-    @property
-    def reference_date(self):
-        """Date of the reference phase of the stack."""
-        # Note this works for either a length-1 tuple (real SLC), or for
-        # the compressed SLC formate (ref, start, end)
-        return self.dates[self.reference_idx][0]
-
-    @property
-    def full_date_range(self) -> tuple[DateOrDatetime, DateOrDatetime]:
-        """Full date range of all SLCs in the ministack."""
-        return (self.reference_date, self.dates[-1][-1])
-
-    @property
-    def full_date_range_str(self) -> str:
-        """Full date range of the ministack as a string, e.g. '20210101_20210202'.
-
-        Includes both compressed + normal SLCs in the range.
-        """
-        return format_dates(*self.full_date_range, fmt=self.file_date_fmt)
 
     @property
     def first_real_slc_idx(self) -> int:
@@ -147,7 +123,6 @@ class BaseStack(BaseModel):
         yield "reference_date", self.reference_date
         yield "file_date_fmt", self.file_date_fmt
         yield "output_folder", self.output_folder
-        yield "reference_idx", self.reference_idx
 
 
 class CompressedSlcInfo(BaseModel):
@@ -356,6 +331,25 @@ class MiniStackInfo(BaseStack):
     Used for planning the processing of a batch of SLCs.
     """
 
+    output_reference_idx: int = Field(
+        0,
+        description="Index of the SLC to use as reference during phase linking",
+    )
+    compressed_reference_idx: int = Field(
+        0,
+        description=(
+            "Index of the SLC to use as during compressed SLC creation. May be"
+            " different than `output_reference_idx`."
+        ),
+    )
+
+    @property
+    def reference_date(self):
+        """Date of the reference phase of the stack."""
+        # Note this works for either a length-1 tuple (real SLC), or for
+        # the compressed SLC formate (ref, start, end)
+        return self.dates[self.output_reference_idx][0]
+
     def get_compressed_slc_info(self) -> CompressedSlcInfo:
         """Get the compressed SLC which will come from this ministack.
 
@@ -387,9 +381,13 @@ class MiniStackPlanner(BaseStack):
     """Class for planning the processing of batches of SLCs."""
 
     max_num_compressed: int = 5
+    output_reference_idx: int = Field(
+        0,
+        description="Index of the SLC to use as reference during phase linking",
+    )
 
     def plan(
-        self, ministack_size: int, manual_idxs: Sequence[int] | None = None
+        self, ministack_size: int, compressed_idxs: Sequence[int] | None = None
     ) -> list[MiniStackInfo]:
         """Create a list of ministacks to be processed."""
         if ministack_size < 2:
@@ -397,7 +395,7 @@ class MiniStackPlanner(BaseStack):
             raise ValueError(msg)
 
         output_ministacks: list[MiniStackInfo] = []
-        manual_idx_list = sorted(manual_idxs or [])
+        compressed_idx_list = sorted(compressed_idxs or [])
 
         # Start of with any compressed SLCs that are passed in
         compressed_slc_infos: list[CompressedSlcInfo] = []
@@ -432,19 +430,23 @@ class MiniStackPlanner(BaseStack):
[…]
