### 2024-09-14 9868c0d
separate ref/compressed idx in `.plan()`, use new api for `run_wrapped_phase_sequential`
