### 2024-09-15 b673fd9
fix tests, make `single` also use `vrt_stack` instead of filename
