### 2024-09-15 f5d5f8a
fix tests for `stack.py`
