+++ tests/test_stack.py
@@ -145,7 +145,8 @@ def run_ministack_planner(files, date_lists, is_compressed):
 
         assert ms.output_folder == expected_out_folders[idx]
 
-    assert all(ms.reference_date == datetime(2022, 1, 1) for ms in ms_list)
+    assert all(ms.compressed_reference_date == datetime(2022, 1, 1) for ms in ms_list)
+    assert all(ms.output_reference_date == datetime(2022, 1, 1) for ms in ms_list)
 
 
 def test_ministack_planner_gtiff(files, date_lists, is_compressed):
@@ -242,25 +243,22 @@ def test_hit_max_compressed(slc_file_list, slc_date_list, is_compressed):
     assert all(len(m.file_list) == 6 for m in ministacks[3:-1])
 
 
-def test_manual_idx_setting(slc_file_list, slc_date_list, is_compressed):
-    """Check the planning works even after going pased the max CCSLCs."""
+def test_compressed_idx_setting(slc_file_list, slc_date_list, is_compressed):
+    """Check the planning works to set a manually passed compressed index."""
     is_compressed = [False] * len(slc_file_list)
     ministack_planner = MiniStackPlanner(
         file_list=slc_file_list,
         dates=slc_date_list,
         is_compressed=is_compressed,
         output_folder=Path("fake_dir"),
-        max_num_compressed=3,
     )
     # Check we can get the compressed SLC info
-    expected_ref_date1 = slc_date_list[1]
-    expected_ref_date2 = slc_date_list[8]
+    expected_ref_date = slc_date_list[8]
 
-    ministacks = ministack_planner.plan(3, manual_idxs=[1, 8])
+    # This currently works ONLY for one single ministack planning
+    # (e.g. manually separating minstack runs and re-passeding only CCSLCs)
+    ministacks = ministack_planner.plan(len(slc_file_list), compressed_idxs=[8])
 
-    assert ministacks[0].reference_date == expected_ref_date1
-    assert ministacks[1].reference_date == expected_ref_date1
-    # Since we did size-3 planning, i.e. (0,3), (3,6), (6,9),...
-    # the 3rd ministack has the changeover to the next manual idx
-    for i in range(2, len(ministacks)):
-        assert ministacks[i].reference_date == expected_ref_date2
+    assert ministacks[0].compressed_reference_date == expected_ref_date
+    # But, the output reference date should be the first one (since it was unset)
+    assert ministacks[0].output_reference_date == slc_date_list[0]
