### 2024-09-15 e6dbe74
let only 1 compressed reference idx be passed to `.plan`
