### 2024-09-15 d3e0d4f
clean up commented, pass the `output_reference_index` to `run_phase_linking` in `single`
