### 2024-09-15 78e6375
Add logic for manually specifying reference dates mid-stack (#334)
* use the `reference_idx` from each `ministack`, dont hardcode to 0

* remove commented `stack.py` code

* Add a `reference_idx` arg to `compress`

* update `MiniStackPlanner` to accept manual reference indexes

* start passing through logic to account for manual reference changeovers

* fix num parallel to respect the file length

* get the linked phase part working with manual references

* move the reference date to be an output option

* get basic single-reference changeover working for interferograms

* specify naive datet
[…]
