### 2024-09-16 5c52f6c
Add ability to specify output bounds as WKT
