### 2024-09-16 780e517
Set the input to EVD to be weighted by correlation
