### 2024-09-16 0a4de23
allow .wkt file too
