### 2024-09-16 f32e3ea
check for none
