+++ .pre-commit-config.yaml
@@ -26,7 +26,7 @@ repos:
         args: [--preview]
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.6.4
+    rev: v0.6.5
     hooks:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix, --show-fixes]
