### 2024-09-17 a16c6b5
Print unwrapping options in `show_versions` (#430)
* Print unwrapping options in `show_versions`

* Fix prints, remove unused private from `wrapped_phase`
