### 2024-09-18 d217ff2
move iono file parsing to separate function
