### 2024-09-18 b29c726
fix test to have `R`, single residual
