### 2024-09-18 96e81cc
Fix ionosphereic date parsing and grouping (#432)
* move iono file parsing to separate function

* Update src/dolphin/workflows/_utils.py
