### 2024-09-19 6ed5c92
add boyd to references
