+++ docs/references.bib
@@ -41,6 +41,18 @@
   keywords = {Big Data,Decorrelation,deformation estimation,differential interferometric synthetic aperture radar (SAR) (DInSAR),distributed scatterers (DSs),error analysis,Fading channels,Moisture,near real-time (NRT) processing,phase inconsistencies,signal decorrelation,Strain,Synthetic aperture radar,Systematics,Time series analysis,time-series analysis}
 }
 
+@article{Boyd2010DistributedOptimizationStatistical,
+  title = {Distributed {{Optimization}} and {{Statistical Learning}} via the {{Alternating Direction Method}} of {{Multipliers}}},
+  author = {Boyd, Stephen},
+  year = {2010},
+  journal = {Foundations and Trends\textregistered{} in Machine Learning},
+  volume = {3},
+  number = {1},
+  pages = {1--122},
+  issn = {1935-8237, 1935-8245},
+  doi = {10.1561/2200000016}
+}
+
 @article{Chen2012IonosphericArtifactsSimultaneous,
   title = {Ionospheric {{Artifacts}} in {{Simultaneous L-Band InSAR}} and {{GPS Observations}}},
   author = {Chen, Jingyi and Zebker, Howard A.},
