### 2024-09-19 48b6d92
Add references to docstrings, background to L1 notebook.
