+++ docs/references.bib
@@ -41,6 +41,19 @@
   keywords = {Big Data,Decorrelation,deformation estimation,differential interferometric synthetic aperture radar (SAR) (DInSAR),distributed scatterers (DSs),error analysis,Fading channels,Moisture,near real-time (NRT) processing,phase inconsistencies,signal decorrelation,Strain,Synthetic aperture radar,Systematics,Time series analysis,time-series analysis}
 }
 
+@article{Berardino2002NewAlgorithmSurface,
+  title = {A {{New Algorithm}} for {{Surface Deformation Monitoring Based}} on {{Small Baseline Differential SAR Interferograms}}},
+  author = {Berardino, Paolo and Fornaro, Dianfranco and Linari, Riccardo and Sansosti, Eugenio},
+  year = {2002},
+  journaltitle = {IEEE International Geoscience and Remote Sensing Symposium},
+  volume = {2},
+  number = {11},
+  pages = {1237--1239},
+  issn = {01962892},
+  doi = {10.1109/IGARSS.2002.1025900},
+  url = {https://ieeexplore.ieee.org/stamp/stamp.jsp?arnumber=1166596}
+}
+
 @article{Boyd2010DistributedOptimizationStatistical,
   title = {Distributed {{Optimization}} and {{Statistical Learning}} via the {{Alternating Direction Method}} of {{Multipliers}}},
   author = {Boyd, Stephen},
@@ -107,6 +120,18 @@
   issn = {0196-2892},
   doi = {10.1109/TGRS.2008.2001756},
   keywords = {Decorrelation,decorrelation models,ENVISAT images,geophysical techniques,geophysics computing,image processing,Information retrieval,interferometric phases,Interferometry,line-of-sight displacement,line-of-sight motion,Maximum likelihood estimation,Monte Carlo simulations,multiimage synthetic aperture radar interferometry,Phase estimation,physical parameters,radar interferometry,Radar scattering,remote sensing by radar,residual topography,SAR interferometry applications,Statistical distributions,statistics,Statistics,stochastic processes,Surfaces,synthetic aperture radar,Synthetic aperture radar interferometry,target statistics,topography (Earth),Yield estimation}
+}
+
+@article{Lauknes2011InSARDeformationTime,
+  title = {{{InSAR Deformation Time Series Using}} an {\textsubscript{1}} -{{Norm Small-Baseline Approach}}},
+  author = {Lauknes, Tom R. and Zebker, Howard A. and Larsen, Yngvar},
+  date = {2011-01},
+  journaltitle = {IEEE Transactions on Geoscience and Remote Sensing},
+  volume = {49},
+  number = {1},
+  pages = {536--546},
+  issn = {1558-0644},
+  doi = {10.1109/TGRS.2010.2051951}
 }
 
 @article{Mirzaee2023NonlinearPhaseLinking,
