### 2024-09-19 5b10f9c
[skip ci] Fix `least_absolute_deviations` docstring
