### 2024-09-19 94a5c85
Use ADMM to minimize L1-norm of `Ax - b` (#431)
* Use ADMM to minimize L1-norm of `Ax - b`

* replace `irls` test with lad test

* fix test to have `R`, single residual

* add boyd to references

* Add references to docstrings, background to L1 notebook.

* [skip ci] Fix `least_absolute_deviations` docstring
