+++ src/dolphin/timeseries.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from tempfile import NamedTemporaryFile
 from typing import Callable, Optional, Protocol, Sequence, TypeVar
 
+import jax
 import jax.numpy as jnp
 import numpy as np
 from jax import Array, jit, lax, vmap
@@ -811,6 +812,7 @@ def invert_unw_network(
     if all(p.exists() for p in out_paths):
         logger.info("All output files already exist, skipping inversion")
         return out_paths
+    logger.info(f"Inverting network using {method.upper()}-norm minimization")
 
     A = get_incidence_matrix(ifg_pairs=ifg_tuples, sar_idxs=sar_dates)
 
@@ -1052,90 +1054,105 @@ def _get_largest_conncomp_mask(
     return isin_largest_conncomp
 
 
-@jit
-def invert_stack_l1(A: ArrayLike, dphi: ArrayLike) -> Array:
-    n_ifgs, n_rows, n_cols = dphi.shape
-
-    # vectorize the solve function to work on 2D and 3D arrays
-    # We are not vectorizing over the A matrix, only the dphi vector
-    # Solve 2d shapes: (nrows, n_ifgs) -> (nrows, n_sar_dates)
-    invert_2d = vmap(irls, in_axes=(None, 1), out_axes=(1, 1))
-    # Solve 3d shapes: (nrows, ncols, n_ifgs) -> (nrows, ncols, n_sar_dates)
-    invert_3d = vmap(invert_2d, in_axes=(None, 2), out_axes=(2, 2))
-    phase, residual_vecs = invert_3d(A, dphi)
-    # Reshape the residuals to be 2D
-    residuals = jnp.sum(residual_vecs, axis=0)
-
-    return phase, residuals
+@jax.jit
+def _shrinkage(a: jnp.ndarray, kappa: float) -> jnp.ndarray:
+    """Apply the shrinkage operator element-wise."""
+    return jnp.maximum(0, a - kappa) - jnp.maximum(0, -a - kappa)
 
 
-@jit
-def irls(
+@jax.jit
+def least_absolute_deviations(
     A: jnp.ndarray,
     b: jnp.ndarray,
-    p: float = 1,
-    max_iters: int = 50,
-    tol: float = 1e-5,
-) -> tuple[jnp.ndarray, float]:
-    """Minimize |Ax - b|_1 using Iteratively reweighted least squares (IRLS).
+    R: jnp.ndarray,
+    rho: float = 0.4,
+    alpha: float = 1.0,
+    max_iter: int = 20,
+) -> jnp.ndarray:
+    """Solve Least Absolute Deviations (LAD) via ADMM.
 
-    See https://en.wikipedia.org/wiki/Iteratively_reweighted_least_squares for
-    algorithm description
+    Solves the following problem via ADMM:
+
+        minimize ||Ax - b||_1
+
+    See [@Boyd2010DistributedOptimizationStatistical] for more on ADMM.
 
     Parameters
     ----------
     A : jnp.ndarray
         The matrix A in the equation Ax = b.
+        Shape is (M, N)
     b : jnp.ndarray
         The vector b in the equation Ax = b.
-    p : float, optional
-        The power parameter for the weights, by default 1.
-    max_iters : int, optional
-        The maximum number of iterations, by default 50.
-    tol : float, optional
-        The tolerance for convergence, by default 1e-5.
+        Shape is (M,)
+    R : jnp.ndarray
+        Precomputed lower-triangular Cholesky factor of A^T A.
+        Shape is (N, N)
+    rho : float, optional
+        The augmented Lagrangian parameter
+        By default 0.4.
+    alpha : float, optional
[…]
