### 2024-09-19 39ae410
Fix handling of `nodata` during timeseries conversion to meters
