### 2024-09-19 32b4df7
Fix handling of `nodata` during timeseries conversion to meters (#433)
* Fix handling of `nodata` during timeseries conversion to meters

* add `nodata` value check to workflow test
