### 2024-09-24 b8ad94c
Dont short-circuit `extra_reference_date` change over (#435)
