### 2024-09-26 ed180b1
Skip network inversion for any set of `n-1` interferograms
