### 2024-09-26 868d70b
Change `is_single_reference` to only look for number of interferograms (#436)
* Skip network inversion for any set of `n-1` interferograms

* dedupe a possible `dayN_dayN` interferogram

* remove comments from requirements files

* fix ref change test
