+++ src/dolphin/timeseries.py
@@ -129,6 +129,7 @@ def run(
 
     condition_func = argmax_index if condition == CallFunc.MAX else argmin_index
     if reference_point == (-1, -1):
+        logger.info("Selecting a reference point for unwrapped interferograms")
         ref_point = select_reference_point(
             condition_file=condition_file,
             output_dir=Path(output_dir),
@@ -142,9 +143,11 @@ def run(
     sar_dates = sorted(set(utils.flatten(ifg_date_pairs)))
     # if we did single-reference interferograms, for `n` sar dates, we will only have
     # `n-1` interferograms. Any more than n-1 ifgs means we need to invert
-    is_single_reference = (len(unwrapped_paths) == len(sar_dates) - 1) and all(
-        pair[0] == ifg_date_pairs[0][0] for pair in ifg_date_pairs
-    )
+    is_single_reference = len(unwrapped_paths) == len(sar_dates) - 1
+    # TODO: Do we ever want to invert this case: the "trivial" network,
+    # which has 1 ifg per date difference, but a moving reference date?
+    # The extra condition to check is
+    # ... and all(pair[0] == ifg_date_pairs[0][0] for pair in ifg_date_pairs)
 
     # check if we even need to invert, or if it was single reference
     inverted_phase_paths: list[Path] = []
@@ -160,8 +163,6 @@ def run(
             wavelength=wavelength,
         )
     else:
-        logger.info("Selecting a reference point for unwrapped interferograms")
-
         logger.info("Inverting network of %s unwrapped ifgs", len(unwrapped_paths))
         inverted_phase_paths = invert_unw_network(
             unw_file_list=unwrapped_paths,
