+++ docs/requirements.txt
@@ -1,5 +1,4 @@
-# mkdocs requirements
-markdown # for mdx_bib
+markdown
 mkdocs
 mkdocs-gen-files
 mkdocs-jupyter
@@ -7,5 +6,5 @@ mkdocs-literate-nav
 mkdocs-material
 mkdocs-section-index
 mkdocstrings[python]
-pybtex # for mdx_bib
+pybtex
 pymdown-extensions
