### 2024-09-30 f967eac
fix `mkdocstrings` link in `CONTRIBUTING`
