### 2024-09-30 5663cfd
Add "Contributing" section to the README
