### 2024-10-01 16be618
Fix the self-referential link to documentation in `getting-started`
