### 2024-10-01 56770dd
Add badge at top for Docs, missing `spurt` unwrapping option to README
