+++ docs/getting-started.md
@@ -1,4 +1,4 @@
---8<-- "README.md:19:95"
+--8<-- "README.md:25:102"
 
 For a more complete tutorial, see our [notebooks](notebooks)
 
