### 2024-10-01 234db5c
Add a sample `dolphin_config.yaml` to the docs, GHA to auto-update
