+++ docs/sample_dolphin_config.yaml
@@ -1,0 +1,330 @@
+input_options:
+  # If passing HDF5/NetCDF files, subdataset to use from CSLC files. .
+  #   Type: string | null.
+  subdataset:
+  # Format of dates contained in CSLC filenames.
+  #   Type: string.
+  cslc_date_fmt: '%Y%m%d'
+# Mask file used to ignore low correlation/bad data (e.g water mask). Convention is 0 for no
+#   data/invalid, and 1 for good data. Dtype must be uint8.
+#   Type: string | null.
+mask_file:
+# Name of sub-directory to use for writing output files.
+#   Type: string.
+work_directory: /Users/staniewi/repos/dolphin
+# Don't resolve filepaths that are given as relative to be absolute.
+#   Type: boolean.
+keep_paths_relative: false
+worker_settings:
+  # Whether to use GPU for processing (if available).
+  #   Type: boolean.
+  gpu_enabled: false
+  # Number of threads to use per worker. This sets the OMP_NUM_THREADS environment variable in
+  #   each python process.
+  #   Type: integer.
+  threads_per_worker: 1
+  # If processing separate spatial bursts, number of bursts to run in parallel for wrapped-
+  #   phase-estimation.
+  #   Type: integer.
+  n_parallel_bursts: 1
+  # Size (rows, columns) of blocks of data to load at a time.
+  #   Type: array.
+  block_shape:
+    - 512
+    - 512
+# Path to output log file (in addition to logging to `stderr`). Default logs to
+#   `dolphin.log` within `work_directory`.
+#   Type: string | null.
+log_file:
+# Time the config file was created.
+#   Type: string.
+creation_time_utc: '2024-10-01T01:37:59.922824Z'
+# REQUIRED: list of CSLC files, or newline-delimited file containing list of CSLC files.
+#   Type: array.
+cslc_file_list: []
+output_options:
+  # Output (x, y) resolution (in units of input data).
+  #   Type: object | null.
+  output_resolution:
+  # Alternative to specifying output resolution: Specify the (x, y) strides (decimation
+  #   factor) to perform while processing input. For example, strides of [4, 2] would turn an
+  #   input resolution of [5, 10] into an output resolution of [20, 20].
+  #   Type: object.
+  strides:
+    x: 1
+    y: 1
+  # Area of interest: [left, bottom, right, top] coordinates. e.g.
+  #   `bbox=[-150.2,65.0,-150.1,65.5]`.
+  #   Type: array | null.
+  bounds:
+  # EPSG code for the `bounds` coordinates, if specified.
+  #   Type: integer.
+  bounds_epsg: 4326
+  # Options for `create_dataset` with h5py.
+  #   Type: object.
+  hdf5_creation_options:
+    chunks:
+      - 128
+      - 128
+    compression: gzip
+    compression_opts: 4
+    shuffle: true
+  # GDAL creation options for GeoTIFF files.
+  #   Type: array.
+  gtiff_creation_options:
+    - COMPRESS=lzw
+    - ZLEVEL=4
+    - BIGTIFF=yes
+    - TILED=yes
+    - INTERLEAVE=band
+    - BLOCKXSIZE=128
+    - BLOCKYSIZE=128
+  # Whether to add overviews to the output GeoTIFF files. This will increase file size, but
+  #   can be useful for visualizing the data with web mapping tools. See
+  #   https://gdal.org/programs/gdaladdo.html for more.
+  #   Type: boolean.
+  add_overviews: true
+  # List of overview levels to create (if `add_overviews=True`).
+  #   Type: array.
+  overview_levels:
+    - 4
+    - 8
+    - 16
+    - 32
+    - 64
+ps_options:
+  # Amplitude dispersion threshold to consider a pixel a PS.
+  #   Type: number.
+  amp_dispersion_threshold: 0.25
[…]
