### 2024-10-01 #6997 @scottstanie
Hi @mikemahoney218 , sorry for the very long delay- I have made responses to all but the "graphical overview/ processing chain diagrams"- should I post the partial responses while I finish that new figure? Or wait to post one reply?

On the GPU setup problem:
Sorry to both reviewers that it was difficult to set up the GPU version! It feels like it can be a moving target as the libraries evolve.

I tried to include that walkthrough of what worked for me https://dolphin-insar.readthedocs.io/en/latest/gpu-setup/ , but I'm hesitant to add many more details beyond linking to the libraries' own GPU setup pages. Each of them has many paragraphs/users raising issues, so I would imaging a `dolphin` page quickly growing stale.

For instance, on the `numba` page, I see they now suggest running
```
conda install -c conda-forge cuda-nvcc cuda-nvrtc "cuda-version>=12.0"
```
since it looks like Nvidia has changed how the distribute their `cudatoolkit` and have their own anaconda channel.

One other thing that could be leading to errors if you have the ability to do `import jax.numpy; jax.numpy.array([1,2,3])` without any GPU warnings/errors: in the configuration `worker_settings`, there is `enable_gpu`. This defaults to false because of the errors you've mentioned, and because our main testing has been to make small AWS CPU machines run well (despite testing on a GPU-compatible workstation).
