### 2024-10-01 #6997 @mikemahoney218
No worries, I know how busy this time of year tends to be for everyone!

> should I post the partial responses while I finish that new figure? Or wait to post one reply?

One reply is probably best!
