+++ .github/workflows/update-sample-config.yaml
@@ -28,11 +28,12 @@ jobs:
           pip install --no-deps .
       - name: Generate default config
         run: |
-          dolphin config --print-empty > docs/sample_dolphin_config.yaml
+          dolphin config --print-empty -o temp_config.yaml
+          sed '/creation_time_utc:/d' temp_config.yaml > docs/sample_dolphin_config.yaml
       - name: Check for changes
         id: git-check
         run: |
-          git add sample_dolphin_config.yaml
+          git add docs/sample_dolphin_config.yaml
           git diff --staged --exit-code || echo "changes=true" >> $GITHUB_OUTPUT
       - name: Create Pull Request
         if: steps.git-check.outputs.changes == 'true'
