### 2024-10-01 771eda1
Add references/url fix from JOSS branch
