+++ docs/references.bib
@@ -41,6 +41,17 @@
   keywords = {Big Data,Decorrelation,deformation estimation,differential interferometric synthetic aperture radar (SAR) (DInSAR),distributed scatterers (DSs),error analysis,Fading channels,Moisture,near real-time (NRT) processing,phase inconsistencies,signal decorrelation,Strain,Synthetic aperture radar,Systematics,Time series analysis,time-series analysis}
 }
 
+@article{Bekaert2021IntroducingOPERAProject,
+  title = {Introducing the {{OPERA Project}} for {{Systematic Surface Water}}, {{Surface Deformation}}, and {{Surface Disturbance Data Products}} from {{Satellite Observations}}},
+  author = {Bekaert, David and Fattahi, Heresh and Jones, John and Hansen, Matthew and Kwoun, Oh-Ig and Lewis, Steven and Meyer, Franz and Osmanoglu, Batuhan and Marshak, Charles and Hamlington, Benjamin and Bato, Mary Grace and Brancato, Virginia and Shiroma, Gustavo and Jung, Jungkyo and Jai, Benhan and Buckley, Sean and Cruz, Jennifer},
+  year = {2021},
+  month = dec,
+  volume = {2021},
+  pages = {G42A-07},
+  urldate = {2024-05-15}
+}
+
+
 @article{Berardino2002NewAlgorithmSurface,
   title = {A {{New Algorithm}} for {{Surface Deformation Monitoring Based}} on {{Small Baseline Differential SAR Interferograms}}},
   author = {Berardino, Paolo and Fornaro, Dianfranco and Linari, Riccardo and Sansosti, Eugenio},
@@ -66,6 +77,26 @@
   doi = {10.1561/2200000016}
 }
 
+@misc{Bradbury2018JAXComposableTransformations,
+  title = {{{JAX}}: Composable Transformations of {{Python}}+{{NumPy}} Programs},
+  author = {Bradbury, James and Frostig, Roy and Hawkins, Peter and Johnson, Matthew James and Leary, Chris and Maclaurin, Dougal and Necula, George and Paszke, Adam and VanderPlas, Jake and {Wanderman-Milne}, Skye and Zhang, Qiao},
+  year = {2018}
+}
+
+@article{Chen2001TwodimensionalPhaseUnwrapping,
+  title = {Two-Dimensional Phase Unwrapping with Use of Statistical Models for Cost Functions in Nonlinear Optimization},
+  author = {Chen, Curtis W. and Zebker, Howard A.},
+  year = {2001},
+  month = feb,
+  journal = {Journal of the Optical Society of America A},
+  volume = {18},
+  number = {2},
+  pages = {338},
+  issn = {1084-7529, 1520-8532},
+  doi = {10.1364/JOSAA.18.000338},
+  urldate = {2024-05-16},
+  langid = {english}
+}
 @article{Chen2012IonosphericArtifactsSimultaneous,
   title = {Ionospheric {{Artifacts}} in {{Simultaneous L-Band InSAR}} and {{GPS Observations}}},
   author = {Chen, Jingyi and Zebker, Howard A.},
@@ -91,6 +122,32 @@
   issn = {19448007},
   doi = {10.1002/2015GL065031},
   keywords = {decorrelation,groundwater,InSAR deformation map,persistent scatterers}
+}
+
+@article{Fattahi2019FRInGEFullResolutionInSAR,
+  title = {{{FRInGE}}; {{Full-Resolution InSAR}} Timeseries Using {{Generalized Eigenvectors}}},
+  author = {Fattahi, H. and Agram, P. S. and Tymofyeyeva, E. and Bekaert, D. P.},
+  year = {2019},
+  month = dec,
+  volume = {2019},
+  pages = {G11B-0514},
+  urldate = {2024-05-16},
[…]
