### 2024-10-01 d9a9bf6
add bash line
