+++ .github/workflows/update-sample-config.yaml
@@ -10,6 +10,9 @@ on:
 jobs:
   update-config:
     runs-on: ubuntu-latest
+    defaults:
+      run:
+        shell: bash -l {0}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
