### 2024-10-01 68465d8
fix the `work_directory` resolving
