### 2024-10-01 f50cd5a
cache mamba envs for the day
