+++ .github/workflows/update-sample-config.yaml
@@ -21,6 +21,8 @@ jobs:
         with:
           environment-file: conda-env.yml
           environment-name: dolphin-env
+          cache-environment-key: environment-${{ steps.date.outputs.date }}
+          cache-downloads-key: downloads-${{ steps.date.outputs.date }}
           generate-run-shell: false
           create-args: ${{ matrix.deps.spec }}
           condarc: |
