### 2024-10-01 bd9be1d
Add a sample `dolphin_config.yaml` to the docs, GHA to auto-update (#438)
* Add a sample `dolphin_config.yaml` to the docs, GHA to auto-update

* add link to readme to sample file

* fix two warnings from mkdocs

* update for merged main

* remove actual creation time, subject to change

* move the new job into existing workflow

* Revert "move the new job into existing workflow"

This reverts commit a0b0b2cd1e94bf119b87c62e82c0fd423350a519.

* add bash line

* fix the `work_directory` resolving

* add `base` to PR GHA

* cache mamba envs for the day
