+++ .github/workflows/test-build-push.yml
@@ -36,6 +36,9 @@ jobs:
         with:
           environment-file: conda-env.yml
           environment-name: dolphin-env
+          # persist on the same day.
+          cache-environment-key: environment-${{ steps.date.outputs.date }}
+          cache-downloads-key: downloads-${{ steps.date.outputs.date }}
           generate-run-shell: false
           create-args: ${{ matrix.deps.spec }}
           condarc: |
