### 2024-10-03 d054e4c
Bump mamba-org/setup-micromamba from 1 to 2 in the actions group
Bumps the actions group with 1 update: [mamba-org/setup-micromamba](https://github.com/mamba-org/setup-micromamba).


Updates `mamba-org/setup-micromamba` from 1 to 2
- [Release notes](https://github.com/mamba-org/setup-micromamba/releases)
- [Commits](https://github.com/mamba-org/setup-micromamba/compare/v1...v2)

---
updated-dependencies:
- dependency-name: mamba-org/setup-micromamba
  dependency-type: direct:production
  update-type: version-update:semver-major
  dependency-group: actions
...

Signed-off-by: dependabot[bot] <s
[…]
