### 2024-10-03 a1eba7a
Bump mamba-org/setup-micromamba from 1 to 2 in the actions group (#441)
