### 2024-10-05 c0f3791
Add re-referencing for network unwrappers using `extra_reference_date`
