### 2024-10-06 d1f1c70
move out to separate helper function
