### 2024-10-08 aaa73c8
sort output paths
