### 2024-10-08 0c682a5
Convert `glrt` SHP method to use JAX
