+++ src/dolphin/shp/_glrt.py
@@ -1,99 +1,90 @@
 from __future__ import annotations
 
 import csv
-from functools import lru_cache
-from math import log
+from functools import lru_cache, partial
 from pathlib import Path
-from typing import Optional
 
-import numba
-import numpy as np
+import jax.numpy as jnp
+from jax import Array, jit, lax, vmap
 from numpy.typing import ArrayLike
 
-from dolphin._types import Strides
-from dolphin.utils import _get_slices, compute_out_shape
-
-from ._common import remove_unconnected
-
-_get_slices = numba.njit(_get_slices)
+from dolphin._types import HalfWindow
+from dolphin.utils import compute_out_shape
 
 
+@lru_cache
+def _read_cutoff_csv():
+    filename = Path(__file__).parent / "glrt_cutoffs.csv"
+
+    result = {}
+    with open(filename) as file:
+        reader = csv.DictReader(file)
+        for row in reader:
+            n = int(row["N"])
+            alpha = float(row["alpha"])
+            cutoff = float(row["cutoff"])
+            result[(n, alpha)] = cutoff
+
+    return result
+
+
+@partial(jit, static_argnames=["half_window", "strides", "nslc", "alpha"])
 def estimate_neighbors(
     mean: ArrayLike,
     var: ArrayLike,
-    halfwin_rowcol: tuple[int, int],
+    half_window: HalfWindow,
     nslc: int,
-    strides: Optional[dict] = None,
-    alpha: float = 0.05,
-    prune_disconnected: bool = False,
+    strides: tuple[int, int] = (1, 1),
+    alpha: float = 0.001,
 ):
-    """Estimate the number of neighbors based on the GLRT.
-
-    Based on the method described in [@Parizzi2011AdaptiveInSARStack].
-    Assumes Rayleigh distributed amplitudes ([@Siddiqui1962ProblemsConnectedRayleigh])
-
-    Parameters
-    ----------
-    mean : ArrayLike, 2D
-        Mean amplitude of each pixel.
-    var: ArrayLike, 2D
-        Variance of each pixel's amplitude.
-    halfwin_rowcol : tuple[int, int]
-        Half the size of the block in (row, col) dimensions
-    nslc : int
-        Number of images in the stack used to compute `mean` and `var`.
-        Used to compute the degrees of freedom for the t- and F-tests to
-        determine the critical values.
-    strides: dict, optional
-        The (x, y) strides (in pixels) to use for the sliding window.
-        By default {"x": 1, "y": 1}
-    alpha : float, default=0.05
-        Significance level at which to reject the null hypothesis.
-        Rejecting means declaring a neighbor is not a SHP.
-    prune_disconnected : bool, default=False
-        If True, keeps only SHPs that are 8-connected to the current pixel.
-        Otherwise, any pixel within the window may be considered an SHP, even
-        if it is not directly connected.
-
-
-    Notes
-    -----
-    When `strides` is not (1, 1), the output first two dimensions
-    are smaller than `mean` and `var` by a factor of `strides`. This
-    will match the downstream shape of the strided phase linking results.
-
-    Returns
-    -------
-    is_shp : np.ndarray, 4D
-        Boolean array marking which neighbors are SHPs for each pixel in the block.
-        Shape is (out_rows, out_cols, window_rows, window_cols), where
-            `out_rows` and `out_cols` are computed by
-            `[dolphin.io.compute_out_shape][]`
-            `window_rows = 2 * halfwin_rowcol[0] + 1`
[…]
