### 2024-10-10 c48894b
Pass through `phase_linking.output_reference_idx` to avoid reference resets
