### 2024-10-10 aebe38c
for spurt, override the `interferogram_network` configuration
