### 2024-10-10 777ca34
Modify `io.process_blocks` to accept overlapping blocks
Result is similar to `map_overlaps`, but with arbitrary number of input or output rasters
