### 2024-10-10 b3b42e7
create and pass through similarity raster
