### 2024-10-10 d51043f
fix existing similarity file search
