### 2024-10-11 a4dc4dd
use `nan` for nodata, mask pixels with all nan/zero
