### 2024-10-11 0cb7642
fix strides and tests, add edge for extra data
