+++ src/dolphin/workflows/single.py
@@ -33,6 +33,7 @@ class OutputFile:
     dtype: DTypeLike
     strides: Optional[dict[str, int]] = None
     nbands: int = 1
+    nodata: float = 0
 
 
 @atomic_output(output_arg="output_folder", is_dir=True)
@@ -124,7 +125,12 @@ def run_wrapped_phase_single(
         ),
         OutputFile(output_folder / f"shp_counts_{start_end}.tif", np.uint16, strides),
         OutputFile(output_folder / f"eigenvalues_{start_end}.tif", np.float32, strides),
-        OutputFile(output_folder / f"estimator_{start_end}.tif", np.int8, strides),
+        OutputFile(
+            output_folder / f"estimator_{start_end}.tif",
+            np.int8,
+            strides,
+            nodata=255,
+        ),
         OutputFile(output_folder / f"avg_coh_{start_end}.tif", np.uint16, strides),
     ]
     for op in output_files:
@@ -135,7 +141,7 @@ def run_wrapped_phase_single(
             dtype=op.dtype,
             strides=op.strides,
             nbands=op.nbands,
-            nodata=0,
+            nodata=op.nodata,
         )
 
     # Iterate over the output grid
@@ -182,7 +188,7 @@ def run_wrapped_phase_single(
         neighbor_arrays = shp.estimate_neighbors(
             halfwin_rowcol=(yhalf, xhalf),
             alpha=shp_alpha,
-            strides=strides,
+            strides=tuple(strides_tup),
             mean=amp_mean[in_rows, in_cols] if amp_mean is not None else None,
             var=amp_variance[in_rows, in_cols] if amp_variance is not None else None,
             nslc=shp_nslc,
@@ -265,7 +271,7 @@ def run_wrapped_phase_single(
         )
 
         # All other outputs are strided (smaller in size)
-        out_datas = [
+        out_datas: list[np.ndarray | None] = [
             pl_output.temp_coh,
             pl_output.shp_counts,
             pl_output.eigenvalues,
@@ -275,8 +281,9 @@ def run_wrapped_phase_single(
         for data, output_file in zip(out_datas, output_files[1:]):
             if data is None:  # May choose to skip some outputs, e.g. "avg_coh"
                 continue
+            trimmed_data = data[out_trim_rows, out_trim_cols]
             writer.queue_write(
-                data[out_trim_rows, out_trim_cols],
+                _erode_edge_pixels(trimmed_data, nodata=output_file.nodata),
                 output_file.filename,
                 out_rows.start,
                 out_cols.start,
@@ -410,3 +417,15 @@ def setup_output_folder(
 
         phase_linked_slc_files.append(output_path)
     return phase_linked_slc_files
+
+
+def _erode_edge_pixels(arr: np.ndarray, nodata: float, n: int = 1) -> np.ndarray:
+    from scipy import ndimage
+
+    mask = arr == nodata if not np.isnan(nodata) else np.isnan(arr)
+    mask_expanded = ndimage.binary_dilation(
+        mask, structure=np.ones((1 + 2 * n, 1 + 2 * n))
+    )
+    out = arr.copy()
+    out[mask_expanded] = nodata
+    return out
