### 2024-10-11 e4c7378
fix sim tests
