### 2024-10-12 26fa88f
add `IfgStackReader` for pre-formed nearest-3 interferograms
