+++ src/dolphin/unwrap/_unwrap_3d.py
@@ -4,7 +4,9 @@ from pathlib import Path
 from typing import Sequence
 
 import numpy as np
+import rasterio as rio
 from opera_utils import get_dates
+from rasterio.windows import Window
 from scipy import ndimage, signal
 
 from dolphin import io, utils
@@ -63,18 +65,35 @@ def unwrap_spurt(
     mrg_settings = MergerSettings(**options.merger_settings.model_dump())
 
     # Using default Hop3Graph
-    # TODO: this is a weird hack.. if there are 15 dates, there are 14 interferograms
-    # the spurt cli expects one of the filenames to be all 0s? maybe?
-    # But also still expects them to be date1_date2.int.tif?
-    g_time = Hop3Graph(len(ifg_filenames) + 1)
+    ifg_date_tuples = [get_dates(f) for f in ifg_filenames]
+    all_dates = sorted(set(utils.flatten(ifg_date_tuples)))
+    g_time = Hop3Graph(len(all_dates))
     logger.info(f"Using Hop3 Graph in time with { g_time.npoints } epochs.")
 
-    date_str_to_file = _map_date_str_to_file(ifg_filenames)
-    stack = SLCStackReader(
-        slc_files=date_str_to_file,
-        temp_coh_file=temporal_coherence_file,
-        temp_coh_threshold=options.temporal_coherence_threshold,
-    )
+    if len(all_dates) == len(ifg_filenames) + 1:
+        # Single reference interferograms: do they all have same first date?
+        if len({d[0] for d in ifg_date_tuples}) > 1:
+            errmsg = "interferograms for spurt must be single reference."
+            raise ValueError(errmsg)
+        logger.info(
+            "Converting single-reference interferograms to Hop3 during unwrapping"
+        )
+        date_str_to_file = _map_date_str_to_file(ifg_filenames)
+        stack = SLCStackReader(
+            slc_files=date_str_to_file,
+            temp_coh_file=temporal_coherence_file,
+            temp_coh_threshold=options.temporal_coherence_threshold,
+        )
+    elif len(ifg_filenames) == len(g_time.links):
+        logger.info("Using pre-formed nearest 3 interferograms")
+        stack = IfgStackReader(
+            ifg_filenames=ifg_filenames,
+            quality_file=temporal_coherence_file,
+            threshold=options.temporal_coherence_threshold,
+        )
+    else:
+        raise ValueError("spurt requires nearest-3 interferograms, or single-reference")
+
     # Run the workflow
     # Generate tiles
     get_tiles(stack, gen_settings, tile_settings)
@@ -103,22 +122,17 @@ def unwrap_spurt(
 def _map_date_str_to_file(
     ifg_filenames: Sequence[PathOrStr], date_fmt: str = "%Y%m%d"
 ) -> dict[str, PathOrStr | None]:
-    # Then list individual SLCs
-    dates = [get_dates(f) for f in ifg_filenames]
-    if len({d[0] for d in dates}) > 1:
-        errmsg = "interferograms for spurt must be single reference."
-        raise ValueError(errmsg)
+    date_tuples = sorted([get_dates(f) for f in ifg_filenames])
 
-    secondary_dates = [d[1] for d in dates]
-    first_date = dates[0][0].strftime(date_fmt)
+    secondary_dates = [d[1] for d in date_tuples]
+    first_date = date_tuples[0][0].strftime(date_fmt)
     date_strings = [utils.format_dates(d, fmt=date_fmt) for d in secondary_dates]
 
-    date_str_to_file: dict[str, PathOrStr | None] = dict(
-        zip(date_strings, ifg_filenames)
-    )
     # first date - set to None
     # None is special case for reference epoch
[…]
