### 2024-10-12 e4f5ad1
add spurt to install
