+++ .github/workflows/test-build-push.yml
@@ -50,6 +50,7 @@ jobs:
             "snaphu>=0.4.0" \
             "opera-utils>=0.4.1" \
             git+https://github.com/isce-framework/tophu@main \
+          pip install git+https://github.com/isce-framework/spurt@main \
             git+https://github.com/isce-framework/whirlwind@40defb38d2d6deca2819934788ebbc57e418e32d
           pip install --no-deps .
       - name: Install test dependencies
