### 2024-10-12 c5767e7
fix optional type check
