### 2024-10-12 ca8e463
revert spurt IfgReader
