+++ src/dolphin/workflows/config/_displacement.py
@@ -250,6 +250,11 @@ class DisplacementWorkflow(WorkflowBase):
         # Modify interferogram options if using spurt for 3d unwrapping,
         # which only does nearest-3 interferograms
         if self.unwrap_options.unwrap_method == UnwrapMethod.SPURT:
-            logger.info("Using spurt: forming nearest-3 interferograms.")
-            self.interferogram_network.reference_idx = None
-            self.interferogram_network.max_bandwidth = 3
+            logger.info(
+                "Using spurt: will form single reference interferograms, later convert"
+                " to nearest-3"
+            )
+            self.interferogram_network.reference_idx = 0
+            # Force all other network options to None
+            for attr in ["max_bandwidth", "max_temporal_baseline", "indexes"]:
+                setattr(self.interferogram_network, attr, None)
