### 2024-10-12 f6aed32
final workaround to get spurt working
