### 2024-10-12 7db4675
work around for both spurt and for snaphu
