### 2024-10-12 cb32fc7
Add stitched phase cosine similarity raster to outputs (#446)
* Modify `io.process_blocks` to accept overlapping blocks

Result is similar to `map_overlaps`, but with arbitrary number of input or output rasters

* create and pass through similarity raster

* fix existing similarity file search

* use `nan` for nodata, mask pixels with all nan/zero

* fix sim tests
