+++ src/dolphin/workflows/displacement.py
@@ -35,6 +35,7 @@ class OutputPaths(NamedTuple):
     stitched_ps_file: Path
     stitched_amp_dispersion_file: Path
     stitched_shp_count_file: Path
+    stitched_similarity_file: Path
     unwrapped_paths: list[Path] | None
     conncomp_paths: list[Path] | None
     timeseries_paths: list[Path] | None
@@ -128,6 +129,7 @@ def run(
     ps_file_list: list[Path] = []
     amp_dispersion_file_list: list[Path] = []
     shp_count_file_list: list[Path] = []
+    similarity_file_list: list[Path] = []
     # The comp_slc tracking object is a dict, since we'll need to organize
     # multiple comp slcs by burst (they'll have the same filename)
     comp_slc_dict: dict[str, list[Path]] = {}
@@ -161,15 +163,22 @@ def run(
         for fut in fut_to_burst:
             burst = fut_to_burst[fut]
 
-            cur_ifg_list, comp_slcs, temp_coh, ps_file, amp_disp_file, shp_count = (
-                fut.result()
-            )
+            (
+                cur_ifg_list,
+                comp_slcs,
+                temp_coh,
+                ps_file,
+                amp_disp_file,
+                shp_count,
+                similarity,
+            ) = fut.result()
             ifg_file_list.extend(cur_ifg_list)
             comp_slc_dict[burst] = comp_slcs
             temp_coh_file_list.append(temp_coh)
             ps_file_list.append(ps_file)
             amp_dispersion_file_list.append(amp_disp_file)
             shp_count_file_list.append(shp_count)
+            similarity_file_list.append(similarity)
 
     # ###################################
     # 2. Stitch burst-wise interferograms
@@ -186,12 +195,14 @@ def run(
         stitched_ps_file,
         stitched_amp_dispersion_file,
         stitched_shp_count_file,
+        stitched_similarity_file,
     ) = stitching_bursts.run(
         ifg_file_list=ifg_file_list,
         temp_coh_file_list=temp_coh_file_list,
         ps_file_list=ps_file_list,
         amp_dispersion_list=amp_dispersion_file_list,
         shp_count_file_list=shp_count_file_list,
+        similarity_file_list=similarity_file_list,
         stitched_ifg_dir=cfg.interferogram_network._directory,
         output_options=cfg.output_options,
         file_date_fmt=cfg.input_options.cslc_date_fmt,
@@ -212,6 +223,7 @@ def run(
             stitched_ps_file=stitched_ps_file,
             stitched_amp_dispersion_file=stitched_amp_dispersion_file,
             stitched_shp_count_file=stitched_shp_count_file,
+            stitched_similarity_file=stitched_similarity_file,
             unwrapped_paths=None,
             conncomp_paths=None,
             timeseries_paths=None,
@@ -354,6 +366,7 @@ def run(
         stitched_ps_file=stitched_ps_file,
         stitched_amp_dispersion_file=stitched_amp_dispersion_file,
         stitched_shp_count_file=stitched_shp_count_file,
+        stitched_similarity_file=stitched_similarity_file,
         unwrapped_paths=unwrapped_paths,
         # TODO: Let's keep the unwrapped_paths since all the outputs are
         # corresponding to those and if we have a network unwrapping, the
