### 2024-10-12 e776da6
Add logic to reset the `extra_reference_date` after a network unwrapping (#442)
* Add re-referencing for network unwrappers using `extra_reference_date`

* move out to separate helper function

* sort output paths

* delete

* add failing test

* for spurt, override the `interferogram_network` configuration

* add `IfgStackReader` for pre-formed nearest-3 interferograms

* add spurt to install

* fix optional type check

* revert spurt IfgReader

* final workaround to get spurt working

* work around for both spurt and for snaphu

* try adjusting deps

* fix disp reference test
