+++ src/dolphin/timeseries.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import logging
+from datetime import datetime
 from enum import Enum
 from pathlib import Path
 from tempfile import NamedTemporaryFile
@@ -17,7 +18,7 @@ from scipy import ndimage
 from dolphin import DateOrDatetime, io, utils
 from dolphin._overviews import ImageType, create_overviews
 from dolphin._types import PathOrStr, ReferencePoint
-from dolphin.utils import flatten, format_dates, full_suffix
+from dolphin.utils import flatten, format_dates, full_suffix, get_nearest_date_idx
 from dolphin.workflows import CallFunc
 
 T = TypeVar("T")
@@ -55,6 +56,7 @@ def run(
     reference_point: tuple[int, int] = (-1, -1),
     wavelength: float | None = None,
     add_overviews: bool = True,
+    extra_reference_date: datetime | None = None,
 ) -> tuple[list[Path], ReferencePoint]:
     """Invert the unwrapped interferograms, estimate timeseries and phase velocity.
 
@@ -108,6 +110,9 @@ def run(
     add_overviews : bool, optional
         If True, creates overviews of the new velocity raster.
         Default is True.
+    extra_reference_date : datetime.datetime, optional
+        If provided, makes another set of interferograms referenced to this
+        for all dates later than it.
 
     Returns
     -------
@@ -203,7 +208,64 @@ def run(
             num_threads=num_threads,
         )
 
-    return inverted_phase_paths, ref_point
+    if extra_reference_date is None:
+        final_out = inverted_phase_paths
+    else:
+        final_out = _redo_reference(inverted_phase_paths, extra_reference_date)
+
+    return final_out, ref_point
+
+
+def _redo_reference(
+    inverted_phase_paths: Sequence[Path], extra_reference_date: datetime
+):
+    """Reset the reference date in `inverted_phase_paths`.
+
+    Affects all files whose secondary is after `extra_reference_date`.
+
+    E.g Given the (day 1, day 2), ..., (day 1, day N) pairs, outputs
+    (1, 2), (1, 3), ...(1, r), (r, r+1), ..., (r, N)
+    where r is the index of the `extra_reference_date`
+    """
+    output_path = inverted_phase_paths[0].parent
+    inverted_date_pairs: list[tuple[datetime, datetime]] = [
+        get_dates(p.stem)[:2] for p in inverted_phase_paths
+    ]
+    secondary_dates = [pair[1] for pair in inverted_date_pairs]
+    extra_ref_idx = get_nearest_date_idx(
+        secondary_dates, requested=extra_reference_date
+    )
+    ref_date = secondary_dates[extra_ref_idx]
+    logger.info(f"Re-referencing later timeseries files to {ref_date}")
+    extra_ref_img = inverted_phase_paths[extra_ref_idx]
+    ref = io.load_gdal(extra_ref_img, masked=True)
+
+    # Use a temp directory while re-referencing
+    extra_out_dir = inverted_phase_paths[0].parent / "extra"
+    extra_out_dir.mkdir(exist_ok=True)
+
+    for idx in range(extra_ref_idx + 1, len(inverted_date_pairs)):
+        # To create the interferogram (r, r+1), we subtract
+        # (1, r) from (1, r+1)
+        cur_img = inverted_phase_paths[idx]
+        new_stem = format_dates(ref_date, secondary_dates[idx])
+        cur_output_name = extra_out_dir / f"{new_stem}.tif"
+        cur = io.load_gdal(cur_img, masked=True)
+        new_out = cur - ref
+        io.write_arr(
+            arr=new_out, like_filename=extra_ref_img, output_name=cur_output_name
+        )
+
[…]
