+++ tests/test_workflows_displacement.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import importlib.util
 from pathlib import Path
 
 import pytest
@@ -188,14 +189,20 @@ def test_separate_workflow_runs(slc_file_list, tmp_path):
     assert [f.name for f in ifgs3_b] == [f.name for f in ifgs3]
 
 
-def test_displacement_run_extra_reference_date(opera_slc_files: list[Path], tmpdir):
+@pytest.mark.parametrize("unwrap_method", ["phass", "spurt"])
+def test_displacement_run_extra_reference_date(
+    opera_slc_files: list[Path], tmpdir, unwrap_method: str
+):
+    if unwrap_method == "spurt" and importlib.util.find_spec("spurt") is None:
+        pytest.skip(reason="spurt unwrapper not installed")
+
     with tmpdir.as_cwd():
         cfg = config.DisplacementWorkflow(
             # start_date = 20220101
             # shape = (4, 128, 128)
             # First one is COMPRESSED_
             output_options={"extra_reference_date": "2022-01-03"},
-            unwrap_options={"unwrap_method": "phass"},
+            unwrap_options={"unwrap_method": unwrap_method},
             cslc_file_list=opera_slc_files,
             input_options={"subdataset": "/data/VV"},
             phase_linking={
@@ -210,16 +217,28 @@ def test_displacement_run_extra_reference_date(opera_slc_files: list[Path], tmpd
 
         # The unwrapped/timeseries files should have a changeover to the new reference
         assert paths.unwrapped_paths is not None
-        unw_names = [pp.name for pp in paths.unwrapped_paths]
-        assert unw_names == [
-            "20220101_20220102.unw.tif",
-            "20220101_20220103.unw.tif",
-            "20220103_20220104.unw.tif",
-        ]
         assert paths.timeseries_paths is not None
+
         ts_names = [pp.name for pp in paths.timeseries_paths]
         assert ts_names == [
             "20220101_20220102.tif",
             "20220101_20220103.tif",
             "20220103_20220104.tif",
         ]
+
+        unw_names = [pp.name for pp in paths.unwrapped_paths]
+        if cfg.unwrap_options.unwrap_method == "spurt":
+            assert unw_names == [
+                "20220101_20220102.unw.tif",
+                "20220101_20220103.unw.tif",
+                "20220101_20220104.unw.tif",
+                "20220102_20220103.unw.tif",
+                "20220102_20220104.unw.tif",
+                "20220103_20220104.unw.tif",
+            ]
+        else:
+            assert unw_names == [
+                "20220101_20220102.unw.tif",
+                "20220101_20220103.unw.tif",
+                "20220103_20220104.unw.tif",
+            ]
