### 2024-10-13 d5fcf94
add a post processing step to fill nans after running spurt
