+++ src/dolphin/unwrap/_post_process.py
@@ -1,8 +1,13 @@
+import logging
+
 import numpy as np
 from numba import njit, stencil
-from numpy.typing import ArrayLike
+from numpy.typing import ArrayLike, NDArray
+from scipy.interpolate import NearestNDInterpolator
 
 TWOPI = 2 * np.pi
+
+logger = logging.getLogger(__name__)
 
 
 @njit(nogil=True)
@@ -54,32 +59,65 @@ def rewrap_to_twopi(arr: ArrayLike) -> np.ndarray:
     return np.mod(np.pi + arr, TWOPI) - np.pi
 
 
-def _get_ambiguities(unw: ArrayLike, round_decimals: int = 4) -> np.ndarray:
+def get_2pi_ambiguities(
+    unw: NDArray[np.floating], round_decimals: int = 4
+) -> NDArray[np.floating]:
+    """Find the number of 2pi offsets from [-pi, pi) at each pixel of `unw`."""
     mod_2pi_image = np.mod(np.pi + unw, TWOPI) - np.pi
     re_wrapped = np.round(mod_2pi_image, round_decimals)
     return np.round((unw - re_wrapped) / (TWOPI), round_decimals - 1)
 
 
-def _fill_masked_ambiguities(
-    amb_image: ArrayLike, mask: ArrayLike, filter_sigma: int = 60
-) -> np.ndarray:
-    from dolphin.filtering import gaussian_filter_nan
+def interpolate_masked_gaps(
+    unw: NDArray[np.float_], ifg: NDArray[np.complex64]
+) -> None:
+    """Perform phase unwrapping using nearest neighbor interpolation of ambiguities.
 
-    masked_ambs = amb_image.copy()
-    masked_ambs[mask] = np.nan
-    ambs_filled = np.round(gaussian_filter_nan(amb_image, filter_sigma))
+    Overwrites `unw`'s masked pixels with the interpolated values.
 
-    out_filled = amb_image.copy()
-    out_filled[mask] = ambs_filled[mask]
-    return out_filled
+    This function takes an input unwrapped phase array containin NaNs at masked pixel.
+    It calculates the phase ambiguity, K, at the attempted unwrapped pixels, then
+    interpolates the ambiguities to fill the gaps.
+    The masked pixels get the value of the original wrapped phase + 2pi*K.
 
+    Parameters
+    ----------
+    unw : NDArray[np.float]
+        Input unwrapped phase array with NaN values for masked areas.
+    ifg : NDArray[np.complex64]
+        Corresponding wrapped interferogram phase
 
-def _smooth_masked_areas(unw, mask, filter_sigma: int = 60):
-    amb = _get_ambiguities(unw)
-    amb_filled = _fill_masked_ambiguities(amb, mask, filter_sigma=filter_sigma)
+    Returns
+    -------
+    np.ndarray
+        Fully unwrapped phase array with interpolated values for previously
+        masked areas.
 
-    out = unw.copy()
-    rewrapped_phase_vec = np.mod(np.pi + unw[mask], TWOPI) - np.pi
-    new_amb_vec = amb_filled[mask]
-    out[mask] = rewrapped_phase_vec + (new_amb_vec * TWOPI)
-    return out
+    """
+    # Create masks for valid areas
+    ifg_valid = ~np.isnan(ifg) & (ifg != 0)
+    unw_valid = ~np.isnan(unw)
+
+    # Identify areas to interpolate: where ifg is valid but unw is not
+    interpolate_mask = ifg_valid & ~unw_valid
+
+    # If there's nothing to interpolate, we're done
+    if not np.any(interpolate_mask):
+        return
+
+    # Calculate ambiguities for valid unwrapped pixels
[…]
