### 2024-10-13 919da2d
Apply nan/0 mask to output of `goldstein`
Closes #251, adds citation to paper.
