+++ src/dolphin/goldstein.py
@@ -1,14 +1,16 @@
 import numpy as np
-from numpy.typing import ArrayLike
+from numpy.typing import NDArray
 
 
-def goldstein(phase: ArrayLike, alpha: float, psize: int = 32) -> np.ndarray:
+def goldstein(
+    phase: NDArray[np.complex_] | NDArray[np.float_], alpha: float, psize: int = 32
+) -> np.ndarray:
     """Apply the Goldstein adaptive filter to the given data.
 
     Parameters
     ----------
     phase : np.ndarray
-        2D complex array containing the data to be filtered.
+        2D array of floating point phase, or complex data, to be filtered.
     alpha : float
         Filtering parameter for Goldstein algorithm
         Must be between 0 (no filtering) and 1 (maximum filtering)
@@ -20,18 +22,22 @@ def goldstein(phase: ArrayLike, alpha: float, psize: int = 32) -> np.ndarray:
     -------
         2D numpy array of filtered data.
 
+    References
+    ----------
+    [@Goldstein1998RadarInterferogramFiltering]
+
     """
 
-    def apply_pspec(data: ArrayLike):
+    def apply_pspec(data: NDArray[np.complex_]) -> np.ndarray:
         # NaN is allowed value
         if alpha < 0:
             raise ValueError(f"alpha must be >= 0, got {alpha = }")
 
-        wgt = np.power(np.abs(data) ** 2, alpha / 2)
-        data = wgt * data
+        weight = np.power(np.abs(data) ** 2, alpha / 2)
+        data = weight * data
         return data
 
-    def make_wgt(nxp: int, nyp: int) -> np.ndarray:
+    def make_weight(nxp: int, nyp: int) -> np.ndarray:
         # Create arrays of horizontal and vertical weights
         wx = 1.0 - np.abs(np.arange(nxp // 2) - (nxp / 2.0 - 1.0)) / (nxp / 2.0 - 1.0)
         wy = 1.0 - np.abs(np.arange(nyp // 2) - (nyp / 2.0 - 1.0)) / (nyp / 2.0 - 1.0)
@@ -39,16 +45,16 @@ def goldstein(phase: ArrayLike, alpha: float, psize: int = 32) -> np.ndarray:
         # the top-left quadrant of the weight matrix
         quadrant = np.outer(wy, wx)
         # Create a full weight matrix by mirroring the quadrant along both axes
-        wgt = np.block(
+        weight = np.block(
             [
                 [quadrant, np.flip(quadrant, axis=1)],
                 [np.flip(quadrant, axis=0), np.flip(np.flip(quadrant, axis=0), axis=1)],
             ]
         )
-        return wgt
+        return weight
 
     def patch_goldstein_filter(
-        data: ArrayLike, wgt: ArrayLike, psize: int
+        data: NDArray[np.complex_], weight: NDArray[np.float_], psize: int
     ) -> np.ndarray:
         """Apply the filter to a single patch of data.
 
@@ -56,7 +62,7 @@ def goldstein(phase: ArrayLike, alpha: float, psize: int = 32) -> np.ndarray:
         ----------
         data : np.ndarray
             2D complex array containing the data to be filtered.
-        wgt : np.ndarray
+        weight : np.ndarray
             weight matrix for summing neighboring data
         psize : int
             edge length of square FFT area
@@ -70,30 +76,39 @@ def goldstein(phase: ArrayLike, alpha: float, psize: int = 32) -> np.ndarray:
         data = np.fft.fft2(data, s=(psize, psize))
         data = apply_pspec(data)
         data = np.fft.ifft2(data, s=(psize, psize))
-        return wgt * data
+        return weight * data
 
[…]
