+++ docs/references.bib
@@ -165,6 +165,18 @@
   keywords = {3-D,4-D and multidimensional (Multi-D) SAR imaging,Covariance matrices,Covariance matrix decomposition,differential SAR tomography,differential synthetic aperture radar (SAR) interferometry (DInSAR),Interferometry,Monitoring,principal component analysis (PCA),SAR interferometry (InSAR),SAR tomography,Scattering,Spatial resolution,Synthetic aperture radar,Tomography}
 }
 
+@article{Goldstein1998RadarInterferogramFiltering,
+  title = {Radar Interferogram Filtering for Geophysical Applications},
+  author = {Goldstein, Richard M. and Werner, Charles L.},
+  year = {1998},
+  journal = {Geophysical Research Letters},
+  volume = {25},
+  number = {21},
+  pages = {4035--4038},
+  issn = {1944-8007},
+  doi = {10.1029/1998GL900033}
+}
+
 @article{Guarnieri2008ExploitationTargetStatistics,
   title = {On the {{Exploitation}} of {{Target Statistics}} for {{SAR Interferometry Applications}}},
   author = {Guarnieri, A. M. and Tebaldini, S.},
