### 2024-10-13 7c28697
remove deprecated `complex_`
also see if `nlooks` is causing whirlwind to freeze
