### 2024-10-14 26a8217
fix numpy 2.0 deprecations by ruff
