+++ src/dolphin/atmosphere/weather_model.py
@@ -342,10 +342,10 @@ class WeatherModel(ABC):
         # Get the integrated ZTD
         wet_total, hydro_total = np.zeros(wet.shape), np.zeros(hydro.shape)
         for level in range(wet.shape[2]):
-            wet_total[..., level] = 1e-6 * np.trapz(
+            wet_total[..., level] = 1e-6 * np.trapezoid(
                 wet[..., level:], x=self._zs[level:], axis=2
             )
-            hydro_total[..., level] = 1e-6 * np.trapz(
+            hydro_total[..., level] = 1e-6 * np.trapezoid(
                 hydro[..., level:], x=self._zs[level:], axis=2
             )
         self._hydrostatic_ztd = hydro_total
