+++ .github/workflows/test-build-push.yml
@@ -51,7 +51,7 @@ jobs:
             "opera-utils>=0.4.1" \
             git+https://github.com/isce-framework/tophu@main \
             git+https://github.com/isce-framework/whirlwind@40defb38d2d6deca2819934788ebbc57e418e32d
-          python -m pip install git+https://github.com/isce-framework/spurt@main
+          python -m pip install git+https://github.com/scottstanie/spurt@use-mp-spawn
           python -m pip install --no-deps .
       - name: Install test dependencies
         run: |
