### 2024-10-14 f5fc4c3
Add a post processing step to fill `nan`s after running `spurt` (#448)
* add a post processing step to fill nans after running spurt

* dont lose profile during overwrite
