### 2024-10-14 3571819
Apply nan/0 mask to output of `goldstein` filter (#447)
* Apply nan/0 mask to output of `goldstein`

Closes #251, adds citation to paper.

* remove deprecated `complex_`

also see if `nlooks` is causing whirlwind to freeze

* fix numpy 2.0 deprecations by ruff

* fix sizes

* test spurt branch
