### 2024-10-14 215929b
Use `dataclasses` instead of `NamedTuple` for displacement, stitched outputs (#449)
* Use `dataclasses` instead of `NamedTuple` for displacement, stitched outputs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
