### 2024-10-14 2bc45e3
Add `outside_input_range` options to `get_nearest_date_idx`
