### 2024-10-14 183455b
Pass through `units` during `timeseries._redo_reference`
