+++ src/dolphin/timeseries.py
@@ -161,7 +161,7 @@ def run(
             "Skipping inversion step: only single reference interferograms exist."
         )
         # Copy over the unwrapped paths to `timeseries/`
-        inverted_phase_paths = _convert_and_reference(
+        final_out = _convert_and_reference(
             unwrapped_paths,
             output_dir=output_dir,
             reference_point=ref_point,
@@ -178,6 +178,11 @@ def run(
             wavelength=wavelength,
             method=method,
         )
+        if extra_reference_date is None:
+            final_out = inverted_phase_paths
+        else:
+            final_out = _redo_reference(inverted_phase_paths, extra_reference_date)
+
     if add_overviews:
         logger.info("Creating overviews for timeseries images")
         create_overviews(inverted_phase_paths, image_type=ImageType.UNWRAPPED)
@@ -207,11 +212,6 @@ def run(
             block_shape=block_shape,
             num_threads=num_threads,
         )
-
-    if extra_reference_date is None:
-        final_out = inverted_phase_paths
-    else:
-        final_out = _redo_reference(inverted_phase_paths, extra_reference_date)
 
     return final_out, ref_point
 
@@ -243,6 +243,7 @@ def _redo_reference(
     # Use a temp directory while re-referencing
     extra_out_dir = inverted_phase_paths[0].parent / "extra"
     extra_out_dir.mkdir(exist_ok=True)
+    units = io.get_raster_units(inverted_phase_paths[0])
 
     for idx in range(extra_ref_idx + 1, len(inverted_date_pairs)):
         # To create the interferogram (r, r+1), we subtract
@@ -253,7 +254,10 @@ def _redo_reference(
         cur = io.load_gdal(cur_img, masked=True)
         new_out = cur - ref
         io.write_arr(
-            arr=new_out, like_filename=extra_ref_img, output_name=cur_output_name
+            arr=new_out,
+            like_filename=extra_ref_img,
+            output_name=cur_output_name,
+            units=units,
         )
 
     for idx, p in enumerate(inverted_phase_paths):
