### 2024-10-14 8d46e32
fix other ts file references
