### 2024-10-15 b25b003
Fix the relative index used to make compressed SLCs
