### 2024-10-16 2a9caca
scrap sample config ci
