### 2024-10-17 7beede0
Form nearest-3 interferograms when computing similarity
