+++ src/dolphin/workflows/single.py
@@ -11,7 +11,7 @@ import numpy as np
 from numpy.typing import DTypeLike
 from tqdm.auto import tqdm
 
-from dolphin import io, shp
+from dolphin import io, shp, similarity
 from dolphin._decorators import atomic_output
 from dolphin._types import Filename, HalfWindow, Strides
 from dolphin.io import EagerLoader, StridedBlockManager, VRTStack
@@ -294,6 +294,15 @@ def run_wrapped_phase_single(
 
     logger.info("Repacking for more compression")
     io.repack_rasters(phase_linked_slc_files, keep_bits=12)
+
+    logger.info("Creating similarity raster on outputs")
+    similarity.create_similarities(
+        phase_linked_slc_files,
+        output_file=output_folder / f"similarity_{start_end}.tif",
+        num_threads=1,
+        add_overviews=False,
+        nearest_n=3,
+    )
 
     written_comp_slc = output_files[0]
 
