### 2024-10-17 fcf1405
Pass `similarity` to unwrapping, use as interpolation metric
