### 2024-10-18 eefff8a
Fix relative compressed reference index
Subtracting the `first_real_slc_idx` was not correct:
The stack planner would put 0 as the compressed index
for all ministacks.
Considering the 2nd ministack, if the `output_reference_idx`
was 0, then `compressed_reference_idx` would be 0. Then we
were slicing away the compressed, but also subtracting the
number of compressed. The led to *neagtive indexes*.
