+++ src/dolphin/phase_link/_compress.py
@@ -9,6 +9,7 @@ from dolphin.utils import upsample_nearest
 def compress(
     slc_stack: ArrayLike,
     pl_cpx_phase: ArrayLike,
+    first_real_slc_idx: int = 0,
     slc_mean: ArrayLike | None = None,
     reference_idx: int | None = None,
 ):
@@ -21,6 +22,10 @@ def compress(
     pl_cpx_phase : ArrayLike
         The estimated complex phase from phase linking.
         shape  = (nslc, rows // strides.y, cols // strides.x)
+    first_real_slc_idx : int
+        Index within `slc_stack` where the "real" SLCs start.
+        Indexes before this (which are assumed to be already Compressed SLCs) are
+        excluded from the dot product during compression
     slc_mean : ArrayLike, optional
         The mean SLC magnitude, shape (rows, cols), to use as output pixel magnitudes.
         If None, the mean is computed from the input SLC stack.
@@ -39,6 +44,7 @@ def compress(
         pl_referenced = pl_cpx_phase * pl_cpx_phase[reference_idx][None, :, :].conj()
     else:
         pl_referenced = pl_cpx_phase
+    pl_referenced = pl_referenced[first_real_slc_idx:, :, :]
     # If the output is downsampled, we need to make `pl_cpx_phase` the same shape
     # as the output
     pl_estimate_upsampled = upsample_nearest(pl_referenced, slc_stack.shape[1:])
