### 2024-10-18 05d80b0
add additional pre-set methods for making compressed SLCs
