### 2024-10-18 1f6cfa8
shrink size of slcs to match in compress
