### 2024-10-18 3008edf
Add ability to specify output bounds as WKT (#427)
* Add ability to specify output bounds as WKT

* allow .wkt file too

* check for none

* try fixing GHA

* scrap sample config ci
