+++ src/dolphin/workflows/wrapped_phase.py
@@ -85,6 +85,7 @@ def run(
     mask_filename = _get_mask(
         output_dir=cfg.work_directory,
         output_bounds=cfg.output_options.bounds,
+        output_bounds_wkt=cfg.output_options.bounds_wkt,
         output_bounds_epsg=cfg.output_options.bounds_epsg,
         like_filename=vrt_stack.outfile,
         cslc_file_list=non_compressed_slcs,
@@ -448,6 +449,7 @@ def _get_input_dates(
 def _get_mask(
     output_dir: Path,
     output_bounds: Bbox | tuple[float, float, float, float] | None,
+    output_bounds_wkt: str | None,
     output_bounds_epsg: int,
     like_filename: Filename,
     cslc_file_list: Sequence[Filename],
@@ -467,11 +469,12 @@ def _get_mask(
 
     mask_filename: Path | None = None
     # Also mask outside the area of interest if we've specified a small bounds
-    if output_bounds is not None:
+    if output_bounds is not None or output_bounds_wkt is not None:
         # Make a mask just from the bounds
         bounds_mask_filename = output_dir / "bounds_mask.tif"
         masking.create_bounds_mask(
             bounds=output_bounds,
+            bounds_wkt=output_bounds_wkt,
             bounds_epsg=output_bounds_epsg,
             output_filename=bounds_mask_filename,
             like_filename=like_filename,
