+++ src/dolphin/workflows/config/_common.py
@@ -19,6 +19,7 @@ from pydantic import (
 from dolphin import __version__ as _dolphin_version
 from dolphin._types import Bbox
 from dolphin.io import DEFAULT_HDF5_OPTIONS, DEFAULT_TIFF_OPTIONS
+from dolphin.stack import CompressedSlcPlan
 
 from ._enums import ShpMethod
 from ._yaml_model import YamlModel
@@ -129,6 +130,7 @@ class PhaseLinkingOptions(BaseModel, extra="forbid"):
             "`n` interferograms. `baseline_line` must be positive."
         ),
     )
+    compressed_slc_plan: CompressedSlcPlan = CompressedSlcPlan.ALWAYS_FIRST
 
 
 class InterferogramNetwork(BaseModel, extra="forbid"):
