+++ src/dolphin/workflows/unwrapping.py
@@ -23,8 +23,8 @@ def run(
     cor_file_list: Sequence[Path],
     nlooks: float,
     unwrap_options: UnwrapOptions,
-    temporal_coherence_file: PathOrStr | None = None,
-    similarity_file: PathOrStr | None = None,
+    temporal_coherence_filename: PathOrStr | None = None,
+    similarity_filename: PathOrStr | None = None,
     mask_file: PathOrStr | None = None,
     add_overviews: bool = True,
 ) -> tuple[list[Path], list[Path]]:
@@ -41,9 +41,9 @@ def run(
     unwrap_options : UnwrapOptions
         [`UnwrapOptions`][dolphin.workflows.config.UnwrapOptions] config object
         with parameters for running unwrapping jobs.
-    temporal_coherence_file : Filename, optional
+    temporal_coherence_filename : Filename, optional
         Path to temporal coherence file from phase linking.
-    similarity_file : Filename, optional
+    similarity_filename : Filename, optional
         Path to phase cosine similarity file from phase linking.
     mask_file : PathOrStr, optional
         Path to boolean mask indicating nodata areas.
@@ -88,8 +88,8 @@ def run(
         output_path=output_path,
         unwrap_options=unwrap_options,
         nlooks=nlooks,
-        temporal_coherence_file=temporal_coherence_file,
-        similarity_file=similarity_file,
+        temporal_coherence_filename=temporal_coherence_filename,
+        similarity_filename=similarity_filename,
         mask_filename=output_mask,
         scratchdir=unwrap_scratchdir,
     )
