### 2024-10-20 60b2eb4
Call spurt from `subprocess` to avoid fork issues
