+++ src/dolphin/unwrap/_unwrap_3d.py
@@ -1,5 +1,6 @@
 import logging
 import shutil
+import subprocess
 from collections.abc import Sequence
 from pathlib import Path
 
@@ -9,7 +10,7 @@ from numpy.typing import NDArray
 from opera_utils import get_dates
 from scipy import ndimage, signal
 
-from dolphin import io, utils
+from dolphin import io
 from dolphin._types import PathOrStr
 from dolphin.workflows.config import SpurtOptions
 
@@ -25,110 +26,79 @@ def unwrap_spurt(
     ifg_filenames: Sequence[PathOrStr],
     output_path: PathOrStr,
     temporal_coherence_file: PathOrStr,
-    cor_filenames: Sequence[PathOrStr] | None = None,
+    # cor_filenames: Sequence[PathOrStr] | None = None,
     mask_filename: PathOrStr | None = None,
     options: SpurtOptions = DEFAULT_OPTIONS,
     scratchdir: PathOrStr | None = None,
 ) -> tuple[list[Path], list[Path]]:
-    """Perform 3D unwrapping using `spurt`."""
-    from spurt.graph import Hop3Graph
-    from spurt.io import SLCStackReader
-    from spurt.workflows.emcf import (
-        GeneralSettings,
-        MergerSettings,
-        SolverSettings,
-        TilerSettings,
-        compute_phasediff_deciles,
-        get_bulk_offsets,
-        get_tiles,
-        merge_tiles,
-        unwrap_tiles,
-    )
-
-    if existing_unw_files := sorted(Path(output_path).glob(f"*{UNW_SUFFIX}")):
-        logger.info(f"Found {len(existing_unw_files)} unwrapped files")
-        existing_ccl_files = sorted(Path(output_path).glob(f"*{CONNCOMP_SUFFIX}"))
-        return existing_unw_files, existing_ccl_files
-
-    if cor_filenames is not None:
-        assert len(ifg_filenames) == len(cor_filenames)
+    """Perform 3D unwrapping using `spurt` via subprocess call."""
     if mask_filename is not None:
-        # TODO: Combine this with the temporal coherence to pass one 0/1 mask
-        # This will still work for spurt, since it runs `> threshold`, which is
-        # always true once we set our desired pixels to 1 and undesired to 0
-        _mask = io.load_gdal(mask_filename)
+        logger.warning("use of `mask_filename` not yet implemented for spurt")
 
-    if scratchdir is None:
-        scratchdir = Path(output_path) / "scratch"
-    gen_settings = GeneralSettings(
-        output_folder=output_path,
-        intermediate_folder=scratchdir,
-        **options.general_settings.model_dump(),
+    cmd = [
+        "python",
+        "-m",
+        "spurt.workflows.emcf",
+        "-i",
+        str(
+            Path(ifg_filenames[0]).parent
+        ),  # Assuming all ifgs are in the same directory
+        "-o",
+        str(output_path),
+        "--tempdir",
+        str(scratchdir or Path(output_path) / "scratch"),
+        "-c",
+        str(options.temporal_coherence_threshold),
+    ]
+    if not options.general_settings.use_tiles:
+        cmd.append("--singletile")
+
+    # Tiler Settings
+    cmd.extend(
+        [
+            "--pts-per-tile",
+            str(options.tiler_settings.target_points_per_tile),
+        ]
     )
 
-    tile_settings = TilerSettings(**options.tiler_settings.model_dump())
-    slv_settings = SolverSettings(**options.solver_settings.model_dump())
-    mrg_settings = MergerSettings(**options.merger_settings.model_dump())
-
-    # Using default Hop3Graph
[…]
