+++ docs/references.bib
@@ -100,14 +100,29 @@
 }
 
 @article{Fattahi2019FRInGEFullResolutionInSAR,
-  title = {{{FRInGE}}; {{Full-Resolution InSAR}} Timeseries Using {{Generalized Eigenvectors}}},
-  author = {Fattahi, H. and Agram, P. S. and Tymofyeyeva, E. and Bekaert, D. P.},
-  year = {2019},
-  month = dec,
+  author = {{Fattahi}, H. and {Agram}, P.S. and {Tymofyeyeva}, E. and {Bekaert}, D.P.},
+  title = {{FRInGE; Full-Resolution InSAR timeseries using Generalized Eigenvectors}},
+  keywords = {1209 Tectonic deformation, GEODESY AND GRAVITY, 1211 Non-tectonic deformation, GEODESY AND GRAVITY, 1240 Satellite geodesy: results, GEODESY AND GRAVITY, 1241 Satellite geodesy: technical issues, GEODESY AND GRAVITY},
+  booktitle = {AGU Fall Meeting Abstracts},
+  year = 2019,
   volume = {2019},
+  month = dec,
+  eid = {G11B-0514},
   pages = {G11B-0514},
-  urldate = {2024-05-16},
-  keywords = {1209 Tectonic deformation,1211 Non-tectonic deformation,1240 Satellite geodesy: results,1241 Satellite geodesy: technical issues,GEODESY AND GRAVITY}
+  adsurl = {https://ui.adsabs.harvard.edu/abs/2019AGUFM.G11B0514F},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{Ferretti2001PermanentScattersSAR,
+  title = {Permanent {{Scatters}} in {{SAR Interferometry}}},
+  author = {Ferretti, Alessandro and Prati, Claudio and Rocca, Fabrio},
+  year = {2001},
+  journal = {IEEE Transactions on Geoscience and Remote Sensing},
+  volume = {39},
+  number = {1},
+  pages = {8--20},
+  issn = {01962892},
+  doi = {10.1109/36.898661}
 }
 
 @article{Ferretti2011NewAlgorithmProcessing,
@@ -138,6 +153,18 @@
   issn = {1558-0644},
   doi = {10.1109/TGRS.2014.2352853},
   keywords = {3-D,4-D and multidimensional (Multi-D) SAR imaging,Covariance matrices,Covariance matrix decomposition,differential SAR tomography,differential synthetic aperture radar (SAR) interferometry (DInSAR),Interferometry,Monitoring,principal component analysis (PCA),SAR interferometry (InSAR),SAR tomography,Scattering,Spatial resolution,Synthetic aperture radar,Tomography}
+}
+
+@article{Goldstein1998RadarInterferogramFiltering,
+  title = {Radar Interferogram Filtering for Geophysical Applications},
+  author = {Goldstein, Richard M. and Werner, Charles L.},
+  year = {1998},
+  journal = {Geophysical Research Letters},
+  volume = {25},
+  number = {21},
+  pages = {4035--4038},
+  issn = {1944-8007},
+  doi = {10.1029/1998GL900033}
 }
 
 @article{Guarnieri2008ExploitationTargetStatistics,
@@ -211,7 +238,7 @@
   publisher = {IEEE},
   address = {Valencia},
   doi = {10.1109/IGARSS.2018.8517504},
-  langid = {english},
+  langid = {english}
 }
 
 @article{Siddiqui1962ProblemsConnectedRayleigh,

[…]
