+++ docs/getting-started.md
@@ -1,4 +1,4 @@
---8<-- "README.md:19:104"
+--8<-- "README.md:19:107"
 
 For a more complete tutorial, see our [tutorials page](tutorials.md)
 
