### 2024-10-20 #6997 @scottstanie
Thanks for the patience @McWhity and @mikemahoney218

Two most notable changes are:

> In the beginning it was hard for me to determine which algorithms are used by dolphin. I suggest to incorporate a graphical overview in the JOSS Paper as well as in the Documentation. I try to explain best I can what I mean. Basically it could be kind of a workflow diagram for the default processing chains including which parts can be easily changed.

I have [created a second figure](https://github.com/isce-framework/dolphin/blob/joss/paper/figures/dolphin-modules.pdf) outlining the main displacement workflow overview. This shows a high level of the processing modules that can be configured.

>You mentioned other software/libraries like Miaplpy or TomoSAR. I see the advantages of dolphin but I think it would be good to clearly state them. Thus I would like see a short discussion of advantages and disadvantages.

I've added details in this commit added [here](https://github.com/isce-framework/dolphin/commit/d7b1f567c432cc9656d6ec742c6d6994863c26d7), noting that

>... there remained a need for software capable of handling the heavy computational demands of large-scale InSAR processing. For example, the TomoSAR library currently requires tens of gigabytes of memory to process more than a small area of interest, while FRInGE and MiaplPy are unable to offer speedups to users who want to process data at a coarser output grid than the full SLC resolution. Additionally, both FRInGE and MiaplPy were designed to process single batches of SLC images... Dolphin was developed to process both historical archives and incrementally handle new data in near-real time.

Other responses:

**README**

> README: Please add a section about contributing.

We have added this in [here](https://github.com/isce-framework/dolphin/commit/5663cfd0682e031368008e982e5d59d921b47c37)

> last sentence of part “Usage”: Is the link to the documentation necessary? It just directs you to the Table of contents.

Good catch! That has been removed  [here](https://github.com/isce-framework/dolphin/commit/16be6187f303405696d5a6f13bfd9bb2fdbd5579) in favor of having the "docs" badge at the top of the readme (as well as the link in the repo description).

>  link to mkdocstrings: page is somehow strange

Thank you, fixed in [here](https://github.com/isce-framework/dolphin/commit/f967eac7112e13e803395516bafce0ef997f6dde)

> - Link to ESA’s Burst database directs you to “404 Page not found”

Fixed [here](https://github.com/isce-framework/dolphin/commit/cb60cc7efb2d469cc1d72d776789ad6ffc41c958)

### Missing section/information

>  For users it might be good if you put a default dolphin_config.yaml file in the github repository as well as in the documentation.

I added one in the docs here https://github.com/isce-framework/dolphin/pull/438.

>Not everyone will read the JOSS paper....

I plan to add links to the JOSS paper, as well as references to find more details about background in the JOSS paper upon publication.

> - Some more information on how to change between SNAPHU, PHASS and ICU would be also nice. Are there different parameter that can be changed for the different algorithms? What is PHASS and ICU? In the README you link to “isce3” but that does not help me to find more information about PHASS or ICU. If there is documentation to PHASS and ICU just pinpoint directly to it.

I [added to the README](https://github.com/isce-framework/dolphin/commit/72ce526f738707fab1b930ec62ffbec326164532) that the unwrapped can be changed with the `unwrap_method` option. For the unwrappers, I've tried to leave out detailed descriptions to avoid stale/incorrect information. For instance, the isce3 bindings to SNAPHU have been removed since the paper was originally added, and we've tried to emphasize those options in isce3 less. I am currently waiting on a documentation push from the isce3 developers as NISAR nears launch so I can point users at more details about ICU and PHASS.
