### 2024-10-21 9d3ede8
Replace CSVs with `chi2.ppf` for GLRT test
