+++ src/dolphin/workflows/config/_common.py
@@ -109,7 +109,11 @@ class PhaseLinkingOptions(BaseModel, extra="forbid"):
     shp_method: ShpMethod = ShpMethod.GLRT
     shp_alpha: float = Field(
         0.005,
-        description="Significance level (probability of false alarm) for SHP tests.",
+        description=(
+            "Significance level (probability of false alarm) for SHP tests. Lower"
+            " numbers include more pixels within the multilook window during covariance"
+            " estimation."
+        ),
         gt=0.0,
         lt=1.0,
     )
