### 2024-10-21 768a635
Manually make temp coh mask to prepare for combined masking
