### 2024-10-21 2a8da7d
fix test with mkdir
