### 2024-10-21 53df4b8
Call spurt from `subprocess` to avoid fork issues (#455)
* Call spurt from `subprocess` to avoid fork issues

* Manually make temp coh mask to prepare for combined masking

* fix test with mkdir
