+++ src/dolphin/unwrap/_unwrap_3d.py
@@ -1,5 +1,6 @@
 import logging
 import shutil
+import subprocess
 from collections.abc import Sequence
 from pathlib import Path
 
@@ -9,7 +10,7 @@ from numpy.typing import NDArray
 from opera_utils import get_dates
 from scipy import ndimage, signal
 
-from dolphin import io, utils
+from dolphin import io
 from dolphin._types import PathOrStr
 from dolphin.workflows.config import SpurtOptions
 
@@ -25,110 +26,101 @@ def unwrap_spurt(
     ifg_filenames: Sequence[PathOrStr],
     output_path: PathOrStr,
     temporal_coherence_file: PathOrStr,
-    cor_filenames: Sequence[PathOrStr] | None = None,
+    # cor_filenames: Sequence[PathOrStr] | None = None,
     mask_filename: PathOrStr | None = None,
     options: SpurtOptions = DEFAULT_OPTIONS,
     scratchdir: PathOrStr | None = None,
 ) -> tuple[list[Path], list[Path]]:
-    """Perform 3D unwrapping using `spurt`."""
-    from spurt.graph import Hop3Graph
-    from spurt.io import SLCStackReader
-    from spurt.workflows.emcf import (
-        GeneralSettings,
-        MergerSettings,
-        SolverSettings,
-        TilerSettings,
-        compute_phasediff_deciles,
-        get_bulk_offsets,
-        get_tiles,
-        merge_tiles,
-        unwrap_tiles,
+    """Perform 3D unwrapping using `spurt` via subprocess call."""
+    # NOTE: we are working around spurt currently wanting "temporal_coherence.tif",
+    # and a temporal coherence threshold.
+    # we'll make our own mask of 0=bad, 1=good, then pass a threshold of 0.5
+    temp_coh = io.load_gdal(temporal_coherence_file, masked=True).filled(0)
+    # Mark the "bad" pixels (good=1, bad=0, following the unwrapper mask convention)
+    temp_coh_mask = temp_coh > options.temporal_coherence_threshold
+    if mask_filename:
+        nodata_mask = io.load_gdal(mask_filename).astype(bool)
+        # A good pixel has to be 1 in both masks
+        combined_mask = temp_coh_mask & nodata_mask
+    else:
+        combined_mask = temp_coh_mask
+    # We name it "temporal_coherence.tif" so spurt reads it.
+    # Also make it float32 as though it were temp coh
+    scratch_path = Path(scratchdir) if scratchdir else Path(output_path) / "scratch"
+    scratch_path.mkdir(exist_ok=True, parents=True)
+    combined_mask_filename = scratch_path / "temporal_coherence.tif"
+    io.write_arr(
+        arr=combined_mask.astype("float32"), output_name=combined_mask_filename
     )
 
-    if existing_unw_files := sorted(Path(output_path).glob(f"*{UNW_SUFFIX}")):
-        logger.info(f"Found {len(existing_unw_files)} unwrapped files")
-        existing_ccl_files = sorted(Path(output_path).glob(f"*{CONNCOMP_SUFFIX}"))
-        return existing_unw_files, existing_ccl_files
+    # Symlink the interferograms to the same scratch path so spurt finds everything
+    # expected in the one directory
+    for fn in ifg_filenames:
+        new_path = scratch_path / Path(fn).name
+        new_path.symlink_to(fn)
 
-    if cor_filenames is not None:
-        assert len(ifg_filenames) == len(cor_filenames)
-    if mask_filename is not None:
-        # TODO: Combine this with the temporal coherence to pass one 0/1 mask
-        # This will still work for spurt, since it runs `> threshold`, which is
-        # always true once we set our desired pixels to 1 and undesired to 0
[…]
