### 2024-10-21 eb00339
Replace CSVs with `chi2.ppf` for GLRT test (#456)
* Replace CSVs with `chi2.ppf` for GLRT test

* default 0.001

* test smaller alphas

* specify log in temp dir
