+++ src/dolphin/workflows/config/_common.py
@@ -108,8 +108,12 @@ class PhaseLinkingOptions(BaseModel, extra="forbid"):
     )
     shp_method: ShpMethod = ShpMethod.GLRT
     shp_alpha: float = Field(
-        0.005,
-        description="Significance level (probability of false alarm) for SHP tests.",
+        0.001,
+        description=(
+            "Significance level (probability of false alarm) for SHP tests. Lower"
+            " numbers include more pixels within the multilook window during covariance"
+            " estimation."
+        ),
         gt=0.0,
         lt=1.0,
     )
