### 2024-10-21 2b01391
lower defaults for interp
