### 2024-10-22 2ef8204
Only do nearest-3 similarity if were unwrapping short baseline
