### 2024-11-04 #6997 @mikemahoney218
Awesome! Thank you so much @McWhity  and @margauxmouchene  for your thoughtful reviews here.

@scottstanie, at this point could you walk through the bullets in the post above?
