### 2024-11-04 #6997 @scottstanie
Yes thank you all! I am going to work later today to take care of @McWhity 's last suggestions to make the workflow diagram more visible, and to make sure the main branch is in sync will all the changes to `joss`.
