### 2024-11-05 #6997 @scottstanie
Ok I believe I've fixed the divergence of [fixes in this PR](https://github.com/isce-framework/dolphin/pull/481), so that is merged into the main branch.

@mikemahoney218  Quick question on the release for Zenodo- is making a release of the main branch is okay? Or is it supposed to be a release containing the `paper.pdf` (i.e. the `joss` branch for me)?
