### 2024-11-05 #6997 @mikemahoney218
Main branch is perfect. We view the Zenodo as being the archive of record of the software, while _we_ are the archive of record for the paper -- so there's no need to make new archives if you need to edit the paper, and there's no need to include paper files in the software archive.
