### 2024-11-05 #6997 @scottstanie
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
>
>  * Double check authors and affiliations (including ORCIDs)

👍

>
> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

https://github.com/isce-framework/dolphin/releases/tag/v0.32.0

>
> * Archive the release on Zenodo/figshare/etc and post the DOI here.

https://zenodo.org/records/14041320

>     * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
>     * Make sure that the license listed for the archive is the same as the software license.
>
👍
