### 2024-11-11 #6997 @mikemahoney218
@scottstanie Are you willing to set the archive to use the same license as the repository?
