### 2024-11-11 #6997 @mikemahoney218
Otherwise, the paper looks good -- only blocker on my end is the different (and non-open source) license on the archive version.
