### 2024-11-11 #6997 @scottstanie
Oh sorry! I don't know why Zenodo didn't save the change to the license with the authors, I was having weird browser problems.

It now should match the repo license.
