### 2024-11-11 #6997 @scottstanie
Ah, Zenodo made a new version from the 0.33 release- I changed the version 0.32 release metadata here: https://zenodo.org/records/14041320
