### 2024-11-11 #6997 @scottstanie
Both versions now have the fixed license. I'll add a `CITATION.cff` to the repo so that it doesn't keep defaulting to the wrong one on zenodo.
