### 2024-11-11 #6997 @scottstanie
Also I just wanted to note that I fixed my ORCID number, which somehow got mangled in the `paper.md`.
