### 2024-11-12 #6997 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
