### 2024-11-12 #6997 @kthyng
@scottstanie We recommend that the title and author list for the metadata in your software archive (Zenodo) exactly match those in your JOSS paper to create a cohesive archive, though this is not a strict requirement. Do you want to make this change?
