### 2024-11-12 #6997 @scottstanie
Hi @kthyng , as far as I can tell, the Zenodo archive that I [listed above for 0.32.0](https://zenodo.org/records/14041320) has the same author list as the paper. Is there something different for that version? Or do you mean the 0.33.0 version?
